### 2024-11-25 528a5ce
Changed variable name of well test ALQ
