+++ opm/simulators/wells/WellInterface_impl.hpp
@@ -1039,14 +1039,14 @@ namespace Opm
                                                         phase_usage,
                                                         ecl_well_map,
                                                         deferred_logger);
-        auto [max_alq, success] = glift->wellTestALQ();
+        auto [wtest_alq, success] = glift->wellTestALQ();
         std::string msg;
         const auto& unit_system = schedule.getUnits();
         if (success) {
-            well_state.setALQ(well_name, max_alq);
+            well_state.setALQ(well_name, wtest_alq);
             msg = fmt::format(
                 "GLIFT WTEST: Well {} : Setting ALQ to optimized value = {}",
-                well_name, unit_system.from_si(UnitSystem::measure::gas_surface_rate, max_alq));
+                well_name, unit_system.from_si(UnitSystem::measure::gas_surface_rate, wtest_alq));
         }
         else {
             if (!gl_well.use_glo()) {
