### 2024-11-25 4d5a67f
added: ability to pass command line parameters from python
