### 2024-11-25 #7524 @multitalentloes
Hello Daniel, thank you very much for initiating the pre-review.

The contributions described in this article predominantly reside in the `gpuistl` folder. Redoing the work of the bot with  `cloc` on that folder yields something like this:
```
$ cloc opm-simulators/opm/simulators/linalg/gpuistl/
      59 text files.
      59 unique files.
       0 files ignored.

Language                     files          blank        comment           code
C/C++ Header                    42            804           2616           3193
C++                             12            316            270           2136
CUDA                             5            154            133           1232
SUM:                            59           1274           3019           6561
```

There are other minor parts of the repository which have been added to support the `gpu-ISTL` framework, such as adjustments in:
* GraphColoring.hpp
* FlexibleSolver_impl.hpp
* PreconditionerFactory_impl.hpp
* CMakeLists.txt
* CMakeLists_files.txt

Regarding the length of the article; I think the section demonstrating that the current framework can yield significant speedup over conventional processors when running OPM Flow makes an important point, so I would like to keep it in some shape or form. We can maybe move `Appendix A`, and maybe into the repository somewhere, or just remove it? Additionally we can replace the section about the build system supporting both AMD and Nvidia with a few sentences in another section. These two changes would probably remove a few hundred words. Would that be properly in line with the typical JOSS format?
