### 2024-11-25 5290b38
adding some information regarding the connection
for CompWell
