+++ flowexperimental/comp/wells/CompWellInterface_impl.hpp
@@ -17,6 +17,8 @@
   along with OPM.  If not, see <http://www.gnu.org/licenses/>.
 */
 
+#include "CompConnectionData.hpp"
+
 #include <string>
 
 namespace Opm
@@ -24,11 +26,34 @@ namespace Opm
 template <typename TypeTag>
 CompWellInterface<TypeTag>::
 CompWellInterface(const Well& well,
-                  const int index_of_well)
+                  const int index_of_well,
+                  const std::vector<CompConnectionData<Scalar> >& well_connection_data)
     : well_ecl_(well)
     , index_of_well_(index_of_well)
     , reference_depth_(well.getRefDepth())
 {
+    number_of_connection_ = well_connection_data.size();
+    {
+        well_cells_.resize(number_of_connection_);
+        well_index_.resize(number_of_connection_);
+        saturation_table_number_.resize(number_of_connection_, 0);
+        int connection_idx = 0;
+        for (const auto& connection_data : well_connection_data) {
+            well_cells_[connection_idx] = connection_data.cell_index;
+            well_index_[connection_idx] = connection_data.connection_transmissibility_factor;
+            saturation_table_number_[connection_idx] = connection_data.satnum_id;
+            ++connection_idx;
+        }
+        // TODO: saturation_table_number
+    }
+
+}
+
+template <typename TypeTag>
+void
+CompWellInterface<TypeTag>::
+init() {
+    // more things to add here
 }
 
 } // end of namespace Opm
