### 2024-11-25 dd94f15
adding some information regarding the connection
for CompWell
