### 2024-11-25 2b0bb18
adding some information regarding the connection
for CompWell
