+++ opm/simulators/wells/BlackoilWellModelGeneric.hpp
@@ -375,7 +375,8 @@ protected:
 
     bool checkGroupHigherConstraints(const Group& group,
                                      DeferredLogger& deferred_logger,
-                                     const int reportStepIdx);
+                                     const int reportStepIdx,
+                                     const int max_number_of_group_switch);
 
     void updateAndCommunicateGroupData(const int reportStepIdx,
                                        const int iterationIdx,
@@ -599,7 +600,7 @@ protected:
 
     // Store maps of group name and new group controls for output
     std::map<std::string, std::vector<std::string>> switched_prod_groups_;
-    std::map<std::pair<std::string, Phase>, std::string> switched_inj_groups_;
+    std::map<std::pair<std::string, Phase>, std::vector<std::string>> switched_inj_groups_;
     // Store map of group name and close offending well for output
     std::map<std::string, std::pair<std::string, std::string>> closed_offending_wells_;
 
