### 2024-11-25 b32e2f6
Guard against small rates in fraction calculation
