### 2024-11-25 56897e6
Guard Hypre preconditioner creation with HAVE_HYPRE macro
