### 2024-11-25 #7524 @danielskatz
Thanks. And your suggestions for reducing the paper's word count seem reasonable.

Please feel free to make changes, then use the command `@editorialbot generate pdf` to make a new PDF.

editorialbot commands need to be the first entry in a new comment.

You can also use `editorialbot check repository` to check a bunch of things, including the word count for the paper.
