+++ opm/simulators/flow/Main.cpp
@@ -40,6 +40,10 @@
 #include <HYPRE_utilities.h>
 #endif
 
+#if HAVE_AMGX
+#include <amgx_c.h>
+#endif
+
 namespace Opm {
 
 Main::Main(int argc, char** argv, bool ownMPI)
@@ -95,6 +99,10 @@ Main::~Main()
 
 #if HAVE_HYPRE
     HYPRE_Finalize();
+#endif
+
+#if HAVE_AMGX
+    AMGX_SAFE_CALL(AMGX_finalize());
 #endif
 
     if (ownMPI_) {
@@ -179,6 +187,10 @@ void Main::initMPI()
 #else
     HYPRE_Init();
 #endif
+#endif
+
+#if HAVE_AMGX
+    AMGX_SAFE_CALL(AMGX_initialize());
 #endif
 }
 
