+++ opm/simulators/linalg/PreconditionerFactory_impl.hpp
@@ -53,6 +53,9 @@
 #include <opm/simulators/linalg/HyprePreconditioner.hpp>
 #endif
 
+#if HAVE_AMGX
+#include <opm/simulators/linalg/AmgxPreconditioner.hpp>
+#endif
 
 namespace Opm {
 
@@ -547,6 +550,16 @@ struct StandardPreconditioners<Operator, Dune::Amg::SequentialInformation> {
                     return getRebuildOnUpdateWrapper<Dune::Amg::FastAMG<O, V>>(op, crit, parms);
                 }
             });
+
+#if HAVE_AMGX
+            // Only add AMGX for scalar matrices
+            if constexpr (M::block_type::rows == 1 && M::block_type::cols == 1) {
+                F::addCreator("amgx", [](const O& op, const P& prm, const std::function<V()>&, std::size_t) {
+                    return std::make_shared<Amgx::AmgxPreconditioner<M, V, V>>(op.getmat(), prm);
+                });
+            }
+#endif
+
 #if HAVE_HYPRE
             // Only add Hypre for scalar matrices
             if constexpr (M::block_type::rows == 1 && M::block_type::cols == 1 &&
