+++ opm/simulators/linalg/AmgxPreconditioner.hpp
@@ -1,0 +1,318 @@
+/*
+  Copyright 2024 SINTEF AS
+  Copyright 2024 Equinor ASA
+
+  This file is part of the Open Porous Media project (OPM).
+
+  OPM is free software: you can redistribute it and/or modify
+  it under the terms of the GNU General Public License as published by
+  the Free Software Foundation, either version 3 of the License, or
+  (at your option) any later version.
+
+  OPM is distributed in the hope that it will be useful,
+  but WITHOUT ANY WARRANTY; without even the implied warranty of
+  MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
+  GNU General Public License for more details.
+
+  You should have received a copy of the GNU General Public License
+  along with OPM.  If not, see <http://www.gnu.org/licenses/>.
+*/
+
+#ifndef OPM_AMGX_PRECONDITIONER_HEADER_INCLUDED
+#define OPM_AMGX_PRECONDITIONER_HEADER_INCLUDED
+
+#include <opm/common/ErrorMacros.hpp>
+#include <opm/common/TimingMacros.hpp>
+#include <opm/simulators/linalg/PreconditionerWithUpdate.hpp>
+#include <opm/simulators/linalg/PropertyTree.hpp>
+
+#include <dune/common/fmatrix.hh>
+#include <dune/istl/bcrsmatrix.hh>
+
+#include <amgx_c.h>
+
+#include <vector>
+
+namespace Amgx {
+
+/**
+ * @brief Configuration structure for AMGX parameters.
+ *
+ * This structure holds the configuration parameters for the AMGX solver.
+ */
+struct AmgxConfig {
+    int determinism_flag = 0;
+    int print_grid_stats = 0;
+    int print_solve_stats = 0;
+    std::string solver = "AMG";
+    std::string algorithm = "CLASSICAL";
+    std::string interpolator = "D2";
+    std::string selector = "PMIS";
+    std::string smoother = "BLOCK_JACOBI";
+    int presweeps = 3;
+    int postsweeps = 3;
+    double strength_threshold = 0.5;
+    int max_iters = 1;
+
+    explicit AmgxConfig(const Opm::PropertyTree& prm) {
+        determinism_flag = prm.get<int>("determinism_flag", determinism_flag);
+        print_grid_stats = prm.get<int>("print_grid_stats", print_grid_stats);
+        print_solve_stats = prm.get<int>("print_solve_stats", print_solve_stats);
+        solver = prm.get<std::string>("solver", solver);
+        algorithm = prm.get<std::string>("algorithm", algorithm);
+        interpolator = prm.get<std::string>("interpolator", interpolator);
+        selector = prm.get<std::string>("selector", selector);
+        smoother = prm.get<std::string>("smoother", smoother);
+        presweeps = prm.get<int>("presweeps", presweeps);
+        postsweeps = prm.get<int>("postsweeps", postsweeps);
+        strength_threshold = prm.get<double>("strength_threshold", strength_threshold);
+        max_iters = prm.get<int>("max_iters", max_iters);
+    }
+
+    std::string toString() const {
+        return "config_version=2, "
+               "determinism_flag=" + std::to_string(determinism_flag) + ", "
+               "print_grid_stats=" + std::to_string(print_grid_stats) + ", "
+               "print_solve_stats=" + std::to_string(print_solve_stats) + ", "
+               "solver=" + solver + ", "
+               "algorithm=" + algorithm + ", "
+               "interpolator=" + interpolator + ", "
+               "selector=" + selector + ", "
+               "smoother=" + smoother + ", "
+               "presweeps=" + std::to_string(presweeps) + ", "
[…]
