+++ opm-simulators-prereqs.cmake
@@ -10,6 +10,7 @@ set (opm-simulators_CONFIG_VAR
   HAVE_OPENCL
   HAVE_OPENCL_HPP
   HAVE_AMGCL
+  HAVE_AMGX
   HAVE_VEXCL
   HAVE_ROCALUTION
   HAVE_ROCSPARSE
