### 2024-11-25 1c315b0
Fix unintentional short-circuit plus a couple of minor nits
