### 2024-11-25 6cf1f95
beginning working on CompWellStates
it is not clear what are needed there yet.
