### 2024-11-25 0ca28b2
beginning working on CompWellStates
it is not clear what are needed there yet.
