### 2024-11-25 5461bb6
beginning working on CompWellStates
it is not clear what are needed there yet.
