### 2024-11-26 7f6fbf5
shorten paper
