+++ paper/paper.md
@@ -46,70 +46,19 @@ Several standalone libraries GPU-accelerate linear algebra operations. Some also
 
 The `gpu-ISTL` library encompasses numerical algorithms, Dune adapters, and essential linear algebra components such as `GpuVector` and `GpuSparseMatrix` that leverage optimized libraries from Nvidia and AMD for efficient mathematical operations. Preconditioners such as Jacobi, ILU(0), and DILU are implemented with custom GPU kernels to enhance the performance of the bi-conjugate gradient stabilized (BiCGSTAB) method within `gpu-ISTL`. Adapter classes like `SolverAdapter` and `PreconditionerAdapter` allows mixing GPU and CPU solvers or preconditioners for ease of development, testing and validation. Moreover, `GpuOwnerOverlapCopy` extends the MPI functionality in Dune to support multi-GPU simulations, including CUDA-aware MPI for Nvidia cards to accelerate inter-process memory transfers. The library also provides capabilities for autotuning thread-block sizes in user-implemented GPU kernels. Figure 1 contains a simplified class diagram with an overview of some of the components that constitute `gpu-ISTL`. Furthermore, Figure 2 provides a sequence diagram for the call stack of a single linear solve in OPM Flow using the gpu-ISTL framework. We note that the sequence diagram is for the MPI free version. In Figure 2 we denote the (indirect) calls to specific GPU kernels as an arrow from the corresponding `gpu-ISTL` classes to the right-hand red box representing the GPU.
 
-![Class diagram showing a simplified view of how `gpu-ISTL` is implemented. The coloured backgrounds indicate which namespace the classes belong to. Colous on individual classes correspond to what conceptual part of the simulator they are a part of. White represents linear solvers, red represents the classes enabling multiprocess simulations, orange represents the implementation linear solver preconditioners, and green represents general linear algebra functionality.](figures/mpiparallel_gpuistl.png)
+![Class diagram showing a simplified view of how `gpu-ISTL` is implemented. The colored backgrounds indicate which namespace the classes belong to. Colors on individual classes correspond to what conceptual part of the simulator they are a part of. White represents linear solvers, red represents the classes enabling multiprocess simulations, orange represents the implementation linear solver preconditioners, and green represents general linear algebra functionality.](figures/mpiparallel_gpuistl.png)
 
 ![A sequence diagram for a single linear solver in OPM Flow using the `gpu-ISTL` framework. For readability, components from OPM are marked in green, components from DUNE in yellow and components from `gpu-ISTL` are marked in red. The classes GpuVector, GpuSparseMatrix and GpuILU0 of `gpu-ISTL` will (through some indirection) call GPU kernels, which is visualized as a big red block on the right-hand side of the diagram.](figures/timeloopgpuistlproper.png)
 
-# GPU Vendor support
-
[…]
