### 2024-11-26 60153a9
working on the initialization of the CompWellState
