### 2024-11-26 5e22db4
working on the initialization of the CompWellState
