### 2024-11-26 187ef57
working on the initialization of the CompWellState
