### 2024-11-26 e4403ba
working on the initialization of the CompWellState
