### 2024-11-26 6fa9c25
Move Hypre initialization and finalization out to Main
