### 2024-11-27 0238f45
Use formattedErrors() and simplify.
No longer calling dump() means we avoid the extra dumping to stderr of
the warnings and errors.
