+++ opm/simulators/flow/GenericCpGridVanguard.cpp
@@ -463,17 +463,6 @@ void GenericCpGridVanguard<ElementMapper,GridView,Scalar>::doCreateGrids_(Eclips
 
     cartesianIndexMapper_ = std::make_unique<CartesianIndexMapper>(*grid_);
 
-    // --- Add LGRs and update Leaf Grid View ---
-    // Check if input file contains Lgrs.
-    const auto& lgrs = eclState.getLgrs();
-    const auto lgrsSize = lgrs.size();
-    // If there are lgrs, create the grid with them, and update the leaf grid view.
-    if (lgrsSize)
-    {
-        OpmLog::info("\nAdding LGRs to the grid and updating its leaf grid view");
-        this->addLgrsUpdateLeafView(lgrs, lgrsSize, *(this->grid_));
-    }
-
 #if HAVE_MPI
     {
         const bool has_numerical_aquifer = eclState.aquifer().hasNumericalAquifer();
@@ -509,7 +498,7 @@ void GenericCpGridVanguard<ElementMapper,GridView,Scalar>::doCreateGrids_(Eclips
     }
 #endif
 
-    // --- Copy grid with LGRs to equilGrid_ ---
+    // --- Copy grid to equilGrid_ ---
     // We use separate grid objects: one for the calculation of the initial
     // condition via EQUIL and one for the actual simulation. The reason is
     // that the EQUIL code is allergic to distributed grids and the
@@ -518,14 +507,36 @@ void GenericCpGridVanguard<ElementMapper,GridView,Scalar>::doCreateGrids_(Eclips
     //
     // After loadbalance, grid_ will contain a global and distribute view.
     // equilGrid_ being a shallow copy only the global view.
-    if (mpiRank == 0)
+    //
+    // --- Add LGRs  ---
+    // Check if input file contains Lgrs.
+    const auto& lgrs = eclState.getLgrs();
+    const auto lgrsSize = lgrs.size();
+    if (lgrsSize == 0)
     {
-        equilGrid_.reset(new Dune::CpGrid(*grid_));
-        equilCartesianIndexMapper_ = std::make_unique<CartesianIndexMapper>(*equilGrid_);
+        if (mpiRank == 0)
+        {
+            equilGrid_.reset(new Dune::CpGrid(*grid_));
+            equilCartesianIndexMapper_ = std::make_unique<CartesianIndexMapper>(*equilGrid_);
 
-        eclState.reset_actnum(UgGridHelpers::createACTNUM(*grid_));
+            eclState.reset_actnum(UgGridHelpers::createACTNUM(*grid_));
+            eclState.set_active_indices(this->grid_->globalCell());
+        }
+    }
+    else { // If there are lgrs, create the grid with them, and update the leaf grid view.
+        equilGrid_.reset(new Dune::CpGrid(*grid_));
+        if (lgrsSize)
+        {
+            OpmLog::info("\nAdding LGRs to the equilGrid and updating its leaf grid view");
+           
+            addLgrsUpdateLeafView(lgrs, lgrsSize, *equilGrid_);
+        }
+        equilCartesianIndexMapper_ = std::make_unique<CartesianIndexMapper>(*equilGrid_);
+        
+        eclState.reset_actnum(UgGridHelpers::createACTNUM(*equilGrid_));
         eclState.set_active_indices(this->grid_->globalCell());
     }
+    
 
     {
         auto size = removed_cells.size();
