### 2024-11-28 32dd0a1
beginning working on the primvary variable and assembly
