+++ flowexperimental/comp/wells/SingleCompWellState_impl.hpp
@@ -26,8 +26,8 @@ ConnectionData(std::size_t num_connection,
                std::size_t num_phases,
                std::size_t num_components)
   : pressure(num_connection)
-  , phase_rates(num_connection * num_phases)
-  , reservoir_rates(num_connection * num_phases)
+  , surface_phase_rates(num_connection * num_phases)
+  , reservoir_phase_rates(num_connection * num_phases)
   , total_molar_fractions(num_connection * num_components)
   , tranmissibility_factor(num_connection)
   , satnum_id(num_connection)
@@ -58,13 +58,13 @@ SingleCompWellState(const std::string& well_name,
                     const std::vector<CompConnectionData<Scalar>>& connections,
                     bool is_producer)
    : name(well_name)
-   , pu(phase_usage)
+   , phase_usage(phase_usage)
    , producer(is_producer)
    , surface_phase_rates(phase_usage.num_phases)
    , phase_fractions(phase_usage.num_phases)
    , reservoir_phase_rates(phase_usage.num_phases)
    , total_molar_fractions(comp_config.numComps())
-   , connection_data(connections.size(), phase_usage.num_phases, comp_config.numComps())
+   , connection_data(connections, phase_usage, comp_config)
 {
 
 }
@@ -72,8 +72,7 @@ SingleCompWellState(const std::string& well_name,
 template <typename Scalar>
 void SingleCompWellState<Scalar>::
 update_injector_targets(const Well& well,
-                        const SummaryState& st)
-{
+                        const SummaryState& st) {
     const auto& inj_controls = well.injectionControls(st);
     const auto cmode_is_bhp = (inj_controls.cmode == Well::InjectorCMode::BHP);
     assert(cmode_is_bhp && "Only BHP control mode is supported for now");
@@ -89,6 +88,30 @@ update_injector_targets(const Well& well,
     assert(this->total_molar_fractions.size() == inj_composition.size());
     // TODO: this might not be correct when crossing flow is involved
     this->total_molar_fractions = inj_composition;
+
+    // we initialize all open wells with a rate to avoid singularities
+    Scalar inj_surf_rate = 10.0 * Opm::unit::cubic(Opm::unit::meter) / Opm::unit::day;
+    if (inj_controls.cmode == Well::InjectorCMode::RATE) {
+        inj_surf_rate = inj_controls.surface_rate;
+    }
+
+    switch (inj_controls.injector_type) {
+        case InjectorType::WATER:
+            assert(phase_usage.phase_used[BlackoilPhases::Aqua]);
+            this->surface_phase_rates[phase_usage.phase_pos[BlackoilPhases::Aqua]] = inj_surf_rate;
+            break;
+        case InjectorType::GAS:
+            assert(phase_usage.phase_used[BlackoilPhases::Vapour]);
+            this->surface_phase_rates[phase_usage.phase_pos[BlackoilPhases::Vapour]] = inj_surf_rate;
+            break;
+        case InjectorType::OIL:
+            assert(phase_usage.phase_used[BlackoilPhases::Liquid]);
+            this->surface_phase_rates[phase_usage.phase_pos[BlackoilPhases::Liquid]] = inj_surf_rate;
+            break;
+        case InjectorType::MULTI:
+            // Not currently handled, keep zero init.
+            break;
+    }
 }
 
 template <typename Scalar>
@@ -98,14 +121,41 @@ update_producer_targets(const Well& well,
 {
     const auto& prod_controls = well.productionControls(st);
 
+    const auto cmode_is_undefined = (prod_controls.cmode == Well::ProducerCMode::CMODE_UNDEFINED);
[…]
