### 2024-11-28 193e66b
beginning working on the primvary variable and assembly
