### 2024-11-28 c9f6f8f
beginning working on the primvary variable and assembly
