### 2024-11-28 8f1f79e
beginning working on the primvary variable and assembly
