### 2024-11-28 4bb50eb
Simply return thpres_ vector as restart vector
