### 2024-11-28 4555716
Ensure all processors have the same thpres_ array also when we do not restart
