### 2024-11-29 3966d0f
For convective mixing, store the saturated dissolution factor.
