### 2024-11-29 67765a3
fixed: well model could deadlock
if umfpack encounters a singular matrix on any process,
the exception propagated up across communication boundaries
leading to deadlock as the other processes tried to do global
reductions
