### 2024-12-01 297523e
Enable building without MPI
