+++ opm/simulators/timestepping/AdaptiveTimeStepping.hpp
@@ -29,6 +29,9 @@
 #include <opm/simulators/timestepping/TimeStepControlInterface.hpp>
 
 #if HAVE_MPI
+#define RESERVOIR_COUPLING_ENABLED
+#endif
+#ifdef RESERVOIR_COUPLING_ENABLED
 #include <opm/simulators/flow/ReservoirCoupling.hpp>
 #include <opm/simulators/flow/ReservoirCouplingMaster.hpp>
 #include <opm/simulators/flow/ReservoirCouplingSlave.hpp>
@@ -88,8 +91,9 @@ class AdaptiveTimeStepping
 {
 private:
     using Scalar = GetPropType<TypeTag, Properties::Scalar>;
+#ifdef RESERVOIR_COUPLING_ENABLED
     using TimePoint = ReservoirCoupling::TimePoint;
-
+#endif
     template <class Solver>
     class SolutionTimeErrorSolverWrapper : public RelativeChangeInterface
     {
@@ -126,10 +130,10 @@ private:
         void maybeModifySuggestedTimeStepAtBeginningOfReportStep_(const double originalTimeStep);
         bool maybeUpdateTuning_(double elapsed, double dt, int sub_step_number) const;
         double maxTimeStep_() const;
+        SimulatorReport runStepOriginal_();
+#ifdef RESERVOIR_COUPLING_ENABLED
         ReservoirCouplingMaster& reservoirCouplingMaster_();
         ReservoirCouplingSlave& reservoirCouplingSlave_();
-        SimulatorReport runStepOriginal_();
-#if HAVE_MPI
         SimulatorReport runStepReservoirCouplingMaster_();
         SimulatorReport runStepReservoirCouplingSlave_();
 #endif
@@ -214,8 +218,10 @@ public:
     bool operator==(const AdaptiveTimeStepping<TypeTag>& rhs);
 
     static void registerParameters();
+#ifdef RESERVOIR_COUPLING_ENABLED
     void setReservoirCouplingMaster(ReservoirCouplingMaster *reservoir_coupling_master);
     void setReservoirCouplingSlave(ReservoirCouplingSlave *reservoir_coupling_slave);
+#endif
     void setSuggestedNextStep(const double x);
     double suggestedNextStep() const;
 
@@ -273,7 +279,7 @@ protected:
 
     //! < shut problematic wells when time step size in days are less than this
     double min_time_step_before_shutting_problematic_wells_;
-#if HAVE_MPI
+#ifdef RESERVOIR_COUPLING_ENABLED
     ReservoirCouplingMaster *reservoir_coupling_master_ = nullptr;
     ReservoirCouplingSlave *reservoir_coupling_slave_ = nullptr;
 #endif
