+++ opm/simulators/flow/FlowMain.hpp
@@ -32,6 +32,9 @@
 #include <opm/simulators/flow/FlowUtils.hpp>
 #include <opm/simulators/flow/SimulatorFullyImplicitBlackoil.hpp>
 
+#if HAVE_MPI
+#define RESERVOIR_COUPLING_ENABLED
+#endif
 #if HAVE_DUNE_FEM
 #include <dune/fem/misc/mpimanager.hh>
 #else
@@ -360,7 +363,11 @@ namespace Opm {
         // Callback that will be called from runSimulatorInitOrRun_().
         int runSimulatorRunCallback_()
         {
+#ifdef RESERVOIR_COUPLING_ENABLED
             SimulatorReport report = simulator_->run(*simtimer_, this->argc_, this->argv_);
+#else
+            SimulatorReport report = simulator_->run(*simtimer_);
+#endif
             runSimulatorAfterSim_(report);
             return report.success.exit_status;
         }
@@ -368,7 +375,11 @@ namespace Opm {
         // Callback that will be called from runSimulatorInitOrRun_().
         int runSimulatorInitCallback_()
         {
+#ifdef RESERVOIR_COUPLING_ENABLED
             simulator_->init(*simtimer_, this->argc_, this->argv_);
+#else
+            simulator_->init(*simtimer_);
+#endif
             return EXIT_SUCCESS;
         }
 

[…]
