### 2024-12-01 402bb85
Enable building without MPI
