### 2024-12-01 8698e15
Enable building without MPI
