+++ opm/simulators/flow/Main.cpp
@@ -61,7 +61,9 @@ namespace Opm {
 Main::Main(int argc, char** argv, bool ownMPI)
     : argc_(argc), argv_(argv), ownMPI_(ownMPI)
 {
+#if HAVE_MPI
     maybeSaveReservoirCouplingSlaveLogFilename_();
+#endif
     if (ownMPI_) {
         initMPI();
     }
@@ -145,6 +147,7 @@ Main::~Main()
 #endif
 }
 
+#if HAVE_MPI
 void Main::maybeSaveReservoirCouplingSlaveLogFilename_()
 {
     // If first command line argument is "--slave-log-file=<filename>",
@@ -167,7 +170,8 @@ void Main::maybeSaveReservoirCouplingSlaveLogFilename_()
         }
     }
 }
-
+#endif
+#if HAVE_MPI
 void Main::maybeRedirectReservoirCouplingSlaveOutput_() {
     if (!this->reservoirCouplingSlaveOutputFilename_.empty()) {
         std::string filename = this->reservoirCouplingSlaveOutputFilename_
@@ -188,6 +192,7 @@ void Main::maybeRedirectReservoirCouplingSlaveOutput_() {
         close(fd);
     }
 }
+#endif
 
 void Main::setArgvArgc_(const std::string& filename)
 {
@@ -221,9 +226,9 @@ void Main::initMPI()
     FlowGenericVanguard::setCommunication(std::make_unique<Parallel::Communication>());
 
     handleTestSplitCommunicatorCmdLine_();
-    maybeRedirectReservoirCouplingSlaveOutput_();
 
 #if HAVE_MPI
+    maybeRedirectReservoirCouplingSlaveOutput_();
     if (test_split_comm_ && FlowGenericVanguard::comm().size() > 1) {
         int world_rank = FlowGenericVanguard::comm().rank();
         int color = (world_rank == 0);
