### 2024-12-01 a500868
Enable building without MPI
