+++ opm/simulators/flow/BlackoilModel.hpp
@@ -134,7 +134,7 @@ struct EnableDispersion<TypeTag, TTag::FlowProblem>
 
 template<class TypeTag>
 struct EnableConvectiveMixing<TypeTag, TTag::FlowProblem>
-{ static constexpr bool value = true; };
+{ static constexpr bool value = false; };
 
 template<class TypeTag>
 struct WellModel<TypeTag, TTag::FlowProblem>
