### 2024-12-02 5ae50c9
Simplify storage of communicators
We don't need unique ptrs for the communicators. These are just simple
integers that can be copied into a std::vector.
