### 2024-12-02 92e0d2e
-made more flexible structure for endStep - added numOverlap to flowgeneric
