### 2024-12-02 506ca46
-made more flexible structure for endStep - added numOverlap to flowgeneric
