+++ opm/simulators/flow/FlowGenericVanguard.hpp
@@ -262,7 +262,7 @@ public:
     bool addCorners() const
     { return addCorners_; }
 
-    bool numOverlap() const
+    int numOverlap() const
     { return numOverlap_; }
 
     /*!
