### 2024-12-02 e9ae7cd
-made more flexible structure for endStep - added numOverlap to flowgeneric
