### 2024-12-02 0290fd0
Add support for partial resetup with AMGX
