### 2024-12-02 ebde3cf
Eliminate TimePoint class
Make fuzzy comparison between two dates explicit.
