+++ opm/simulators/timestepping/AdaptiveTimeStepping.hpp
@@ -95,9 +95,6 @@ class AdaptiveTimeStepping
 {
 private:
     using Scalar = GetPropType<TypeTag, Properties::Scalar>;
-#ifdef RESERVOIR_COUPLING_ENABLED
-    using TimePoint = ReservoirCoupling::TimePoint;
-#endif
     template <class Solver>
     class SolutionTimeErrorSolverWrapper : public RelativeChangeInterface
     {
