### 2024-12-02 cd9fb6c
Eliminate TimePoint class
Make fuzzy comparison between two dates explicit.
