+++ opm/simulators/flow/ReservoirCoupling.hpp
@@ -37,157 +37,38 @@ enum class MessageTag : int {
     MasterGroupNamesSize,
 };
 
-// This class represents a time point.
-// It is currently used to represent an epoch time (a double value in seconds since the epoch),
-// or an elapsed time (a double value in seconds since the start of the simulation).
-// To avoid numerical issues when adding or subtracting time points and then later comparing
-// for equality with for example a given report date, we use a tolerance value.
-class TimePoint {
-private:
-    double time;
-    // TODO: Epoch values often lies in the range of [1e9,1e11], so a tolerance value of 1e-10
-    //     might be a little too small. However, for elapsed time values, the range is often
-    //     in the range of [0, 1e8], so a tolerance value of 1e-10 should be sufficient.
-    // NOTE: 1 nano-second = 1e-9 seconds
-    static constexpr double tol = 1e-10; // Tolerance value
-
-public:
-    TimePoint() : time(0.0) {}
-    explicit TimePoint(double t) : time(t) {}
-    TimePoint(const TimePoint& other) : time(other.time) {}
-
-    // Assignment operator for double
-    TimePoint& operator=(double t) {
-        time = t;
-        return *this;
-    }
-
-    // Copy assignment operator
-    TimePoint& operator=(const TimePoint& other) {
-        if (this != &other) {
-            time = other.time;
-        }
-        return *this;
-    }
-
-    double getTime() const { return time; }
-
-    // Equality operator
-    bool operator==(const TimePoint& other) const {
-        return std::abs(time - other.time) < tol;
-    }
-
-    // Inequality operator
-    bool operator!=(const TimePoint& other) const {
-        return !(*this == other);
-    }
-
-    // Less than operator
-    bool operator<(const TimePoint& other) const {
-        return (time < other.time) && !(*this == other);
-    }
-
-    // Comparison operator: double < TimePoint
-    friend bool operator<(double lhs, const TimePoint& rhs) {
-        return lhs < rhs.time;
-    }
-
-   // Comparison operator: TimePoint < double
-    bool operator<(double rhs) const {
-        return time < rhs;
-    }
-
-    // Less than or equal to operator
-    bool operator<=(const TimePoint& other) const {
-        return (time < other.time) || (*this == other);
-    }
-
-    // Comparison operator: double <= TimePoint
-    friend bool operator<=(double lhs, const TimePoint& rhs) {
-        return lhs <= rhs.time;
-    }
-
-    // Comparison operator: TimePoint <= double
-    bool operator<=(double rhs) const {
-        return time <= rhs;
-    }
-
-    // Greater than operator
-    bool operator>(const TimePoint& other) const {
-        return (time > other.time) && !(*this == other);
-    }
-
-    // Comparison operator: double > TimePoint
-    friend bool operator>(double lhs, const TimePoint& rhs) {
-        return lhs > rhs.time;
-    }
-
-    // Comparison operator: TimePoint > double
-    bool operator>(double rhs) const {
-        return time > rhs;
-    }
-
-    // Greater than or equal to operator
-    bool operator>=(const TimePoint& other) const {
-        return (time > other.time) || (*this == other);
-    }
-
-    // Comparison operator: TimePoint >= double
-    bool operator>=(double rhs) const {
-        return time >= rhs;
-    }
-
-    // Comparison operator: double >= TimePoint
-    friend bool operator>=(double lhs, const TimePoint& rhs) {
-        return lhs >= rhs.time;
-    }
-
-    // Addition operator: TimePoint + TimePoint (summing their times)
-    TimePoint operator+(const TimePoint& other) const {
-        return TimePoint(time + other.time);
-    }
-
-    // Addition operator: TimePoint + double
-    TimePoint operator+(double delta) const {
[…]
