### 2024-12-02 879fa72
Eliminate TimePoint class
Make fuzzy comparison between two dates explicit.
