### 2024-12-03 3f628f0
Use regular OpenMP for loop, then iterate over chunks inside.
Avoiding the mutex-based ThreadedEntityIterator gives a good speedup,
the overhead from the single extra loop over the grid for initialization
should be small in comparison.
