### 2024-12-03 82a2b39
-add blockoutput of stress
