### 2024-12-03 473a16f
-add blockoutput of stress
