### 2024-12-03 45883f3
Conversion of std::time_t to double
Clarify the limits for conversion of std::time_t to double
