### 2024-12-03 c653d18
Clarify how the timestep is selected
Clarify how the master timestep is computed based on the slaves next
report date or the slaves start date.
