### 2024-12-04 738eb76
Only restrict oscillation for iter > nupcol
