### 2024-12-04 3b67d6d
Small fixes for hypre integration (#1)
* use appropriate hypre init function

* properly handle hypre without device support

* don't add hypre preconditioner to factory if scalar type does not match
