### 2024-12-04 90136ff
Propagate need for recreating solver properly for two level methods
