### 2024-12-04 e60aa7d
Propagate need for recreating solver properly for two level methods
