+++ CMakeLists_files.cmake
@@ -492,6 +492,10 @@ if(HDF5_FOUND)
   list(APPEND TEST_SOURCE_FILES tests/test_HDF5Serializer.cpp)
 endif()
 
+if(HYPRE_FOUND)
+  list(APPEND TEST_SOURCE_FILES tests/test_hyprepreconditioner.cpp)
+endif()
+
 list (APPEND TEST_DATA_FILES
   tests/equil_base.DATA
   tests/equil_capillary.DATA
