+++ CMakeLists.txt
@@ -726,12 +726,9 @@ if(USE_HYPRE)
   find_package(HYPRE)
   if(HYPRE_FOUND)
     set(HAVE_HYPRE 1)
+    target_link_libraries(opmsimulators PUBLIC HYPRE::HYPRE)
   else()
     message(WARNING "Hypre requested but not found. Continuing without Hypre support.")
     set(USE_HYPRE OFF)
   endif()
-endif()
-
-if(HYPRE_FOUND)
-  target_link_libraries(opmsimulators PUBLIC HYPRE::HYPRE)
 endif()
