+++ opm/simulators/linalg/gpuistl/GpuBuffer.cpp
@@ -176,6 +176,19 @@ GpuBuffer<T>::copyFromHost(const T* dataPointer, size_t numberOfElements)
 
 template <class T>
 void
+GpuBuffer<T>::copyFromHost(const T* dataPointer, size_t numberOfElements, cudaStream_t stream)
+{
+    if (numberOfElements > size()) {
+        OPM_THROW(std::runtime_error,
+                  fmt::format("Requesting to copy too many elements. buffer has {} elements, while {} was requested.",
+                              size(),
+                              numberOfElements));
+    }
+    OPM_GPU_SAFE_CALL(cudaMemcpyAsync(data(), dataPointer, numberOfElements * sizeof(T), cudaMemcpyHostToDevice, stream));
+}
+
+template <class T>
+void
 GpuBuffer<T>::copyToHost(T* dataPointer, size_t numberOfElements) const
 {
     assertSameSize(numberOfElements);
@@ -188,6 +201,14 @@ GpuBuffer<T>::copyFromHost(const std::vector<T>& data)
 {
     copyFromHost(data.data(), data.size());
 }
+
+template <class T>
+void
+GpuBuffer<T>::copyFromHost(const std::vector<T>& data, cudaStream_t stream)
+{
+    copyFromHost(data.data(), data.size(), stream);
+}
+
 template <class T>
 void
 GpuBuffer<T>::copyToHost(std::vector<T>& data) const
