### 2024-12-05 c43ed3e
add first version of cudagraph for the apply
