+++ opm/simulators/linalg/gpuistl/GpuDILU.hpp
@@ -25,7 +25,8 @@
 #include <opm/simulators/linalg/gpuistl/GpuSparseMatrix.hpp>
 #include <opm/simulators/linalg/gpuistl/detail/kernel_enums.hpp>
 #include <vector>
-
+#include <map>
+#include <utility>
 
 
 namespace Opm::gpuistl
@@ -153,6 +154,16 @@ private:
     int m_lowerSolveThreadBlockSize = -1;
     int m_moveThreadBlockSize = -1;
     int m_DILUFactorizationThreadBlockSize = -1;
+
+    std::map<std::pair<field_type*, const field_type*>, cudaGraph_t> m_apply_graphs;
+    bool m_update_graph_captured {false};
+    cudaGraph_t m_update_graph;
+    std::map<std::pair<field_type*, const field_type*>, cudaGraphExec_t> m_executableGraphs;
+    cudaStream_t stream;
+    cudaGraph_t graph;
+    cudaGraphExec_t instance;
+    bool m_cudagraphInitialized {false};
+    cudaGraph_t m_cudagraphInstance;
 };
 } // end namespace Opm::gpuistl
 
