### 2024-12-05 d394a1f
use cudagraphs in gpu DILU and ILU0 apply
is also supported in HIP, though not does not
seem to affect performance in any clear way.
1.1 to 1.2 speedup in Nvidia GPUs.
