### 2024-12-05 4894c9f
add first version of cudagraph for the apply
