### 2024-12-05 66fa971
use cudagraphs in gpu DILU and ILU0 apply expected to give performance gain on current nvidia cards, no significant change on AMD cards for cases of reasonable size
uses new GPU convenence objects
