### 2024-12-05 d8a7875
fixed: always calculate initial FIP
this can be used by various Schedule keywords even if FIP is not in RPTSOL
