### 2024-12-05 1536de0
Document the HyprePreconditioner class
