+++ opm/simulators/linalg/HyprePreconditioner.hpp
@@ -39,22 +39,30 @@
 
 namespace Hypre {
 
-/// Wrapper for Hypre's BoomerAMG preconditioner
+/**
+ * @brief Wrapper for Hypre's BoomerAMG preconditioner.
+ *
+ * This class provides an interface to the BoomerAMG preconditioner from the Hypre library.
+ * It is designed to work with matrices, update vectors, and defect vectors specified by the template parameters.
+ *
+ * @tparam M The matrix type the preconditioner is for.
+ * @tparam X The type of the update vector.
+ * @tparam Y The type of the defect vector.
+ */
 template<class M, class X, class Y>
 class HyprePreconditioner : public Dune::PreconditionerWithUpdate<X,Y> {
 public:
-    //! \brief The matrix type the preconditioner is for
-    using matrix_type = M;
-    //! \brief The domain type of the preconditioner
-    using domain_type = X;
-    //! \brief The range type of the preconditioner
-    using range_type = Y;
-    //! \brief The field type of the preconditioner
-    using field_type = typename X::field_type;
 
-    // Constructor
+    /**
+     * @brief Constructor for the HyprePreconditioner class.
+     *
+     * Initializes the preconditioner with the given matrix and property tree.
+     *
+     * @param A The matrix for which the preconditioner is constructed.
+     * @param prm The property tree containing configuration parameters.
+     */
     HyprePreconditioner (const M& A, const Opm::PropertyTree prm)
-        : A_(A), prm_(prm)
+        : A_(A)
     {
         OPM_TIMEBLOCK(prec_construct);
 
@@ -64,7 +72,7 @@ public:
             OPM_THROW(std::runtime_error, "HyprePreconditioner is currently only implemented for sequential runs");
         }
 
-        use_gpu_ = prm_.get<bool>("use_gpu", false);
+        use_gpu_ = prm.get<bool>("use_gpu", false);
 
         // Set memory location and execution policy
 #if HYPRE_USING_CUDA || HYPRE_USING_HIP
@@ -89,13 +97,13 @@ public:
         HYPRE_BoomerAMGCreate(&solver_);
 
         // Set parameters from property tree with defaults
-        HYPRE_BoomerAMGSetPrintLevel(solver_, prm_.get<int>("print_level", 0));
-        HYPRE_BoomerAMGSetMaxIter(solver_, prm_.get<int>("max_iter", 1));
-        HYPRE_BoomerAMGSetStrongThreshold(solver_, prm_.get<double>("strong_threshold", 0.5));
-        HYPRE_BoomerAMGSetAggTruncFactor(solver_, prm_.get<double>("agg_trunc_factor", 0.3));
-        HYPRE_BoomerAMGSetInterpType(solver_, prm_.get<int>("interp_type", 6));
-        HYPRE_BoomerAMGSetMaxLevels(solver_, prm_.get<int>("max_levels", 15));
-        HYPRE_BoomerAMGSetTol(solver_, prm_.get<double>("tolerance", 0.0));
+        HYPRE_BoomerAMGSetPrintLevel(solver_, prm.get<int>("print_level", 0));
+        HYPRE_BoomerAMGSetMaxIter(solver_, prm.get<int>("max_iter", 1));
+        HYPRE_BoomerAMGSetStrongThreshold(solver_, prm.get<double>("strong_threshold", 0.5));
+        HYPRE_BoomerAMGSetAggTruncFactor(solver_, prm.get<double>("agg_trunc_factor", 0.3));
+        HYPRE_BoomerAMGSetInterpType(solver_, prm.get<int>("interp_type", 6));
+        HYPRE_BoomerAMGSetMaxLevels(solver_, prm.get<int>("max_levels", 15));
[…]
