### 2024-12-05 0821f5f
Document the HyprePreconditioner class
