+++ opm/simulators/linalg/HyprePreconditioner.hpp
@@ -39,20 +39,28 @@
 
 namespace Hypre {
 
-/// Wrapper for Hypre's BoomerAMG preconditioner
+/**
+ * @brief Wrapper for Hypre's BoomerAMG preconditioner.
+ *
+ * This class provides an interface to the BoomerAMG preconditioner from the Hypre library.
+ * It is designed to work with matrices, update vectors, and defect vectors specified by the template parameters.
+ *
+ * @tparam M The matrix type the preconditioner is for.
+ * @tparam X The type of the update vector.
+ * @tparam Y The type of the defect vector.
+ */
 template<class M, class X, class Y>
 class HyprePreconditioner : public Dune::PreconditionerWithUpdate<X,Y> {
 public:
-    //! \brief The matrix type the preconditioner is for
-    using matrix_type = M;
-    //! \brief The domain type of the preconditioner
-    using domain_type = X;
-    //! \brief The range type of the preconditioner
-    using range_type = Y;
-    //! \brief The field type of the preconditioner
-    using field_type = typename X::field_type;
 
-    // Constructor
+    /**
+     * @brief Constructor for the HyprePreconditioner class.
+     *
+     * Initializes the preconditioner with the given matrix and property tree.
+     *
+     * @param A The matrix for which the preconditioner is constructed.
+     * @param prm The property tree containing configuration parameters.
+     */
     HyprePreconditioner (const M& A, const Opm::PropertyTree prm)
         : A_(A), prm_(prm)
     {
@@ -141,7 +149,11 @@ public:
         update();
     }
 
-    // Destructor
+    /**
+     * @brief Destructor for the HyprePreconditioner class.
+     *
+     * Cleans up resources allocated by the preconditioner.
+     */
     ~HyprePreconditioner() {
         if (solver_) {
             HYPRE_BoomerAMGDestroy(solver_);
@@ -169,15 +181,36 @@ public:
         }
     }
 
+    /**
+     * @brief Updates the preconditioner with the current matrix values.
+     *
+     * This method should be called whenever the matrix values change.
+     */
     void update() override {
         OPM_TIMEBLOCK(prec_update);
         copyMatrixToHypre();
         HYPRE_BoomerAMGSetup(solver_, parcsr_A_, par_b_, par_x_);
     }
 
-    void pre(X& /*x*/, Y& /*b*/) override {
+    /**
+     * @brief Pre-processing step before applying the preconditioner.
+     *
+     * This method is currently a no-op.
+     *
+     * @param v The update vector.
+     * @param d The defect vector.
+     */
+    void pre(X& /*v*/, Y& /*d*/) override {
     }
 
+    /**
+     * @brief Applies the preconditioner to a vector.
+     *
+     * Performs one AMG V-cycle to solve the system.
+     *
+     * @param v The update vector.
+     * @param d The defect vector.
+     */
     void apply(X& v, const Y& d) override {
         OPM_TIMEBLOCK(prec_apply);
 
@@ -191,13 +224,30 @@ public:
         copyVectorFromHypre(v);
     }
 
-    void post(X& /*x*/) override {
+    /**
+     * @brief Post-processing step after applying the preconditioner.
+     *
+     * This method is currently a no-op.
+     *
+     * @param v The update vector.
+     */
+    void post(X& /*v*/) override {
     }
 
+    /**
+     * @brief Returns the solver category.
+     *
+     * @return The solver category, which is sequential.
+     */
[…]
