### 2024-12-05 095c045
Code improvment for the group control switching
Pass the controls as enums not strings
Avoid pair as key in map
