+++ opm/simulators/wells/BlackoilWellModel_impl.hpp
@@ -749,29 +749,42 @@ namespace Opm {
         // report group switching
         if (this->terminal_output_) {
 
-            for (const auto& [name, to] : this->switched_prod_groups_) {
+            for (const auto& [name, ctrls] : this->switched_prod_groups_) {
                 const Group::ProductionCMode& oldControl = this->prevWGState().group_state.production_control(name);
                 std::string from = Group::ProductionCMode2String(oldControl);
-                if (to.back() != from) {
+                std::string to = Group::ProductionCMode2String(ctrls.back());
+                if (ctrls.back() != oldControl) {
                     std::string msg = "    Production Group " + name
                     + " control mode changed from ";
                     msg += from;
-                    msg += " to " + to.back();
+                    msg += " to " + to;
                     local_deferredLogger.info(msg);
                 }
             }
-            for (const auto& [key, to] : this->switched_inj_groups_) {
-                const std::string& name = key.first;
-                const Opm::Phase& phase = key.second;
 
-                const Group::InjectionCMode& oldControl = this->prevWGState().group_state.injection_control(name, phase);
-                std::string from = Group::InjectionCMode2String(oldControl);
-                if (to.back() != from) {
-                    std::string msg = "    Injection Group " + name
-                    + " control mode changed from ";
-                    msg += from;
-                    msg += " to " + to.back();
-                    local_deferredLogger.info(msg);
+            for (const auto& [grname, grdata] : this->switched_inj_groups_) {
+                //const std::string& name = key.first;
+                //const Opm::Phase& phase = key.second;
+                const Phase all[] = {Phase::WATER, Phase::OIL, Phase::GAS};
+                for (Phase phase : all) {
+                    const auto& ctrls = grdata[static_cast<std::underlying_type_t<Phase>>(phase)];
+
+                    if (ctrls.empty()) {
+                        continue;
+                    }
+                    if ( !this->prevWGState().group_state.has_injection_control(grname, phase))
+                        continue;
+
+                    const Group::InjectionCMode& oldControl = this->prevWGState().group_state.injection_control(grname, phase);
+                    std::string from = Group::InjectionCMode2String(oldControl);
+                    std::string to = Group::InjectionCMode2String(ctrls.back());
+                    if (ctrls.back() != oldControl) {
+                        std::string msg = "    Injection Group " + grname
+                        + " control mode changed from ";
+                        msg += from;
+                        msg += " to " + to;
+                        local_deferredLogger.info(msg);
+                    }
                 }
             }
         }
