+++ tests/test_RestartSerialization.cpp
@@ -332,8 +332,9 @@ public:
         last_valid_wgstate_ = WGState<double>::serializationTestObject(dummy);
         nupcol_wgstate_ = WGState<double>::serializationTestObject(dummy);
         last_glift_opt_time_ = 5.0;
-        switched_prod_groups_ = {{"test4", {"test5", "test6"}}};
-        switched_inj_groups_ = {{{"test4", Phase::SOLVENT}, {"test5", "test6"}}};
+        switched_prod_groups_ = {{"test4", {Group::ProductionCMode::NONE, Group::ProductionCMode::ORAT}}};
+        const auto controls = {Group::InjectionCMode::NONE, Group::InjectionCMode::RATE, Group::InjectionCMode::RATE };
+        switched_inj_groups_ = {{"test4", {controls, {}, controls} }};
         closed_offending_wells_ = {{"test4", {"test5", "test6"}}};
     }
 
