+++ opm/simulators/wells/BlackoilWellModelGeneric.hpp
@@ -599,8 +599,8 @@ protected:
     bool wellStructureChangedDynamically_{false};
 
     // Store maps of group name and new group controls for output
-    std::map<std::string, std::vector<std::string>> switched_prod_groups_;
-    std::map<std::pair<std::string, Phase>, std::vector<std::string>> switched_inj_groups_;
+    std::map<std::string, std::vector<Group::ProductionCMode>> switched_prod_groups_;
+    std::map<std::string, std::array<std::vector<Group::InjectionCMode>, 3>> switched_inj_groups_;
     // Store map of group name and close offending well for output
     std::map<std::string, std::pair<std::string, std::string>> closed_offending_wells_;
 
