+++ opm/simulators/wells/BlackoilWellModelConstraints.hpp
@@ -74,8 +74,8 @@ public:
     bool updateGroupIndividualControl(const Group& group,
                                       const int reportStepIdx,
                                       const int max_number_of_group_switch,
-                                      std::map<std::pair<std::string,Phase>,std::vector<std::string>>& switched_inj,
-                                      std::map<std::string, std::vector<std::string>>& switched_prod,
+                                      std::map<std::string, std::array<std::vector<Group::InjectionCMode>, 3>>& switched_inj,
+                                      std::map<std::string, std::vector<Group::ProductionCMode>>& switched_prod,
                                       std::map<std::string, std::pair<std::string, std::string>>& closed_offending_wells,
                                       GroupState<Scalar>& group_state,
                                       WellState<Scalar>& well_state,
