+++ opm/simulators/wells/BlackoilWellModelConstraints.cpp
@@ -562,8 +562,8 @@ bool BlackoilWellModelConstraints<Scalar>::
 updateGroupIndividualControl(const Group& group,
                              const int reportStepIdx,
                              const int max_number_of_group_switch,
-                             std::map<std::pair<std::string,Phase>,std::vector<std::string>>& switched_inj,
-                             std::map<std::string, std::vector<std::string>>& switched_prod,
+                             std::map<std::string, std::array<std::vector<Group::InjectionCMode>, 3>>& switched_inj,
+                             std::map<std::string, std::vector<Group::ProductionCMode>>& switched_prod,
                              std::map<std::string, std::pair<std::string, std::string>>& closed_offending_wells,
                              GroupState<Scalar>& group_state,
                              WellState<Scalar>& well_state,
@@ -577,31 +577,32 @@ updateGroupIndividualControl(const Group& group,
             if (!group.hasInjectionControl(phase)) {
                 continue;
             }
-
             bool group_is_oscillating = false;
-            if (switched_inj.count({group.name(), phase}) > 0) {
-                for (const auto& key : switched_inj[{group.name(), phase}]) {
-                    if (std::count(switched_inj[{group.name(), phase}].begin(), switched_inj[{group.name(), phase}].end(), key) >= max_number_of_group_switch) {
-                        const size_t sz = switched_inj[{group.name(), phase}].size();
-                        // only output if the two last keys are not the same.
-                        if (key !=  switched_inj[{group.name(), phase}][sz-2]) {
-                            if (wellModel_.comm().rank() == 0) {
-                                std::ostringstream os;
-                                os << phase;
-                                const std::string msg =
-                                    fmt::format("Group control for {} injector group {} is oscillating. Group control kept at {}.",
-                                                std::move(os).str(),
-                                                group.name(),
-                                                key);
-                                deferred_logger.info(msg);
-                            }
-                            switched_inj[{group.name(), phase}].push_back(key);
-                        }
-                        group_is_oscillating = true;
-                        break;
+            if (auto groupPos = switched_inj.find(group.name()); groupPos != switched_inj.end()) {
+                auto& ctrls = groupPos->second[static_cast<std::underlying_type_t<Phase>>(phase)];
+                for (const auto& ctrl : ctrls) {
+                    if (std::count(ctrls.begin(), ctrls.end(), ctrl) < max_number_of_group_switch) {
+                        continue;
                     }
+
+                    if (ctrls.back() != *(ctrls.end() - 2)) {
+                        if (wellModel_.comm().rank() == 0 ) {
+                            std::ostringstream os;
+                            os << phase;
+                            const std::string msg =
+                                fmt::format("Group control for {} injector group {} is oscillating. Group control kept at {}.",
+                                            std::move(os).str(),
+                                            group.name(),
+                                            Group::InjectionCMode2String(ctrl));
+                            deferred_logger.info(msg);
+                        }
+                        ctrls.push_back(ctrl);
+                    }
+                    group_is_oscillating = true;
+                    break;
                 }
             }
+
             if (group_is_oscillating) {
                 continue;
             }
@@ -611,8 +612,8 @@ updateGroupIndividualControl(const Group& group,
                                                                             phase);
             if (changed_this.first != Group::InjectionCMode::NONE)
             {
-                switched_inj[{group.name(), phase}].push_back(
-                             Group::InjectionCMode2String(changed_this.first));
+                switched_inj[group.name()][static_cast<std::underlying_type_t<Phase>>(phase)].push_back(
+                             changed_this.first);
 
                 this->actionOnBrokenConstraints(group, changed_this.first, phase,
                                                 group_state, deferred_logger);
[…]

[…]
