### 2024-12-05 de95cd6
WIP in assembling of the well equations
