### 2024-12-05 0fd057b
WIP in assembling of the well equations
