### 2024-12-06 a758b79
added: support WCYCLE
