+++ opm/simulators/wells/GasLiftStage2.cpp
@@ -831,7 +831,8 @@ computeDelta(const std::string& well_name, bool add)
         // only get deltas for wells owned by this rank
         if (this->well_state_.wellIsOwned(well.indexOfWell(), well_name)) {
             const auto& well_ecl = well.wellEcl();
-            Scalar factor = well_ecl.getEfficiencyFactor();
+            Scalar factor = well_ecl.getEfficiencyFactor() *
+                            this->well_state_[well_name].efficiency_scaling_factor;
             auto& [delta_oil, delta_gas, delta_water, delta_alq] = delta;
             delta_oil = factor * (gi.new_oil_rate - state.oilRate());
             delta_gas = factor * (gi.new_gas_rate - state.gasRate());
@@ -1177,7 +1178,8 @@ computeDelta(const std::string& well_name)
         // only get deltas for wells owned by this rank
         if (this->parent.well_state_.wellIsOwned(well.indexOfWell(), well_name)) {
             const auto& well_ecl = well.wellEcl();
-            Scalar factor = well_ecl.getEfficiencyFactor();
+            Scalar factor = well_ecl.getEfficiencyFactor() *
+                            1.0;//this->well_state_[well_name].efficiency_scaling_factor;
             auto& [delta_oil, delta_gas, delta_water, delta_alq] = delta;
             delta_oil = factor * (gi.new_oil_rate - state.oilRate());
             delta_gas = factor * (gi.new_gas_rate - state.gasRate());

[…]
