### 2024-12-07 4074245
Fix typo in Equinor ASA
