### 2024-12-07 3313fc2
Mark GRUPSLAV and GRUPMAST as unsupported
Don't remove these entries from the map before reservoir coupling is
completely supported.
