+++ opm/simulators/utils/UnsupportedFlowKeywords.cpp
@@ -242,7 +242,9 @@ const KeywordValidation::UnsupportedKeywords& unsupportedKeywords()
         {"GRAVDRB", {true, std::nullopt}},
         {"GRAVDRM", {true, std::nullopt}},
         {"GRDREACH", {true, std::nullopt}},
+        {"GRUPMAST", {true, std::nullopt}},
         {"GRUPRIG", {true, std::nullopt}},
+        {"GRUPSLAV", {true, std::nullopt}},
         {"GRUPTARG", {true, std::nullopt}},
         {"GSATINJE", {true, std::nullopt}},
         {"GSEPCOND", {true, std::nullopt}},
