### 2024-12-07 e332db8
Use Dune::MPITraits to determine MPI datatype
Determine size of std::size_t correctly for all platforms using
Dune::MPITraits<std::size_t>::getType()
