### 2024-12-08 c9e78fe
Clearify that errhandler is a handle
It is safe to free the error handler after MPI_Comm_set_errhandler()
has been called
