### 2024-12-09 1c8bf2f
Remove duplicate headers
