### 2024-12-10 1180ad2
Allow for the third group switch
