### 2024-12-10 3daec4d
added option for edgconformal processing in flow (cpgrid and polygrid?)
