+++ opm/simulators/flow/FlowGenericVanguard.hpp
@@ -75,6 +75,7 @@ struct ActionParsingStrictness { static constexpr auto value = "normal"; };
 /// 0: simple, 1: Zoltan, 2: METIS, 3: Zoltan with a all cells of a well
 /// represented by one vertex in the graph, see GridEnums.hpp
 struct PartitionMethod { static constexpr int value = 3; };
+struct EdgeConformal { static constexpr bool value = false; };    
 
 struct SchedRestart{ static constexpr bool value = false; };
 struct SerialPartitioning{ static constexpr bool value = false; };
@@ -256,6 +257,8 @@ public:
     { return ownersFirst_; }
 
 #if HAVE_MPI
+    bool edgeConformal() const
+    { return edgeConformal_; }
     /*!
      * \brief Parameter deciding which partition method to use
      */
@@ -357,6 +360,7 @@ protected:
 
     bool ownersFirst_;
 #if HAVE_MPI
+    bool edgeConformal_;
     Dune::PartitionMethod partitionMethod_;
     bool serialPartitioning_;
     double imbalanceTol_;
