### 2024-12-10 f10384c
added option for edgconformal processing in flow (cpgrid and polygrid?)
