+++ opm/simulators/flow/FlowGenericVanguard.hpp
@@ -73,6 +73,7 @@ struct ActionParsingStrictness { static constexpr auto value = "normal"; };
 
  // 0: simple, 1: Zoltan, 2: METIS, see GridEnums.hpp
 struct PartitionMethod { static constexpr int value = 1; };
+struct EdgeConformal { static constexpr bool value = false; };    
 
 struct SchedRestart{ static constexpr bool value = false; };
 struct SerialPartitioning{ static constexpr bool value = false; };
@@ -254,6 +255,8 @@ public:
     { return ownersFirst_; }
 
 #if HAVE_MPI
+    bool edgeConformal() const
+    { return edgeConformal_; }
     /*!
      * \brief Parameter deciding which partition method to use
      */
@@ -349,6 +352,7 @@ protected:
 
     bool ownersFirst_;
 #if HAVE_MPI
+    bool edgeConformal_;
     Dune::PartitionMethod partitionMethod_;
     bool serialPartitioning_;
     double imbalanceTol_;
