### 2024-12-10 3158a38
added option for edgconformal processing in flow (cpgrid and polygrid?)
