+++ opm/simulators/flow/FlowGenericVanguard.cpp
@@ -113,30 +113,35 @@ FlowGenericVanguard::FlowGenericVanguard(SimulationModelParams&& params)
     edgeWeightsMethod_   = Dune::EdgeWeightMethod(Parameters::Get<Parameters::EdgeWeightsMethod>());
 
 #if HAVE_OPENCL || HAVE_ROCSPARSE || HAVE_CUDA
-        numJacobiBlocks_ = Parameters::Get<Parameters::NumJacobiBlocks>();
+    numJacobiBlocks_ = Parameters::Get<Parameters::NumJacobiBlocks>();
 #endif
 
-        ownersFirst_ = Parameters::Get<Parameters::OwnerCellsFirst>();
+    ownersFirst_ = Parameters::Get<Parameters::OwnerCellsFirst>();
+
 #if HAVE_MPI
-        numOverlap_ = Parameters::Get<Parameters::NumOverlap>();
-        addCorners_ = Parameters::Get<Parameters::AddCorners>();
-        partitionMethod_   = Dune::PartitionMethod(Parameters::Get<Parameters::PartitionMethod>());
-        serialPartitioning_ = Parameters::Get<Parameters::SerialPartitioning>();
-        zoltanParams_ = Parameters::Get<Parameters::ZoltanParams>();
+    numOverlap_ = Parameters::Get<Parameters::NumOverlap>();
+    addCorners_ = Parameters::Get<Parameters::AddCorners>();
+    edgeConformal_ = Parameters::Get<Parameters::EdgeConformal>();
+    partitionMethod_ = Dune::PartitionMethod(Parameters::Get<Parameters::PartitionMethod>());
+    serialPartitioning_ = Parameters::Get<Parameters::SerialPartitioning>();
+    zoltanParams_ = Parameters::Get<Parameters::ZoltanParams>();
 
-        metisParams_ = Parameters::Get<Parameters::MetisParams>();
+    metisParams_ = Parameters::Get<Parameters::MetisParams>();
 
-        externalPartitionFile_ = Parameters::Get<Parameters::ExternalPartition>();
+    externalPartitionFile_ = Parameters::Get<Parameters::ExternalPartition>();
 #endif
-        enableDistributedWells_ = Parameters::Get<Parameters::AllowDistributedWells>();
-        enableEclOutput_ = Parameters::Get<Parameters::EnableEclOutput>();
-        allow_splitting_inactive_wells_ = Parameters::Get<Parameters::AllowSplittingInactiveWells>();
-        ignoredKeywords_ = Parameters::Get<Parameters::IgnoreKeywords>();
-        int output_param = Parameters::Get<Parameters::EclOutputInterval>();
-        if (output_param >= 0) {
-            outputInterval_ = output_param;
-        }
-        useMultisegmentWell_ = Parameters::Get<Parameters::UseMultisegmentWell>();
+
+    enableDistributedWells_ = Parameters::Get<Parameters::AllowDistributedWells>();
+    enableEclOutput_ = Parameters::Get<Parameters::EnableEclOutput>();
+    allow_splitting_inactive_wells_ = Parameters::Get<Parameters::AllowSplittingInactiveWells>();
+    ignoredKeywords_ = Parameters::Get<Parameters::IgnoreKeywords>();
+
+    const int output_param = Parameters::Get<Parameters::EclOutputInterval>();
+    if (output_param >= 0) {
+        outputInterval_ = output_param;
+    }
+
+    useMultisegmentWell_ = Parameters::Get<Parameters::UseMultisegmentWell>();
 }
 
 FlowGenericVanguard::SimulationModelParams
