### 2024-12-10 c9513a3
added option for edgconformal processing in flow (cpgrid and polygrid?)
