### 2024-12-10 7cda902
added option for edgconformal processing in flow (cpgrid and polygrid?)
