### 2024-12-10 aafb5b9
added option for edgconformal processing in flow (cpgrid and polygrid?)
