### 2024-12-11 39441d1
Made OpenCL optional.
