### 2024-12-11 dd9dd75
WIP not compiling trying to make a very simple version of the dynamic object being passed around during the update so that the interface looks like what it will be later
