+++ opm/simulators/flow/FIBlackoilModel.hpp
@@ -55,6 +55,7 @@ class FIBlackOilModel : public BlackOilModel<TypeTag>
     using ElementContext = GetPropType<TypeTag, Properties::ElementContext>;
     using ThreadManager = GetPropType<TypeTag, Properties::ThreadManager>;
     using GridView = GetPropType<TypeTag, Properties::GridView>;
+    using FluidSystem = GetPropType<TypeTag, Properties::FluidSystem>;
     using Element = typename GridView::template Codim<0>::Entity;
     using ElementIterator = typename GridView::template Codim<0>::Iterator;
     enum {
@@ -102,6 +103,12 @@ public:
 
     void invalidateAndUpdateIntensiveQuantities(unsigned timeIdx) const
     {
+
+        // TODO: add something here so that the existing code will not be made any slower or cause copies
+        // maybe through a static assert on the type of the fluidsystem
+        // TODO: ensure that this will be an updated version if the static version has changed
+        const auto& FSystem = FluidSystem::getNonStaticInstance();
+
         this->invalidateIntensiveQuantitiesCache(timeIdx);
         //const auto& fluidSystemInstance = FluidSystem::getNonStaticInstance();
         OPM_BEGIN_PARALLEL_TRY_CATCH();
@@ -127,7 +134,7 @@ public:
             for (; it != end; ++it) {
                 const Element& elem = *it;
                 elemCtx.updatePrimaryStencil(elem);
-                elemCtx.updatePrimaryIntensiveQuantities(timeIdx);
+                elemCtx.updatePrimaryIntensiveQuantities(timeIdx, FSystem);
             }
         }
         OPM_END_PARALLEL_TRY_CATCH("InvalideAndUpdateIntensiveQuantities: state error", this->simulator_.vanguard().grid().comm());
