### 2024-12-11 3eeb33d
fixed: remove HYPRE from prereqs
this is controlled by USE_HYPRE and we should not unconditionally
look for it. had to move the code into config hook and split some ifs

pointy hat [at]akva2
