+++ opm/simulators/wells/WellInterface_impl.hpp
@@ -384,6 +384,7 @@ namespace Opm
                 WellTestState& well_test_state,
                 const PhaseUsage& phase_usage,
                 GLiftEclWells& ecl_well_map,
+                std::map<std::string, double>& open_times,
                 DeferredLogger& deferred_logger)
     {
         deferred_logger.info(" well " + this->name() + " is being tested");
@@ -480,6 +481,7 @@ namespace Opm
             // set the status of the well_state to open
             ws.open();
             well_state = well_state_copy;
+            open_times.try_emplace(this->name(), well_test_state.lastTestTime(this->name()));
         }
     }
 
