+++ opm/simulators/wells/BlackoilWellModel_impl.hpp
@@ -33,6 +33,7 @@
 #include <opm/input/eclipse/Schedule/Network/Balance.hpp>
 #include <opm/input/eclipse/Schedule/Network/ExtNetwork.hpp>
 #include <opm/input/eclipse/Schedule/Well/PAvgDynamicSourceData.hpp>
+#include <opm/input/eclipse/Schedule/Well/WellMatcher.hpp>
 #include <opm/input/eclipse/Schedule/Well/WellTestConfig.hpp>
 
 #include <opm/input/eclipse/Units/UnitSystem.hpp>
@@ -177,6 +178,11 @@ namespace Opm {
         const auto& events = this->schedule()[reportStepIdx].wellgroup_events();
         for (auto& wellPtr : this->well_container_) {
             const bool well_opened_this_step = this->report_step_starts_ && events.hasEvent(wellPtr->name(), effective_events_mask);
+            if (well_opened_this_step && this->wellState().well(wellPtr->name()).status == Well::Status::OPEN) {
+                this->well_open_times_.insert_or_assign(wellPtr->name(), this->simulator_.time());
+                this->well_close_times_.erase(wellPtr->name());
+            }
+
             wellPtr->init(&this->phase_usage_, this->depth_, this->gravity_, this->B_avg_, well_opened_this_step);
         }
     }
@@ -672,8 +678,8 @@ namespace Opm {
                 GLiftEclWells ecl_well_map;
                 initGliftEclWellMap(ecl_well_map);
                 well->wellTesting(simulator_, simulationTime, this->wellState(),
-                                  this->groupState(), this->wellTestState(), this->phase_usage_, 
-                                  ecl_well_map, deferred_logger);
+                                  this->groupState(), this->wellTestState(), this->phase_usage_,
+                                  ecl_well_map, this->well_open_times_, deferred_logger);
             } catch (const std::exception& e) {
                 const std::string msg = fmt::format("Exception during testing of well: {}. The well will not open.\n Exception message: {}", wellEcl.name(), e.what());
                 deferred_logger.warning("WELL_TESTING_FAILED", msg);
@@ -919,6 +925,13 @@ namespace Opm {
         if (nw > 0) {
             well_container_.reserve(nw);
 
+            const auto& wmatcher = this->schedule().wellMatcher(report_step);
+            const auto& wcycle = this->schedule()[report_step].wcycle.get();
+            const auto cycle_states = wcycle.wellStatus(this->simulator_.time(),
+                                                         wmatcher,
+                                                         this->well_open_times_,
+                                                         this->well_close_times_);
+
             for (int w = 0; w < nw; ++w) {
                 const Well& well_ecl = this->wells_ecl_[w];
 
@@ -940,6 +953,8 @@ namespace Opm {
                         this->wellState().shutWell(w);
                     }
 
+                    this->well_open_times_.erase(well_name);
+                    this->well_close_times_.erase(well_name);
                     continue;
                 }
 
@@ -957,6 +972,9 @@ namespace Opm {
                         if (!closed_this_step) {
                             this->wellTestState().open_well(well_name);
                             this->wellTestState().open_completions(well_name);
+                            this->well_open_times_.insert_or_assign(well_name,
+                                                                    this->simulator_.time());
+                            this->well_close_times_.erase(well_name);
                         }
                         events.clearEvent(ScheduleEvents::REQUEST_OPEN_WELL);
                     }
@@ -970,12 +988,16 @@ namespace Opm {
                     if (well_ecl.getAutomaticShutIn()) {
                         // shut wells are not added to the well container
                         this->wellState().shutWell(w);
+                        this->well_close_times_.erase(well_name);
+                        this->well_open_times_.erase(well_name);
                         continue;
                     } else {
[…]
