### 2024-12-11 dd223b5
working example with fluidsystem from FIB
