+++ flowexperimental/BlackOilIntensiveQuantitiesGlobalIndex.hpp
@@ -154,7 +154,8 @@ public:
     /*!
      * \copydoc IntensiveQuantities::update
      */
-    void update(const ElementContext& elemCtx, unsigned dofIdx, unsigned timeIdx)
+    template <class DynamicFluidSystem = void*>
+    void update(const ElementContext& elemCtx, unsigned dofIdx, unsigned timeIdx, const DynamicFluidSystem& fluidSystem = nullptr)
     {
         ParentType::update(elemCtx, dofIdx, timeIdx);
 
