### 2024-12-11 c6d9576
Regression test for the partitioner with GraphOfGrid
