### 2024-12-12 5bab785
Fix trivial NETV
This commit includes several fixes related to supporting NETV guide rates that are zero
