+++ opm/simulators/wells/FractionCalculator.hpp
@@ -56,10 +56,12 @@ public:
 
 private:
     std::string parent(const std::string& name);
-    Scalar guideRateSum(const Group& group,
-                        const std::string& always_included_child);
+    std::pair<Scalar,int> guideRateSum(const Group& group,
+                        const std::string& always_included_child,
+                        const bool use_potentials);
     Scalar guideRate(const std::string& name,
-                     const std::string& always_included_child);
+                     const std::string& always_included_child,
+                     const bool use_potentials);
     int groupControlledWells(const std::string& group_name,
                              const std::string& always_included_child);
     GuideRate::RateVector getGroupRateVector(const std::string& group_name);
