+++ opm/simulators/wells/FractionCalculator.cpp
@@ -77,18 +77,25 @@ Scalar FractionCalculator<Scalar>::
 localFraction(const std::string& name,
               const std::string& always_included_child)
 {
-    const Scalar my_guide_rate = guideRate(name, always_included_child);
-    const Group& parent_group = schedule_.getGroup(parent(name), report_step_);
-    const Scalar total_guide_rate = guideRateSum(parent_group, always_included_child);
+    bool only_use_potentials = false;
+    const Scalar my_guide_rate = guideRate(name, always_included_child, only_use_potentials);
 
-    // the total guide gate is the same as my_guide rate
-    // the well/group is probably on its own, i.e. return 1
-    // even is its guide_rate is zero
-    const Scalar guide_rate_epsilon = 1e-12;
-    if ( std::abs(my_guide_rate - total_guide_rate) < guide_rate_epsilon )
+    const Group& parent_group = schedule_.getGroup(parent(name), report_step_);
+    const auto [total_guide_rate, num_active_groups] = guideRateSum(parent_group, always_included_child, only_use_potentials);
+
+    // the group/well "name" is the only active group/well we therefore return 1 as the fraction
+    // even though my_guide_rate may be zero
+    if (num_active_groups == 1)
         return 1.0;
 
-    assert(total_guide_rate > my_guide_rate);
+    if (total_guide_rate == 0 ) {
+        // if the total guide rate is zero (for instance due to netv = 0) we use the potentials
+        // to distribute the group rate
+        only_use_potentials = true;
+        const Scalar my_pot = guideRate(name, always_included_child, only_use_potentials);
+        const Scalar my_total_pot = guideRateSum(parent_group, always_included_child, only_use_potentials).first;
+        return my_pot / my_total_pot;
+    }
     return my_guide_rate / total_guide_rate;
 }
 
@@ -104,11 +111,13 @@ parent(const std::string& name)
 }
 
 template<class Scalar>
-Scalar FractionCalculator<Scalar>::
+std::pair<Scalar, int> FractionCalculator<Scalar>::
 guideRateSum(const Group& group,
-             const std::string& always_included_child)
+             const std::string& always_included_child,
+             const bool use_potentials)
 {
     Scalar total_guide_rate = 0.0;
+    int number_of_included_well_or_group = 0;
     for (const std::string& child_group : group.groups()) {
         bool included = (child_group == always_included_child);
         if (is_producer_) {
@@ -122,7 +131,10 @@ guideRateSum(const Group& group,
                         (ctrl == Group::InjectionCMode::NONE);
         }
         if (included) {
-            total_guide_rate += guideRate(child_group, always_included_child);
+            if(groupControlledWells(child_group, always_included_child) > 0) {
+                number_of_included_well_or_group++;
+                total_guide_rate += guideRate(child_group, always_included_child, use_potentials);
+            }
         }
     }
 
@@ -133,34 +145,35 @@ guideRateSum(const Group& group,
         } else {
             included |= well_state_.isInjectionGrup(child_well);
         }
-
         if (included) {
-            total_guide_rate += guideRate(child_well, always_included_child);
+            number_of_included_well_or_group++;
+            total_guide_rate += guideRate(child_well, always_included_child, use_potentials);
         }
     }
-    return total_guide_rate;
+    return {total_guide_rate, number_of_included_well_or_group};
 }
 
 template<class Scalar>
 Scalar FractionCalculator<Scalar>::
 guideRate(const std::string& name,
-          const std::string& always_included_child)
+          const std::string& always_included_child,
+          const bool use_potentials)
 {
     if (schedule_.hasWell(name, report_step_)) {
         return WellGroupHelpers<Scalar>::getGuideRate(name, schedule_, well_state_, group_state_,
                                                       report_step_, guide_rate_, target_, pu_);
[…]
