### 2024-12-12 ff1970d
Fix the regression test name.
