### 2024-12-13 d8d074a
Compute temperature for xflowing injectors as weighted average of fluids entering the well.
