### 2024-12-16 93de995
Make Sources Buildable on Current Master
Not all 'problem's have a geomechanical model, so limit block level
stress collection to those that do.  While here, also install more
headers to enable downstream/out-of-tree users of the main simulator
classes.  Finally, reduce scope of some objects and be (more)
'const' correct.
