### 2024-12-16 1b094fd
Remove option of comparing only the SMRY file to run-parallel-regressionTest.sh via the flag -s, i.e. revert commit debb7cb036c336be1a16c8156ce663dfa982a031
