### 2024-12-16 410e19e
remove unused fipnum argument
also allows to drop the default value for the tuningUpdater callback
