### 2024-12-17 8fdeb52
Add test for distributed MSWells with two MSWells
