### 2024-12-17 bcb7842
--fixed option for edge conformal
