### 2024-12-17 6ac9b37
lgr experiments need changes in opm-common
