### 2024-12-17 317a54c
Relaxed tolerances for parallel test of 3_A_MPI_MULTFLT_SCHED_MOD2
This seems to be needed for the new partitioning with cells of a well
merged to one vertex.
