### 2024-12-17 ac5b6b5
Select AMGX mode based on template scalar type
