### 2024-12-17 b4b3f0d
Add tests for AMGX preconditioner
