### 2024-12-18 e2818ba
Regression test for --partition-method=3. Required extending macro functionality.
