### 2024-12-18 8a53a4b
Refactor Connection Value Report Function
In particular, split this function into distinct steps:

  1. Computing the linearised IJK index of each connection.
  2. Reporting multiplicative factors such as the CTF, the D factor,
     and the transmissibility multiplier due to rock compaction.
  3. Reporting pressures, flow rates, and productivity indices.
  4. Reporting filter cake values for injectors

While here, also reverse a condition in init() to enable "early
continue" and reduce levels of nesting and to reduce scope of some
objects.
