+++ opm/simulators/wells/WellState.cpp
@@ -23,22 +23,29 @@
 #include <opm/simulators/wells/WellState.hpp>
 
 #include <opm/common/ErrorMacros.hpp>
+
 #include <opm/input/eclipse/Schedule/MSW/WellSegments.hpp>
 #include <opm/input/eclipse/Schedule/Schedule.hpp>
 #include <opm/input/eclipse/Schedule/Well/Well.hpp>
 #include <opm/input/eclipse/Schedule/Well/WellConnections.hpp>
 
-#include <opm/simulators/utils/ParallelCommunication.hpp>
+#include <opm/output/data/Wells.hpp>
+
 #include <opm/simulators/wells/ParallelWellInfo.hpp>
 #include <opm/simulators/wells/PerforationData.hpp>
+
+#include <opm/simulators/utils/ParallelCommunication.hpp>
+
 #include <opm/grid/common/p2pcommunicator.hh>
-#include <opm/output/data/Wells.hpp>
 
 #include <algorithm>
 #include <cassert>
+#include <initializer_list>
 #include <numeric>
 #include <set>
 #include <stdexcept>
+#include <type_traits>
+#include <utility>
 #include <vector>
 
 #include <fmt/format.h>
@@ -267,27 +274,32 @@ void WellState<Scalar>::init(const std::vector<Scalar>& cellPressures,
                              const SummaryState& summary_state,
                              const bool enableDistributedWells)
 {
-    // call init on base class
-    this->base_init(cellPressures, cellTemperatures, wells_ecl, parallel_well_info,
+    // Call init on base class.
+    this->base_init(cellPressures, cellTemperatures,
+                    wells_ecl, parallel_well_info,
                     well_perf_data, summary_state);
+
     this->enableDistributedWells_ = enableDistributedWells;
-    this->global_well_info = std::make_optional<GlobalWellInfo<Scalar>>(schedule,
-                                                                report_step,
-                                                                wells_ecl);
+
+    this->global_well_info.emplace(schedule, report_step, wells_ecl);
+
     well_rates.clear();
+
     this->permanently_inactive_well_names_ = schedule.getInactiveWellNamesAtEnd();
-    for (const auto& wname : schedule.wellNames(report_step))
-    {
+
+    for (const auto& wname : schedule.wellNames(report_step)) {
         well_rates.insert({wname, std::make_pair(false, std::vector<Scalar>(this->numPhases()))});
     }
-    for (const auto& winfo: parallel_well_info)
-    {
+
+    for (const auto& winfo : parallel_well_info) {
         well_rates[winfo.get().name()].first = winfo.get().isOwner();
     }
 
-    const int nw = wells_ecl.size();
+    if (wells_ecl.empty()) {
+        return;
+    }
 
-    if( nw == 0 ) return ;
+    const int nw = wells_ecl.size();
 
     // Initialize perfphaserates_, which must be done here.
     const auto& pu = this->phaseUsage();
@@ -359,71 +371,75 @@ void WellState<Scalar>::init(const std::vector<Scalar>& cellPressures,
 
     // intialize wells that have been there before
     // order may change so the mapping is based on the well name
-    if (prevState && prevState->size() > 0) {
+    if ((prevState != nullptr) && (prevState->size() > 0)) {
         for (int w = 0; w < nw; ++w) {
-            const Well& well = wells_ecl[w];
-            if (well.getStatus() == Well::Status::SHUT) {
+            if (wells_ecl[w].getStatus() == Well::Status::SHUT) {
                 continue;
             }
+
+            const auto old_index = prevState->index(wells_ecl[w].name());
+            if (! old_index.has_value()) {
+                continue;
+            }
+
+            const auto& prev_well = prevState->well(*old_index);
+
             auto& new_well = this->well(w);
-            const auto& old_index = prevState->index(well.name());
[…]
