### 2024-12-18 a9471a3
Add Facility for Collecting Connection Level Fracturing Statistics
In particular,

  * RunningStatistics calculates basic statics measures of a running
    sample
  * ConnFracStatistics aggregates those measures for connection
    level fracture pressure, fracture width, and flow rate
