### 2024-12-18 9651765
Report Connection Level Fracturing Statistics to I/O Layer
Populates the new data::Connection::fract data member.
