### 2024-12-18 443751f
Report Connection Level Fracturing Statistics to I/O Layer
Populates the new data::Connection::fract data member.
