### 2024-12-18 1da573c
Report Connection Level Fracturing Statistics to I/O Layer
Populates the new data::Connection::fract data member.
