### 2024-12-18 0a9b51e
Regression test for --partition-method=3
