+++ opm/simulators/wells/GlobalWellInfo.hpp
@@ -49,7 +49,7 @@ enum class WellStatus;
     under group control.
 */
 
-
+template<class Scalar>
 class GlobalWellInfo {
 public:
 
@@ -64,6 +64,7 @@ public:
         comm.sum( this->m_in_injecting_group.data(), size);
         comm.sum( this->m_in_producing_group.data(), size);
         comm.sum( this->m_is_open.data(), size);
+        comm.min( this->m_efficiency_scaling_factors.data(), size);
     }
 
 
@@ -76,6 +77,8 @@ public:
     const std::string& well_name(std::size_t well_index) const;
     void update_injector(std::size_t well_index, WellStatus well_status, WellInjectorCMode injection_cmode);
     void update_producer(std::size_t well_index, WellStatus well_status, WellProducerCMode production_cmode);
+    void update_efficiency_scaling_factor(std::size_t well_index, const Scalar efficiency_scaling_factor);
+    Scalar efficiency_scaling_factor(const std::string& wname) const;
     void clear();
 
 private:
@@ -85,6 +88,7 @@ private:
     std::vector<int> m_in_injecting_group;       // global_index -> int/bool
     std::vector<int> m_in_producing_group;       // global_index -> int/bool
     std::vector<int> m_is_open;                  // global_index -> int/bool
+    std::vector<Scalar> m_efficiency_scaling_factors; // global_index --> double
 };
 
 
