+++ opm/simulators/wells/BlackoilWellModelGeneric.cpp
@@ -1576,7 +1576,7 @@ calculateEfficiencyFactors(const int reportStepIdx)
     for (auto& well : well_container_generic_) {
         const Well& wellEcl = well->wellEcl();
         Scalar well_efficiency_factor = wellEcl.getEfficiencyFactor() *
-                                        wellState()[well->name()].efficiency_scaling_factor;
+                                        wellState().getGlobalEfficiencyScalingFactor(well->name());
         WellGroupHelpers<Scalar>::accumulateGroupEfficiencyFactor(schedule().getGroup(wellEcl.groupName(),
                                                                                       reportStepIdx),
                                                                   schedule(),
