+++ opm/simulators/wells/WellState.cpp
@@ -271,7 +271,7 @@ void WellState<Scalar>::init(const std::vector<Scalar>& cellPressures,
     this->base_init(cellPressures, cellTemperatures, wells_ecl, parallel_well_info,
                     well_perf_data, summary_state);
     this->enableDistributedWells_ = enableDistributedWells;
-    this->global_well_info = std::make_optional<GlobalWellInfo>(schedule,
+    this->global_well_info = std::make_optional<GlobalWellInfo<Scalar>>(schedule,
                                                                 report_step,
                                                                 wells_ecl);
     well_rates.clear();
@@ -969,6 +969,7 @@ void WellState<Scalar>::updateGlobalIsGrup(const Parallel::Communication& comm)
     this->global_well_info.value().clear();
     for (std::size_t well_index = 0; well_index < this->size(); well_index++) {
         const auto& ws = this->well(well_index);
+        this->global_well_info.value().update_efficiency_scaling_factor(well_index, ws.efficiency_scaling_factor);
         if (ws.producer)
             this->global_well_info.value().update_producer(well_index, ws.status, ws.production_cmode);
         else
