+++ opm/simulators/wells/WellState.hpp
@@ -180,6 +180,11 @@ public:
         return this->global_well_info.value().is_open(name);
     }
 
+    bool getGlobalEfficiencyScalingFactor(const std::string& name) const
+    {
+        return this->global_well_info.value().efficiency_scaling_factor(name);
+    }
+
     Scalar getALQ(const std::string& name) const
     {
         return this->alq_state.get(name);
@@ -373,7 +378,7 @@ private:
     // Use of std::optional<> here is a technical crutch, the
     // WellStateFullyImplicitBlackoil class should be default constructible,
     // whereas the GlobalWellInfo is not.
-    std::optional<GlobalWellInfo> global_well_info;
+    std::optional<GlobalWellInfo<Scalar>> global_well_info;
     ALQState<Scalar> alq_state;
 
     // The well_rates variable is defined for all wells on all processors. The
