+++ opm/simulators/wells/BlackoilWellModel_impl.hpp
@@ -655,7 +655,7 @@ namespace Opm {
             well->init(&this->phase_usage_, depth_, gravity_, B_avg_, true);
 
             Scalar well_efficiency_factor = wellEcl.getEfficiencyFactor() *
-                                            this->wellState()[well_name].efficiency_scaling_factor;
+                                            this->wellState().getGlobalEfficiencyScalingFactor(well_name);
             WellGroupHelpers<Scalar>::accumulateGroupEfficiencyFactor(this->schedule().getGroup(wellEcl.groupName(),
                                                                                                 timeStepIdx),
                                                                       this->schedule(),
@@ -1061,7 +1061,8 @@ namespace Opm {
                                                                           this->well_open_times_,
                                                                           schedule_open))
                 {
-                    this->wellState()[wname].efficiency_scaling_factor = wscale;
+                    if (this->wellState().has(wname))
+                        this->wellState()[wname].efficiency_scaling_factor = wscale;
                     this->schedule_.add_event(ScheduleEvents::WELLGROUP_EFFICIENCY_UPDATE, report_step);
                 }
             }
