+++ parallelTests.cmake
@@ -78,9 +78,10 @@ add_test_compare_parallel_simulation(CASENAME spe3
                                      REL_TOL ${coarse_rel_tol_parallel}
                                      TEST_ARGS --linear-solver-reduction=1e-7 --tolerance-cnv=5e-6 --tolerance-mb=1e-8 --tolerance-wells=1e-7)
 
-add_test_compare_parallel_simulation(CASENAME spe3
+add_test_compare_parallel_simulation(CASENAME spe3_partition_method_3
                                      FILENAME SPE3CASE1
                                      POSTFIX partition_method_3
+                                     DIR spe3
                                      SIMULATOR flow
                                      ABS_TOL ${abs_tol_parallel}
                                      REL_TOL ${coarse_rel_tol_parallel}
