### 2024-12-19 9c42aeb
Fix wrong return type and revert unneccessary changes.
