### 2024-12-19 c17aad1
add parallel regression test for WCYCLE
