### 2024-12-19 8158ae8
allow defaulted group guide rates
clean up
