### 2024-12-19 74d6ac1
Skip allocTrans when enableEclOutput is False and
