### 2024-12-19 88e3e31
Skip allocTrans when enableEclOutput is False and
