+++ opm/simulators/flow/FlowBaseVanguard.hpp
@@ -139,6 +139,7 @@ public:
         externalPartitionFile_ = Parameters::Get<Parameters::ExternalPartition>();
 #endif
         enableDistributedWells_ = Parameters::Get<Parameters::AllowDistributedWells>();
+        enableEclOutput_ = Parameters::Get<Parameters::EnableEclOutput>();
         allow_splitting_inactive_wells_ = Parameters::Get<Parameters::AllowSplittingInactiveWells>();
         ignoredKeywords_ = Parameters::Get<Parameters::IgnoreKeywords>();
         int output_param = Parameters::Get<Parameters::EclOutputInterval>();
