+++ opm/simulators/wells/MultisegmentWellSegments.cpp
@@ -514,7 +514,7 @@ getFrictionPressureLoss(const int seg,
     // at segment node while fraction derivatives are given at upwind node.
 
     if (seg != seg_upwind) {
-        if (!extra_reverse_flow_derivatives){
+        if (!extra_reverse_flow_derivatives) {
             constexpr int WQTotal = Indices::numEq + PrimaryVariables::WQTotal;
             constexpr int SPres = Indices::numEq + PrimaryVariables::SPres;
             density.setDerivative(WQTotal, 0.0);
@@ -522,12 +522,12 @@ getFrictionPressureLoss(const int seg,
             visc.setDerivative(WQTotal, 0.0);
             visc.setDerivative(SPres, 0.0);
         } else {
-            if (PrimaryVariables::has_water){
+            if (PrimaryVariables::has_wfrac_variable) {
                 constexpr int WFrac = Indices::numEq + PrimaryVariables::WFrac;
                 density.setDerivative(WFrac, 0.0);
                 visc.setDerivative(WFrac, 0.0);
             }
-            if (PrimaryVariables::has_gas){
+            if (PrimaryVariables::has_gfrac_variable) {
                 constexpr int GFrac = Indices::numEq + PrimaryVariables::GFrac;
                 density.setDerivative(GFrac, 0.0);
                 visc.setDerivative(GFrac, 0.0);
@@ -600,10 +600,10 @@ pressureDropSpiralICD(const int seg,
             zero_mask[PrimaryVariables::WQTotal] = true;
             zero_mask[PrimaryVariables::SPres] = true;
         } else {
-            if constexpr (PrimaryVariables::has_water){
+            if constexpr (PrimaryVariables::has_wfrac_variable) {
                 zero_mask[PrimaryVariables::WFrac] = true;
             }
-            if constexpr (PrimaryVariables::has_gas){
+            if constexpr (PrimaryVariables::has_gfrac_variable) {
                 zero_mask[PrimaryVariables::GFrac] = true;
             }
             // mass_rate has no extra derivatives (they are organized as in equations)
@@ -712,10 +712,10 @@ pressureDropAutoICD(const int seg,
             zero_mask[PrimaryVariables::WQTotal] = true;
             zero_mask[PrimaryVariables::SPres] = true;
         } else {
-            if (PrimaryVariables::has_water){
+            if (PrimaryVariables::has_wfrac_variable) {
                 zero_mask[PrimaryVariables::WFrac] = true;
             }
-            if (PrimaryVariables::has_gas){
+            if (PrimaryVariables::has_gfrac_variable) {
                 zero_mask[PrimaryVariables::GFrac] = true;
             }
             // mass_rate has no extra derivatives (they are organized as in equations)
@@ -787,7 +787,7 @@ pressureDropValve(const int seg,
     // For reference: the pressure equation assumes pressure/flow derivatives are given 
     // at segment node while fraction derivatives are given at upwind node. 
     if (seg != seg_upwind) {
-        if (!extra_reverse_flow_derivatives){
+        if (!extra_reverse_flow_derivatives) {
             constexpr int WQTotal = Indices::numEq + PrimaryVariables::WQTotal;
             constexpr int SPres = Indices::numEq + PrimaryVariables::SPres;
             density.setDerivative(WQTotal, 0.0);
@@ -795,12 +795,12 @@ pressureDropValve(const int seg,
             visc.setDerivative(WQTotal, 0.0);
             visc.setDerivative(SPres, 0.0);
         } else {
-            if (PrimaryVariables::has_water){
+            if (PrimaryVariables::has_wfrac_variable) {
[…]
