+++ opm/simulators/wells/MultisegmentWellPrimaryVariables.cpp
@@ -225,7 +225,6 @@ copyToWellState(const MultisegmentWellGeneric<Scalar>& mswell,
     const auto pvtReg = std::max(well_.wellEcl().pvt_table_number() - 1, 0);
 
     const PhaseUsage& pu = well_.phaseUsage();
-    assert( FluidSystem::phaseIsActive(FluidSystem::oilPhaseIdx) );
     const int oil_pos = pu.phase_pos[Oil];
 
     auto& ws = well_state.well(well_.indexOfWell());
@@ -236,18 +235,34 @@ copyToWellState(const MultisegmentWellGeneric<Scalar>& mswell,
     auto& segment_pressure = segments.pressure;
     for (std::size_t seg = 0; seg < value_.size(); ++seg) {
         std::vector<Scalar> fractions(well_.numPhases(), 0.0);
-        fractions[oil_pos] = 1.0;
 
-        if (FluidSystem::phaseIsActive(FluidSystem::waterPhaseIdx)) {
-            const int water_pos = pu.phase_pos[Water];
-            fractions[water_pos] = value_[seg][WFrac];
-            fractions[oil_pos] -= fractions[water_pos];
+        if (FluidSystem::phaseIsActive(FluidSystem::oilPhaseIdx)) {
+            fractions[oil_pos] = 1.0;
+            if (FluidSystem::phaseIsActive(FluidSystem::waterPhaseIdx)) {
+                const int water_pos = pu.phase_pos[Water];
+                fractions[water_pos] = value_[seg][WFrac];
+                fractions[oil_pos] -= fractions[water_pos];
+            }
+
+            if (FluidSystem::phaseIsActive(FluidSystem::gasPhaseIdx)) {
+                const int gas_pos = pu.phase_pos[Gas];
+                fractions[gas_pos] = value_[seg][GFrac];
+                fractions[oil_pos] -= fractions[gas_pos];
+            }
         }
+        else if (FluidSystem::phaseIsActive(FluidSystem::waterPhaseIdx)) {
+            const int water_pos = pu.phase_pos[Water];
+            fractions[water_pos] = 1.0;
 
-        if (FluidSystem::phaseIsActive(FluidSystem::gasPhaseIdx)) {
+            if (FluidSystem::phaseIsActive(FluidSystem::gasPhaseIdx)) {
+                const int gas_pos = pu.phase_pos[Gas];
+                fractions[gas_pos] = value_[seg][GFrac];
+                fractions[water_pos] -= fractions[gas_pos];
+            }
+        }
+        else if (FluidSystem::phaseIsActive(FluidSystem::gasPhaseIdx)) {
             const int gas_pos = pu.phase_pos[Gas];
-            fractions[gas_pos] = value_[seg][GFrac];
-            fractions[oil_pos] -= fractions[gas_pos];
+            fractions[gas_pos] = 1.0;
         }
 
         // convert the fractions to be Q_p / G_total to calculate the phase rates
@@ -287,7 +302,7 @@ copyToWellState(const MultisegmentWellGeneric<Scalar>& mswell,
 
         Scalar rsMax = 0.0;
         Scalar rvMax = 0.0;
-        if (FluidSystem::phaseIsActive(FluidSystem::gasPhaseIdx)) {
+        if (FluidSystem::phaseIsActive(FluidSystem::gasPhaseIdx) && FluidSystem::phaseIsActive(FluidSystem::oilPhaseIdx)) {
             // Both oil and gas active.
[…]
