### 2024-12-19 60bcfd7
Remove the mpi calls from the business logic
Implement mmv and mv functions that are called from the business layer that do the communication
