### 2024-12-19 890f285
Remove the mpi calls from the business logic
Implement mmv and mv functions that are called from the business layer that do the communication
