+++ opm/simulators/wells/MSWellHelpers.cpp
@@ -35,6 +35,8 @@
 #include <opm/simulators/utils/DeferredLogger.hpp>
 #include <opm/simulators/utils/DeferredLoggingErrorHelpers.hpp>
 
+#include <opm/simulators/wells/ParallelWellInfo.hpp>
+
 #include <dune/istl/preconditioners.hh>
 #include <dune/istl/solvers.hh>
 
@@ -96,6 +98,45 @@ ValueType OIWEmulsionViscosity(const ValueType& water_viscosity,
 }
 
 namespace Opm::mswellhelpers {
+
+template<class MatrixType>
+ParallellMSWellB<MatrixType>::
+ParallellMSWellB(const MatrixType& B,
+                 const ParallelWellInfo<Scalar>& parallel_well_info)
+    : B_(B), parallel_well_info_(parallel_well_info)
+{}
+
+template<class MatrixType>
+template<class X, class Y>
+void ParallellMSWellB<MatrixType>::
+mv (const X& x, Y& y) const
+{
+    B_.mv(x, y);
+
+    if (this->parallel_well_info_.communication().size() > 1)
+    {
+        // Communicate here to get the contributions from all segments
+        this->parallel_well_info_.communication().sum(y.data(), y.size());
+    }
+}
+
+template<class MatrixType>
+template<class X, class Y>
+void ParallellMSWellB<MatrixType>::
+mmv (const X& x, Y& y) const
+{
+    if (this->parallel_well_info_.communication().size() == 1)
+    {
+        // Do the same thing as before. The else branch
+        // produces different rounding errors and results
+        // slightly different iteration counts / well curves
+        B_.mmv(x, y);
+    } else {
+        Y temp(y);
+        mv(x, temp); // includes parallel reduction
+        y -= temp;
+    }
+}
 
     /// Applies umfpack and checks for singularity
 template <typename MatrixType, typename VectorType>
@@ -308,8 +349,24 @@ ValueType emulsionViscosity(const ValueType& water_fraction,
 
 template<class Scalar, int Dim>
 using Vec = Dune::BlockVector<Dune::FieldVector<Scalar,Dim>>;
-template<class Scalar, int Dim>
-using Mat = Dune::BCRSMatrix<Dune::FieldMatrix<Scalar,Dim,Dim>>;
+template<class Scalar, int M, int N = M>
+using Mat = Dune::BCRSMatrix<Dune::FieldMatrix<Scalar,M,N>>;
+
+#define INSTANTIATE_PARALLELLMSWELLB(T, M, N)                                                                                                                                                                                                                                      \
+    template class ParallellMSWellB<Mat<T,M,N>>;                                                                                                                                                                                                 \
+    template void ParallellMSWellB<Mat<T,M,N>>::mv(Vec<T,N> const&,Vec<T,M>& ) const; \
+    template void ParallellMSWellB<Mat<T,M,N>>::mmv(Vec<T,N> const&,Vec<T,M>& ) const;
+
+#define INSTANTIATE_ALL_PARALLELLMSWELLB(T) \
+    INSTANTIATE_PARALLELLMSWELLB(T, 2, 1)   \
+    INSTANTIATE_PARALLELLMSWELLB(T, 2, 2)   \
+    INSTANTIATE_PARALLELLMSWELLB(T, 2, 6)   \
[…]
