### 2024-12-19 6562503
Add template to assignFieldPropsDoubleOnLeaf
