### 2024-12-20 ece9456
Change msw-simple test to use solver-max-time-step-in-days=15 instead of solver-max-time-step-in-days=10
