+++ opm/simulators/wells/MultisegmentWellEquations.cpp
@@ -152,7 +152,7 @@ apply(const BVector& x, BVector& Ax) const
 
     duneB_.mv(x, Bx);
 
-    if (this->pw_info_.communication().size() == 1) {
+    if (this->pw_info_.communication().size() > 1) {
         // We need to communicate here to get the contributions from all segments
         this->pw_info_.communication().sum(Bx.data(), Bx.size());
     }
