### 2024-12-20 27ec1eb
BlackoilWellModel: add some accessors
