### 2024-12-20 165d0a9
changed: move implementation of WellOperators out of BlackoilWellModel
