### 2024-12-20 4a797a7
simplify: separate out application of a single well
and share code between normal and domain operators
