### 2024-12-20 e5f8ea3
simplify: applyScaleAdd can be shared between regular and domain operators
