### 2024-12-20 4e4e554
dont use implicit potentials for newly opened wells with target zero
