### 2024-12-20 177aa39
dont use implicit potentials for newly opened wells with target zero
