### 2024-12-20 8489343
should not use phase index for canonicalToActiveComponentIndex()
