### 2024-12-20 ed452d8
changed: remove unused code in WellConnectionAuxilliaryModule
