### 2024-12-20 2800f35
changed: move the auxillary well module from BlackoilModule to WellConnectionAuxillary
this is an adapter for the non-linear solver.
it is still model using is-a but this can now be changed if desired.
