+++ opm/simulators/wells/WellConnectionAuxiliaryModule.hpp
@@ -23,51 +23,172 @@
 
 #include <opm/models/discretization/common/baseauxiliarymodule.hh>
 
-#include <vector>
+#include <opm/simulators/flow/SubDomain.hpp>
 
-namespace Dune { class CpGrid; }
+#include <opm/simulators/utils/DeferredLoggingErrorHelpers.hpp>
+#include <opm/simulators/utils/ParallelCommunication.hpp>
 
-namespace Opm
+#if HAVE_MPI
+#include <opm/simulators/utils/MPISerializer.hpp>
+#endif
+
+namespace Opm {
+
+template<class TypeTag, class Model>
+class WellConnectionAuxiliaryModule : public BaseAuxiliaryModule<TypeTag>
 {
-
-class Schedule;
-
-template<class TypeTag>
-class WellConnectionAuxiliaryModule
-    : public BaseAuxiliaryModule<TypeTag>
-{
+    using Grid = GetPropType<TypeTag, Properties::Grid>;
     using GlobalEqVector = GetPropType<TypeTag, Properties::GlobalEqVector>;
     using SparseMatrixAdapter = GetPropType<TypeTag, Properties::SparseMatrixAdapter>;
 
 public:
-
     using NeighborSet = typename
         ::Opm::BaseAuxiliaryModule<TypeTag>::NeighborSet;
 
-    WellConnectionAuxiliaryModule(const Schedule& schedule,
-                                  const Dune::CpGrid& grid)
+    using Domain = SubDomain<Grid>;
+
+    WellConnectionAuxiliaryModule(Model& model, Parallel::Communication comm)
+        : model_(model)
+        , lin_comm_(std::move(comm))
     {
     }
 
-    unsigned numDofs() const
+    unsigned numDofs() const override
     {
         // No extra dofs are inserted for wells.
         return 0;
     }
 
-    void addNeighbors(std::vector<NeighborSet>& neighbors) const
+    void addNeighbors(std::vector<NeighborSet>& neighbors) const override
     {
+        if (!model_.addMatrixContributions()) {
+            return;
+        }
+
+        // Create cartesian to compressed mapping
+        const auto& schedule_wells = model_.schedule().getWellsatEnd();
+        auto possibleFutureConnections = model_.schedule().getPossibleFutureConnections();
+
+#if HAVE_MPI
+        // Communicate Map to other processes, since it is only available on rank 0
+        Parallel::MpiSerializer ser(lin_comm_);
+        ser.broadcast(possibleFutureConnections);
+#endif
+        // initialize the additional cell connections introduced by wells.
+        for (const auto& well : schedule_wells)
+        {
+            std::vector<int> wellCells = model_.getCellsForConnections(well);
+            // Now add the cells of the possible future connections
+            const auto possibleFutureConnectionSetIt = possibleFutureConnections.find(well.name());
+            if (possibleFutureConnectionSetIt != possibleFutureConnections.end()) {
+                for (auto& global_index : possibleFutureConnectionSetIt->second) {
+                    int compressed_idx = model_.compressedIndexForInterior(global_index);
+                    if (compressed_idx >= 0) { // Ignore connections in inactive/remote cells.
+                        wellCells.push_back(compressed_idx);
+                    }
+                }
+            }
+            for (int cellIdx : wellCells) {
+                neighbors[cellIdx].insert(wellCells.begin(),
+                                          wellCells.end());
+            }
+        }
     }
 
-    void applyInitial()
+    void applyInitial() override
     {}
 
-    void linearize(SparseMatrixAdapter& , GlobalEqVector&)
+    void linearize(SparseMatrixAdapter& jacobian, GlobalEqVector& res) override
     {
[…]
