+++ opm/simulators/wells/BlackoilWellModel.hpp
@@ -24,13 +24,11 @@
 #ifndef OPM_BLACKOILWELLMODEL_HEADER_INCLUDED
 #define OPM_BLACKOILWELLMODEL_HEADER_INCLUDED
 
-#include <opm/common/OpmLog/OpmLog.hpp>
+#include <dune/common/fmatrix.hh>
+#include <dune/istl/bcrsmatrix.hh>
+#include <dune/istl/matrixmatrix.hh>
 
-#include <cstddef>
-#include <map>
-#include <memory>
-#include <string>
-#include <vector>
+#include <opm/common/OpmLog/OpmLog.hpp>
 
 #include <opm/grid/utility/SparseTable.hpp>
 
@@ -39,13 +37,18 @@
 #include <opm/input/eclipse/Schedule/Schedule.hpp>
 #include <opm/input/eclipse/Schedule/Well/WellTestState.hpp>
 
-#include <opm/models/discretization/common/baseauxiliarymodule.hh>
+#include <opm/material/densead/Math.hpp>
 
 #include <opm/simulators/flow/countGlobalCells.hpp>
 #include <opm/simulators/flow/FlowBaseVanguard.hpp>
 #include <opm/simulators/flow/SubDomain.hpp>
 
 #include <opm/simulators/linalg/matrixblock.hh>
+
+#include <opm/simulators/timestepping/SimulatorReport.hpp>
+#include <opm/simulators/timestepping/gatherConvergenceReport.hpp>
+
+#include <opm/simulators/utils/DeferredLogger.hpp>
 
 #include <opm/simulators/wells/BlackoilWellModelGasLift.hpp>
 #include <opm/simulators/wells/BlackoilWellModelGeneric.hpp>
@@ -63,22 +66,18 @@
 #include <opm/simulators/wells/StandardWell.hpp>
 #include <opm/simulators/wells/VFPInjProperties.hpp>
 #include <opm/simulators/wells/VFPProdProperties.hpp>
-#include <opm/simulators/wells/WGState.hpp>
+#include <opm/simulators/wells/WellConnectionAuxiliaryModule.hpp>
 #include <opm/simulators/wells/WellGroupHelpers.hpp>
 #include <opm/simulators/wells/WellInterface.hpp>
 #include <opm/simulators/wells/WellProdIndexCalculator.hpp>
 #include <opm/simulators/wells/WellState.hpp>
+#include <opm/simulators/wells/WGState.hpp>
 
-#include <opm/simulators/timestepping/SimulatorReport.hpp>
-#include <opm/simulators/timestepping/gatherConvergenceReport.hpp>
-
-#include <dune/common/fmatrix.hh>
-#include <dune/istl/bcrsmatrix.hh>
-#include <dune/istl/matrixmatrix.hh>
-
-#include <opm/material/densead/Math.hpp>
-
-#include <opm/simulators/utils/DeferredLogger.hpp>
+#include <cstddef>
+#include <map>
+#include <memory>
+#include <string>
+#include <vector>
 
 namespace Opm {
 
@@ -88,7 +87,7 @@ template<class Scalar> class WellContributions;
 
         /// Class for handling the blackoil well model.
         template<typename TypeTag>
-        class BlackoilWellModel : public BaseAuxiliaryModule<TypeTag>
+        class BlackoilWellModel : public WellConnectionAuxiliaryModule<TypeTag, BlackoilWellModel<TypeTag>>
                                 , public BlackoilWellModelGeneric<GetPropType<TypeTag,
                                                                               Properties::Scalar>>
         {
@@ -106,8 +105,9 @@ template<class Scalar> class WellContributions;
             using SparseMatrixAdapter = GetPropType<TypeTag, Properties::SparseMatrixAdapter>;
[…]
