### 2024-12-20 be6ec0e
split out linearization for a single well
use this to unify code between regular and domain
