### 2024-12-20 939f944
Remove unneded capture of 'this'.
