+++ opm/simulators/wells/MultisegmentWellAssemble.cpp
@@ -279,9 +279,7 @@ assemblePressureEq(const int seg,
                    const int outlet_segment_index,
                    const EvalWell& pressure_equation,
                    const EvalWell& outlet_pressure,
-                   Equations& eqns1,
-                   bool wfrac,
-                   bool gfrac) const
+                   Equations& eqns1) const
 {
     /*
         This method does *not* need communication.
@@ -290,10 +288,10 @@ assemblePressureEq(const int seg,
     eqns.residual()[seg][SPres] += pressure_equation.value();
     eqns.D()[seg][seg][SPres][SPres] += pressure_equation.derivative(SPres + Indices::numEq);
     eqns.D()[seg][seg][SPres][WQTotal] += pressure_equation.derivative(WQTotal + Indices::numEq);
-    if (wfrac) {
+    if (has_wfrac_variable) {
         eqns.D()[seg][seg_upwind][SPres][WFrac] += pressure_equation.derivative(WFrac + Indices::numEq);
     }
-    if (gfrac) {
+    if (has_gfrac_variable) {
         eqns.D()[seg][seg_upwind][SPres][GFrac] += pressure_equation.derivative(GFrac + Indices::numEq);
     }
 
