### 2024-12-20 b5c25a2
some debugging code
