+++ regressionTests.cmake
@@ -947,7 +947,8 @@ add_test_compareECLFiles(CASENAME 9_3d_grpctl_stw_model2
                          SIMULATOR flow
                          ABS_TOL ${abs_tol}
                          REL_TOL ${rel_tol}
-                         DIR model2)
+                         DIR model2
+                         TEST_ARGS --time-step-after-event-in-days=1)
 
 add_test_compareECLFiles(CASENAME 9_3d_grpctl_msw_model2
                          FILENAME 9_3D_GINJ_GAS_MAX_EXPORT_MSW
