### 2024-12-20 2af056b
Add LGRs into equilGrid_. Issue: want seq call
