+++ opm/simulators/flow/GenericCpGridVanguard.hpp
@@ -211,7 +211,7 @@ protected:
     void doCreateGrids_(EclipseState& eclState);
     void addLgrsUpdateLeafView(const LgrCollection& lgrCollection,
                                const int lgrsSize,
-                               Dune::CpGrid grid);
+                               const  std::unique_ptr<Dune::CpGrid>& grid);
 
     virtual void allocTrans() = 0;
     virtual double getTransmissibility(unsigned I, unsigned J) const = 0;
