### 2024-12-26 b948686
Updates to AdaptiveTimeStepping_impl.hpp
Include the latest PRs that changed AdaptiveTimeStepping.hpp into
AdaptiveTimeStepping_impl.hpp. The PRs included are:

- remove 16 days restriction for timestep in predicition mode #2970
- remove unused fipnum argument #5803: remove unused fipnum argument
- added: support WCYCLE #5792: added: support WCYCLE
