+++ opm/simulators/timestepping/AdaptiveTimeStepping_impl.hpp
@@ -174,13 +174,16 @@ template <class Solver>
 SimulatorReport
 AdaptiveTimeStepping<TypeTag>::
 step(const SimulatorTimer& simulator_timer,
-                                    Solver& solver,
-                                    const bool is_event,
-                                    const std::vector<int>* fipnum,
-                                    const std::function<bool()> tuning_updater)
+     Solver& solver,
+     const bool is_event,
+     const std::function<bool(const double /*current_time*/,
+                              const double /*dt*/,
+                              const int    /*substep_number*/
+                             )> tuning_updater
+)
 {
     SubStepper sub_stepper{
-        *this, simulator_timer, solver, is_event, fipnum, tuning_updater,
+        *this, simulator_timer, solver, is_event, tuning_updater,
     };
     return sub_stepper.run();
 }
@@ -462,14 +465,16 @@ SubStepper(
     const SimulatorTimer& simulator_timer,
     Solver& solver,
     const bool is_event,
-    const std::vector<int>* fipnum,
-    const std::function<bool()>& tuning_updater
+    const std::function<bool(const double /*current_time*/,
+                             const double /*dt*/,
+                             const int    /*substep_number*/
+                            )>& tuning_updater
+
 )
     : adaptive_time_stepping_{adaptive_time_stepping}
     , simulator_timer_{simulator_timer}
     , solver_{solver}
     , is_event_{is_event}
-    , fipnum_{fipnum}
     , tuning_updater_{tuning_updater}
 {
 }
@@ -538,13 +543,16 @@ maybeModifySuggestedTimeStepAtBeginningOfReportStep_(const double original_time_
     );
 }
 
+// The maybeUpdateTuning_() lambda callback is defined in SimulatorFullyImplicitBlackoil::runStep()
+// It has to be called for each substep since TUNING might have been changed for next sub step due
+// to ACTIONX (via NEXTSTEP) or WCYCLE keywords.
 template<class TypeTag>
 template<class Solver>
 bool
 AdaptiveTimeStepping<TypeTag>::SubStepper<Solver>::
-maybeUpdateTuning_() const
+maybeUpdateTuning_(double elapsed, double dt, int sub_step_number) const
 {
-    return this->tuning_updater_();
+    return this->tuning_updater_(elapsed, dt, sub_step_number);
 }
 
 template<class TypeTag>
@@ -562,16 +570,18 @@ SimulatorReport
 AdaptiveTimeStepping<TypeTag>::SubStepper<Solver>::
 runStepOriginal_()
 {
-    maybeUpdateTuning_();
-    const double original_time_step = this->simulator_timer_.currentStepLength();
+    auto elapsed = this->simulator_timer_.simulationTimeElapsed();
+    auto original_time_step = this->simulator_timer_.currentStepLength();
+    auto report_step = this->simulator_timer_.reportStepNum();
+    maybeUpdateTuning_(elapsed, original_time_step, report_step);
     maybeModifySuggestedTimeStepAtBeginningOfReportStep_(original_time_step);
 
     AdaptiveSimulatorTimer substep_timer{
         this->simulator_timer_.startDateTime(),
         original_time_step,
-        this->simulator_timer_.simulationTimeElapsed(),
+        elapsed,
         suggestedNextTimestep_(),
-        this->simulator_timer_.reportStepNum(),
+        report_step,
         maxTimeStep_()
     };
     SubStepIteration substepIteration{*this, substep_timer, original_time_step, /*final_step=*/true};
@@ -617,7 +627,7 @@ runStepReservoirCouplingMaster_()
     while(!substep_done) {
         reservoirCouplingMaster_().receiveNextReportDateFromSlaves();
         if (iteration == 0) {
[…]
