### 2024-12-27 c1c6c5c
Add test for inserting keyword GLIFTOPT from a Pyaction
