### 2025-01-02 15a9fd1
changed: move GasLift to separate class
move to has-a instead of is-a for modelling in BlackoilWellModel
