+++ opm/simulators/wells/BlackoilWellModel_impl.hpp
@@ -679,10 +679,16 @@ namespace Opm {
 
             try {
                 GLiftEclWells ecl_well_map;
-                initGliftEclWellMap(ecl_well_map);
-                well->wellTesting(simulator_, simulationTime, this->wellState(),
-                                  this->groupState(), this->wellTestState(), this->phase_usage_,
-                                  ecl_well_map, this->well_open_times_, deferred_logger);
+                gaslift_.initGliftEclWellMap(well_container_, ecl_well_map);
+                well->wellTesting(simulator_,
+                                  simulationTime,
+                                  this->wellState(),
+                                  this->groupState(),
+                                  this->wellTestState(),
+                                  this->phase_usage_,
+                                  ecl_well_map,
+                                  this->well_open_times_,
+                                  deferred_logger);
             } catch (const std::exception& e) {
                 const std::string msg = fmt::format("Exception during testing of well: {}. The well will not open.\n Exception message: {}", wellEcl.name(), e.what());
                 deferred_logger.warning("WELL_TESTING_FAILED", msg);
@@ -842,8 +848,9 @@ namespace Opm {
     {
         rate = 0;
 
-        if (!is_cell_perforated_[elemIdx])
+        if (!is_cell_perforated_[elemIdx]) {
             return;
+        }
 
         for (const auto& well : well_container_)
             well->addCellRates(rate, elemIdx);
@@ -862,8 +869,9 @@ namespace Opm {
         rate = 0;
         int elemIdx = context.globalSpaceIndex(spaceIdx, timeIdx);
 
-        if (!is_cell_perforated_[elemIdx])
+        if (!is_cell_perforated_[elemIdx]) {
             return;
+        }
 
         for (const auto& well : well_container_)
             well->addCellRates(rate, elemIdx);
@@ -1235,11 +1243,12 @@ namespace Opm {
     {
 
         DeferredLogger local_deferredLogger;
-        if (this->glift_debug) {
+/*        if (this->glift_debug) {
             const std::string msg = fmt::format(
                 "assemble() : iteration {}" , iterationIdx);
             this->gliftDebug(msg, local_deferredLogger);
         }
+*/
         last_report_ = SimulatorReportSingle();
         Dune::Timer perfTimer;
         perfTimer.start();
@@ -1334,7 +1343,9 @@ namespace Opm {
                                           DeferredLogger& local_deferredLogger)
     {
         auto [well_group_control_changed, more_network_update] =
-                updateWellControls(mandatory_network_balance, local_deferredLogger, relax_network_tolerance);
+                updateWellControls(mandatory_network_balance,
+                                   local_deferredLogger,
+                                   relax_network_tolerance);
 
         bool alq_updated = false;
         OPM_BEGIN_PARALLEL_TRY_CATCH();
@@ -1342,14 +1353,21 @@ namespace Opm {
             // Set the well primary variables based on the value of well solutions
             initPrimaryVariablesEvaluation();
 
-            alq_updated = maybeDoGasLiftOptimize(local_deferredLogger);
+            alq_updated = gaslift_.maybeDoGasLiftOptimize(simulator_,
+                                                          well_container_,
+                                                          this->terminal_output_,
+                                                          this->phase_usage_,
+                                                          this->wellState(),
+                                                          this->groupState(),
+                                                          local_deferredLogger);
 
             prepareWellsBeforeAssembling(dt, local_deferredLogger);
         }
-        OPM_END_PARALLEL_TRY_CATCH_LOG(local_deferredLogger, "updateWellControlsAndNetworkIteration() failed: ",
+        OPM_END_PARALLEL_TRY_CATCH_LOG(local_deferredLogger,
[…]
