### 2025-01-02 89dbda4
changed: move GasLift to separate class
move to has-a instead of is-a for modelling in BlackoilWellModel
