+++ opm/simulators/wells/BlackoilWellModel_impl.hpp
@@ -79,6 +79,7 @@ namespace Opm {
                                             phase_usage,
                                             simulator.gridView().comm())
         , simulator_(simulator)
+        , gaslift_(this->terminal_output_, this->phase_usage_)
     {
         local_num_cells_ = simulator_.gridView().size(0);
 
@@ -677,11 +678,18 @@ namespace Opm {
             }
 
             try {
+                using GLiftEclWells = typename GasLiftGroupInfo<Scalar>::GLiftEclWells;
                 GLiftEclWells ecl_well_map;
-                initGliftEclWellMap(ecl_well_map);
-                well->wellTesting(simulator_, simulationTime, this->wellState(),
-                                  this->groupState(), this->wellTestState(), this->phase_usage_,
-                                  ecl_well_map, this->well_open_times_, deferred_logger);
+                gaslift_.initGliftEclWellMap(well_container_, ecl_well_map);
+                well->wellTesting(simulator_,
+                                  simulationTime,
+                                  this->wellState(),
+                                  this->groupState(),
+                                  this->wellTestState(),
+                                  this->phase_usage_,
+                                  ecl_well_map,
+                                  this->well_open_times_,
+                                  deferred_logger);
             } catch (const std::exception& e) {
                 const std::string msg = fmt::format("Exception during testing of well: {}. The well will not open.\n Exception message: {}", wellEcl.name(), e.what());
                 deferred_logger.warning("WELL_TESTING_FAILED", msg);
@@ -841,8 +849,9 @@ namespace Opm {
     {
         rate = 0;
 
-        if (!is_cell_perforated_[elemIdx])
+        if (!is_cell_perforated_[elemIdx]) {
             return;
+        }
 
         for (const auto& well : well_container_)
             well->addCellRates(rate, elemIdx);
@@ -861,8 +870,9 @@ namespace Opm {
         rate = 0;
         int elemIdx = context.globalSpaceIndex(spaceIdx, timeIdx);
 
-        if (!is_cell_perforated_[elemIdx])
+        if (!is_cell_perforated_[elemIdx]) {
             return;
+        }
 
         for (const auto& well : well_container_)
             well->addCellRates(rate, elemIdx);
@@ -1230,12 +1240,14 @@ namespace Opm {
     assemble(const int iterationIdx,
              const double dt)
     {
-
         DeferredLogger local_deferredLogger;
-        if (this->glift_debug) {
-            const std::string msg = fmt::format(
-                "assemble() : iteration {}" , iterationIdx);
-            this->gliftDebug(msg, local_deferredLogger);
+
+        if constexpr (gaslift_.glift_debug) {
+            if (gaslift_.terminalOutput()) {
+                const std::string msg =
+                    fmt::format("assemble() : iteration {}" , iterationIdx);
+                gaslift_.gliftDebug(msg, local_deferredLogger);
+            }
         }
         last_report_ = SimulatorReportSingle();
         Dune::Timer perfTimer;
@@ -1331,7 +1343,9 @@ namespace Opm {
                                           DeferredLogger& local_deferredLogger)
     {
         auto [well_group_control_changed, more_network_update] =
-                updateWellControls(mandatory_network_balance, local_deferredLogger, relax_network_tolerance);
+                updateWellControls(mandatory_network_balance,
+                                   local_deferredLogger,
+                                   relax_network_tolerance);
 
         bool alq_updated = false;
         OPM_BEGIN_PARALLEL_TRY_CATCH();
@@ -1339,14 +1353,19 @@ namespace Opm {
             // Set the well primary variables based on the value of well solutions
             initPrimaryVariablesEvaluation();
 
-            alq_updated = maybeDoGasLiftOptimize(local_deferredLogger);
+            alq_updated = gaslift_.maybeDoGasLiftOptimize(simulator_,
[…]
