### 2025-01-02 f71a92c
use std::find_if instead of loop
