+++ opm/simulators/timestepping/AdaptiveTimeStepping.hpp
@@ -6,19 +6,11 @@
 #include <dune/common/version.hh>
 #include <dune/istl/istlexception.hh>
 
-#include <opm/common/Exceptions.hpp>
-#include <opm/common/ErrorMacros.hpp>
 #include <opm/common/OpmLog/OpmLog.hpp>
-
-#include <opm/grid/utility/StopWatch.hpp>
-
-#include <opm/input/eclipse/Units/Units.hpp>
-#include <opm/input/eclipse/Units/UnitSystem.hpp>
 
 #include <opm/input/eclipse/Schedule/Tuning.hpp>
 
 #include <opm/models/utils/basicproperties.hh>
-#include <opm/models/utils/parametersystem.hpp>
 #include <opm/models/utils/propertysystem.hh>
 
 #include <opm/simulators/timestepping/AdaptiveSimulatorTimer.hpp>
@@ -28,18 +20,10 @@
 #include <opm/simulators/timestepping/TimeStepControl.hpp>
 #include <opm/simulators/timestepping/TimeStepControlInterface.hpp>
 
-#include <opm/simulators/utils/phaseUsageFromDeck.hpp>
-
-#include <fmt/format.h>
-
-#include <algorithm>
-#include <cassert>
 #include <cmath>
 #include <functional>
 #include <memory>
 #include <set>
-#include <sstream>
-#include <stdexcept>
 #include <string>
 #include <vector>
 
@@ -67,6 +51,7 @@ struct MinTimeStepBasedOnNewtonIterations { static constexpr double value = 0.0;
 
 namespace Opm {
 
+class UnitSystem;
 struct StepReport;
 
 namespace detail {
@@ -114,30 +99,9 @@ void registerAdaptiveParameters();
         AdaptiveTimeStepping() = default;
 
         //! \brief contructor taking parameter object
-        AdaptiveTimeStepping(const UnitSystem& unitSystem,
-                             const double max_next_tstep = -1.0,
-                             const bool terminalOutput = true)
-            : timeStepControl_()
-            , restartFactor_(Parameters::Get<Parameters::SolverRestartFactor<Scalar>>()) // 0.33
-            , growthFactor_(Parameters::Get<Parameters::SolverGrowthFactor<Scalar>>()) // 2.0
-            , maxGrowth_(Parameters::Get<Parameters::SolverMaxGrowth<Scalar>>()) // 3.0
-            , maxTimeStep_(Parameters::Get<Parameters::SolverMaxTimeStepInDays<Scalar>>() * 24 * 60 * 60) // 365.25
-            , minTimeStep_(unitSystem.to_si(UnitSystem::measure::time, Parameters::Get<Parameters::SolverMinTimeStep<Scalar>>())) // 1e-12;
-            , ignoreConvergenceFailure_(Parameters::Get<Parameters::SolverContinueOnConvergenceFailure>()) // false;
-            , solverRestartMax_(Parameters::Get<Parameters::SolverMaxRestarts>()) // 10
-            , solverVerbose_(Parameters::Get<Parameters::SolverVerbosity>() > 0 && terminalOutput) // 2
-            , timestepVerbose_(Parameters::Get<Parameters::TimeStepVerbosity>() > 0 && terminalOutput) // 2
-            , suggestedNextTimestep_((max_next_tstep <= 0 ? Parameters::Get<Parameters::InitialTimeStepInDays>() : max_next_tstep) * 24 * 60 * 60) // 1.0
-            , fullTimestepInitially_(Parameters::Get<Parameters::FullTimeStepInitially>()) // false
-            , timestepAfterEvent_(Parameters::Get<Parameters::TimeStepAfterEventInDays<Scalar>>() * 24 * 60 * 60) // 1e30
-            , useNewtonIteration_(false)
[…]
