+++ opm/simulators/timestepping/AdaptiveTimeStepping_impl.hpp
@@ -1,0 +1,607 @@
+/*
+*/
+#ifndef OPM_ADAPTIVE_TIME_STEPPING_IMPL_HPP
+#define OPM_ADAPTIVE_TIME_STEPPING_IMPL_HPP
+
+// Improve IDE experience
+#ifndef OPM_ADAPTIVE_TIME_STEPPING_HPP
+#include <config.h>
+#include <opm/simulators/timestepping/AdaptiveTimeStepping.hpp>
+#endif
+
+#include <opm/common/Exceptions.hpp>
+#include <opm/common/ErrorMacros.hpp>
+
+#include <opm/grid/utility/StopWatch.hpp>
+
+#include <opm/input/eclipse/Units/Units.hpp>
+#include <opm/input/eclipse/Units/UnitSystem.hpp>
+
+#include <opm/models/utils/parametersystem.hpp>
+
+#include <algorithm>
+#include <cassert>
+#include <cmath>
+#include <sstream>
+#include <stdexcept>
+
+#include <fmt/format.h>
+
+namespace Opm {
+
+template<class TypeTag>
+AdaptiveTimeStepping<TypeTag>::
+AdaptiveTimeStepping(const UnitSystem& unitSystem,
+                     const double max_next_tstep,
+                     const bool terminalOutput)
+    : timeStepControl_()
+    , restartFactor_(Parameters::Get<Parameters::SolverRestartFactor<Scalar>>()) // 0.33
+    , growthFactor_(Parameters::Get<Parameters::SolverGrowthFactor<Scalar>>()) // 2.0
+    , maxGrowth_(Parameters::Get<Parameters::SolverMaxGrowth<Scalar>>()) // 3.0
+    , maxTimeStep_(Parameters::Get<Parameters::SolverMaxTimeStepInDays<Scalar>>() * 24 * 60 * 60) // 365.25
+    , minTimeStep_(unitSystem.to_si(UnitSystem::measure::time, Parameters::Get<Parameters::SolverMinTimeStep<Scalar>>())) // 1e-12;
+    , ignoreConvergenceFailure_(Parameters::Get<Parameters::SolverContinueOnConvergenceFailure>()) // false;
+    , solverRestartMax_(Parameters::Get<Parameters::SolverMaxRestarts>()) // 10
+    , solverVerbose_(Parameters::Get<Parameters::SolverVerbosity>() > 0 && terminalOutput) // 2
+    , timestepVerbose_(Parameters::Get<Parameters::TimeStepVerbosity>() > 0 && terminalOutput) // 2
+    , suggestedNextTimestep_((max_next_tstep <= 0 ? Parameters::Get<Parameters::InitialTimeStepInDays>() : max_next_tstep) * 24 * 60 * 60) // 1.0
+    , fullTimestepInitially_(Parameters::Get<Parameters::FullTimeStepInitially>()) // false
+    , timestepAfterEvent_(Parameters::Get<Parameters::TimeStepAfterEventInDays<Scalar>>() * 24 * 60 * 60) // 1e30
+    , useNewtonIteration_(false)
+    , minTimeStepBeforeShuttingProblematicWells_(Parameters::Get<Parameters::MinTimeStepBeforeShuttingProblematicWellsInDays>() * unit::day)
+
+{
+    init_(unitSystem);
+}
+
+template<class TypeTag>
+AdaptiveTimeStepping<TypeTag>::
+AdaptiveTimeStepping(double max_next_tstep,
+                     const Tuning& tuning,
+                     const UnitSystem& unitSystem,
+                     const bool terminalOutput)
+    : timeStepControl_()
+    , restartFactor_(tuning.TSFCNV)
[…]
