### 2025-01-02 1830cf7
Make a seqential solver for temperature
Use with TEMP
