+++ opm/simulators/flow/FlowProblemBlackoil.hpp
@@ -341,6 +341,13 @@ public:
             this->maxTimeStepAfterWellEvent_ = tuning.TMAXWC;
         }
 
+        // conserve inner energy instead of enthalpy if TEMP is used
+        // or THERMAL and parameter ConserveInnerEnergyThermal is true (default false)
+        bool isThermal = eclState.getSimulationConfig().isThermal();
+        bool isTemp = eclState.getSimulationConfig().isTemp();
+        bool conserveInnerEnergy = isTemp || (isThermal && Parameters::Get<Parameters::ConserveInnerEnergyThermal>());
+        FluidSystem::setEnergyEqualEnthalpy(conserveInnerEnergy);
+
         if (FluidSystem::phaseIsActive(FluidSystem::oilPhaseIdx) &&
             FluidSystem::phaseIsActive(FluidSystem::gasPhaseIdx)) {
             this->maxOilSaturation_.resize(this->model().numGridDof(), 0.0);
@@ -374,7 +381,7 @@ public:
         else {
             this->readInitialCondition_();
         }
-
+        this->temperatureModel_.init();
         this->tracerModel_.prepareTracerBatches();
 
         this->updatePffDofData_();
