+++ opm/simulators/flow/MainDispatchDynamic.cpp
@@ -27,6 +27,7 @@
 #include <flow/flow_biofilm.hpp>
 #include <flow/flow_blackoil.hpp>
 #include <flow/flow_blackoil_legacyassembly.hpp>
+#include <flow/flow_blackoil_temp.hpp>
 #include <flow/flow_brine.hpp>
 #include <flow/flow_brine_precsalt_vapwat.hpp>
 #include <flow/flow_brine_saltprecipitation.hpp>
@@ -74,6 +75,7 @@ int Opm::Main::dispatchDynamic_()
     // TODO: make sure that no illegal combinations like thermal and
     //       twophase are requested.
     const bool thermal = eclipseState_->getSimulationConfig().isThermal();
+    const bool temp = eclipseState_->getSimulationConfig().isTemp();
 
     // Single-phase case
     if (rspec.micp()) {
@@ -128,6 +130,11 @@ int Opm::Main::dispatchDynamic_()
     // Energy case
     else if (thermal) {
         return this->runThermal(phases);
+    }
+
+    // Blackoil case with temperature
+    else if (phases.size() == 4 && temp) {
+        return this->runBlackOilTemp();
     }
 
     // Blackoil case
@@ -423,6 +430,12 @@ int Opm::Main::runBlackOil()
         // support the diffusion module yet.
         return flowBlackoilMain(argc_, argv_, outputCout_, outputFiles_);
     }
-
     return flowBlackoilTpfaMain(argc_, argv_, outputCout_, outputFiles_);
+}
+
+int Opm::Main::runBlackOilTemp()
+{
+    // TEMP is used. Energy equation is solved seperatly
+    // Only 3p-blackoil supported with TEMP option
+    return flowBlackoilTempMain(argc_, argv_, outputCout_, outputFiles_);
 }
