### 2025-01-02 87b5233
Make a seqential solver for temperature
Use with TEMP
