### 2025-01-02 bad7419
comile and dont segfault
