+++ opm/simulators/flow/TemperatureModel.hpp
@@ -1,0 +1,228 @@
+// -*- mode: C++; tab-width: 4; indent-tabs-mode: nil; c-basic-offset: 4 -*-
+// vi: set et ts=4 sw=4 sts=4:
+/*
+  This file is part of the Open Porous Media project (OPM).
+
+  OPM is free software: you can redistribute it and/or modify
+  it under the terms of the GNU General Public License as published by
+  the Free Software Foundation, either version 2 of the License, or
+  (at your option) any later version.
+
+  OPM is distributed in the hope that it will be useful,
+  but WITHOUT ANY WARRANTY; without even the implied warranty of
+  MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
+  GNU General Public License for more details.
+
+  You should have received a copy of the GNU General Public License
+  along with OPM.  If not, see <http://www.gnu.org/licenses/>.
+
+  Consult the COPYING file in the top-level source directory of this
+  module for the precise wording of the license and the list of
+  copyright holders.
+*/
+/**
+ * \file
+ *
+ * \copydoc Opm::TemperatureModel
+ */
+#ifndef OPM_TEMPERATURE_MODEL_HPP
+#define OPM_TEMPERATURE_MODEL_HPP
+
+#include <opm/common/OpmLog/OpmLog.hpp>
+
+#include <opm/models/utils/propertysystem.hh>
+
+#include <opm/simulators/flow/GenericTemperatureModel.hpp>
+//#include <opm/simulators/utils/VectorVectorDataHandle.hpp>
+
+#include <array>
+#include <cstddef>
+#include <memory>
+#include <stdexcept>
+#include <string>
+#include <vector>
+
+namespace Opm::Properties {
+
+template<class TypeTag, class MyTypeTag>
+struct EnableTemperatureModel {
+    using type = UndefinedProperty;
+};
+
+} // namespace Opm::Properties
+
+namespace Opm {
+
+/*!
+ * \ingroup BlackOilSimulator
+ *
+ * \brief A class which handles sequential implicit solution of the energy equation as specified in by TEMP
+ */
+template <class TypeTag>
+class TemperatureModel : public GenericTemperatureModel<GetPropType<TypeTag, Properties::Grid>,
+                                              GetPropType<TypeTag, Properties::GridView>,
+                                              GetPropType<TypeTag, Properties::DofMapper>,
+                                              GetPropType<TypeTag, Properties::Stencil>,
+                                              GetPropType<TypeTag, Properties::FluidSystem>,
+                                              GetPropType<TypeTag, Properties::Scalar>>
+{
+    using BaseType = GenericTemperatureModel<GetPropType<TypeTag, Properties::Grid>,
+                                        GetPropType<TypeTag, Properties::GridView>,
+                                        GetPropType<TypeTag, Properties::DofMapper>,
+                                        GetPropType<TypeTag, Properties::Stencil>,
+                                        GetPropType<TypeTag, Properties::FluidSystem>,
+                                        GetPropType<TypeTag, Properties::Scalar>>;
+    using Simulator = GetPropType<TypeTag, Properties::Simulator>;
+    using GridView = GetPropType<TypeTag, Properties::GridView>;
+    using Grid = GetPropType<TypeTag, Properties::Grid>;
+    using Scalar = GetPropType<TypeTag, Properties::Scalar>;
+    using Stencil = GetPropType<TypeTag, Properties::Stencil>;
+    using FluidSystem = GetPropType<TypeTag, Properties::FluidSystem>;
+    using ElementContext = GetPropType<TypeTag, Properties::ElementContext>;
[…]
