+++ opm/simulators/flow/FlowProblem.hpp
@@ -61,6 +61,7 @@
 #include <opm/simulators/flow/FlowBaseProblemProperties.hpp>
 #include <opm/simulators/flow/FlowUtils.hpp>
 #include <opm/simulators/flow/TracerModel.hpp>
+#include <opm/simulators/flow/TemperatureModel.hpp>
 #include <opm/simulators/flow/Transmissibility.hpp>
 #include <opm/simulators/timestepping/AdaptiveTimeStepping.hpp>
 #include <opm/simulators/timestepping/SimulatorReport.hpp>
@@ -162,7 +163,7 @@ protected:
     using Toolbox = MathToolbox<Evaluation>;
     using DimMatrix = Dune::FieldMatrix<Scalar, dimWorld, dimWorld>;
 
-
+    using TemperatureModel = GetPropType<TypeTag, Properties::TemperatureModel>;
     using TracerModel = GetPropType<TypeTag, Properties::TracerModel>;
     using DirectionalMobilityPtr = Utility::CopyablePtr<DirectionalMobility<TypeTag, Evaluation>>;
 
@@ -224,6 +225,7 @@ public:
         , aquiferModel_(simulator)
         , pffDofData_(simulator.gridView(), this->elementMapper())
         , tracerModel_(simulator)
+        , temperatureModel_(simulator)
     {
         this->enableDriftCompensation_ = Parameters::Get<Parameters::EnableDriftCompensation>();
         this->enableVtkOutput_ = Parameters::Get<Parameters::EnableVtkOutput>();
@@ -391,6 +393,7 @@ public:
         wellModel_.beginTimeStep();
         aquiferModel_.beginTimeStep();
         tracerModel_.beginTimeStep();
+        temperatureModel_.beginTimeStep();
 
     }
 
@@ -446,6 +449,7 @@ public:
         this->wellModel_.endTimeStep();
         this->aquiferModel_.endTimeStep();
         this->tracerModel_.endTimeStep();
+        this->temperatureModel_.endTimeStep();
 
         // Compute flux for output
         this->model().linearizer().updateFlowsInfo();
@@ -1192,6 +1196,7 @@ public:
         serializer(wellModel_);
         serializer(aquiferModel_);
         serializer(tracerModel_);
+        serializer(temperatureModel_);
         serializer(*materialLawManager_);
     }
 
@@ -1379,7 +1384,7 @@ protected:
 
     void readThermalParameters_()
     {
-        if constexpr (enableEnergy)
+        if constexpr (true)
         {
             const auto& simulator = this->simulator();
             const auto& vanguard = simulator.vanguard();
@@ -1693,6 +1698,7 @@ protected:
 
     PffGridVector<GridView, Stencil, PffDofData_, DofMapper> pffDofData_;
     TracerModel tracerModel_;
+    TemperatureModel temperatureModel_;
 
     template<class T>
     struct BCData
@@ -1737,3 +1743,4 @@ protected:
 } // namespace Opm
 
 #endif // OPM_FLOW_PROBLEM_HPP
+
