### 2025-01-02 0298b27
Make a seqential solver for temperature
Use with TEMP
