+++ flowexperimental/BlackOilEnergyIntensiveQuantitiesGlobalIndex.hpp
@@ -121,13 +121,11 @@ public:
                             [[maybe_unused]] unsigned globalSpaceIdx,
                             [[maybe_unused]] unsigned timeIdx)
     {
-        if constexpr (enableTemperature) {
-            // even if energy is conserved, the temperature can vary over the spatial
-            // domain if the EnableTemperature property is set to true
-            auto& fs = this->asImp_().fluidState_;
-            Scalar T = problem.temperature(globalSpaceIdx, timeIdx);
-            fs.setTemperature(T);
-        }
+        // even if energy is conserved, the temperature can vary over the spatial
+        // domain
+        auto& fs = this->asImp_().fluidState_;
+        Scalar T = problem.temperature(globalSpaceIdx, timeIdx);
+        fs.setTemperature(T);
     }
 
     void updateEnergyQuantities_([[maybe_unused]] const Problem& problem,
