+++ opm/simulators/flow/MainDispatchDynamic.cpp
@@ -26,6 +26,7 @@
 
 #include <flow/flow_blackoil.hpp>
 #include <flow/flow_blackoil_legacyassembly.hpp>
+#include <flow/flow_blackoil_temp.hpp>
 #include <flow/flow_brine.hpp>
 #include <flow/flow_brine_precsalt_vapwat.hpp>
 #include <flow/flow_brine_saltprecipitation.hpp>
@@ -73,6 +74,7 @@ int Opm::Main::dispatchDynamic_()
     // TODO: make sure that no illegal combinations like thermal and
     //       twophase are requested.
     const bool thermal = eclipseState_->getSimulationConfig().isThermal();
+    const bool temp = eclipseState_->getSimulationConfig().isTemp();
 
     // Single-phase case
     if (rspec.micp()) {
@@ -122,6 +124,11 @@ int Opm::Main::dispatchDynamic_()
     // Energy case
     else if (thermal) {
         return this->runThermal(phases);
+    }
+
+    // Blackoil case with temperature
+    else if (phases.size() == 4 && temp) {
+        return this->runBlackOilTemp();
     }
 
     // Blackoil case
@@ -400,6 +407,12 @@ int Opm::Main::runBlackOil()
         // support the diffusion module yet.
         return flowBlackoilMain(argc_, argv_, outputCout_, outputFiles_);
     }
-
     return flowBlackoilTpfaMain(argc_, argv_, outputCout_, outputFiles_);
+}
+
+int Opm::Main::runBlackOilTemp()
+{
+    // TEMP is used. Energy equation is solved seperatly
+    // Only 3p-blackoil supported with TEMP option
+    return flowBlackoilTempMain(argc_, argv_, outputCout_, outputFiles_);
 }
