+++ opm/simulators/wells/StandardWell_impl.hpp
@@ -2618,7 +2618,6 @@ namespace Opm
                 result += Base::restrictEval(cq_r_thermal);
             }
         }
-
         return result * this->well_efficiency_factor_;
     }
 } // namespace Opm
