### 2025-01-02 288e97d
Make a seqential solver for temperature
Use with TEMP
