+++ opm/simulators/flow/FlowBaseProblemProperties.hpp
@@ -40,6 +40,7 @@
 #include <opm/simulators/flow/EclWriter.hpp>
 #include <opm/simulators/flow/FlowProblemParameters.hpp>
 #include <opm/simulators/flow/TracerModel.hpp>
+#include <opm/simulators/flow/TemperatureModel.hpp>
 
 #if HAVE_DAMARIS
 #include <opm/simulators/flow/DamarisWriter.hpp>
@@ -89,6 +90,12 @@ struct TracerModel {  using type = UndefinedProperty; };
 template <class TypeTag>
 struct TracerModel<TypeTag, TTag::FlowBaseProblem>
 { using type =  ::Opm::TracerModel<TypeTag>; };
+
+template<class TypeTag, class MyTypeTag>
+struct TemperatureModel {  using type = UndefinedProperty; };
+template <class TypeTag>
+struct TemperatureModel<TypeTag, TTag::FlowBaseProblem>
+{ using type =  ::Opm::TemperatureModel<TypeTag>; };
 
 // Select the element centered finite volume method as spatial discretization
 template<class TypeTag>
