+++ opm/simulators/flow/FlowProblemBlackoil.hpp
@@ -362,6 +362,7 @@ public:
 
         const auto& initconfig = eclState.getInitConfig();
         this->tracerModel_.init(initconfig.restartRequested());
+        this->temperatureModel_.init(initconfig.restartRequested());
         if (initconfig.restartRequested()) {
             this->readEclRestartSolution_();
         }
@@ -1162,8 +1163,9 @@ protected:
 
             // For CO2STORE and H2STORE we need to set the initial temperature for isothermal simulations
             bool isThermal = eclState.getSimulationConfig().isThermal();
+            bool isTemp = eclState.getSimulationConfig().isTemp();
             bool needTemperature = (eclState.runspec().co2Storage() || eclState.runspec().h2Storage());
-            if (!isThermal && needTemperature) {
+            if (!isThermal && !isTemp && needTemperature) {
                 const auto& fp = simulator.vanguard().eclState().fieldProps();
                 elemFluidState.setTemperature(fp.get_double("TEMPI")[elemIdx]);
             }
