### 2025-01-02 1556169
changed: do no assert for user provided data
instead throw an exception
