### 2025-01-02 f0cb059
remove unnecessary std::to_string
format can handle floating point numbers
