+++ opm/simulators/timestepping/AdaptiveTimeStepping_impl.hpp
@@ -544,62 +544,86 @@ template<class TypeTag>
 void AdaptiveTimeStepping<TypeTag>::
 init_(const UnitSystem& unitSystem)
 {
-    // valid are "pid" and "pid+iteration"
-    std::string control = Parameters::Get<Parameters::TimeStepControl>(); // "pid"
-
     const double tol =  Parameters::Get<Parameters::TimeStepControlTolerance>(); // 1e-1
-    if (control == "pid") {
-        timeStepControl_ = std::make_unique<PIDTimeStepControl>(tol);
-        timeStepControlType_ = TimeStepControlType::PID;
-    }
-    else if (control == "pid+iteration") {
-        const int iterations =  Parameters::Get<Parameters::TimeStepControlTargetIterations>(); // 30
-        const double decayDampingFactor = Parameters::Get<Parameters::TimeStepControlDecayDampingFactor>(); // 1.0
-        const double growthDampingFactor = Parameters::Get<Parameters::TimeStepControlGrowthDampingFactor>(); // 3.2
-        timeStepControl_ = std::make_unique<PIDAndIterationCountTimeStepControl>(iterations, decayDampingFactor, growthDampingFactor, tol);
-        timeStepControlType_ = TimeStepControlType::PIDAndIterationCount;
-    }
-    else if (control == "pid+newtoniteration") {
-        const int iterations =  Parameters::Get<Parameters::TimeStepControlTargetNewtonIterations>(); // 8
-        const double decayDampingFactor = Parameters::Get<Parameters::TimeStepControlDecayDampingFactor>(); // 1.0
-        const double growthDampingFactor = Parameters::Get<Parameters::TimeStepControlGrowthDampingFactor>(); // 3.2
-        const double nonDimensionalMinTimeStepIterations = Parameters::Get<Parameters::MinTimeStepBasedOnNewtonIterations>(); // 0.0 by default
-        // the min time step can be reduced by the newton iteration numbers
-        double minTimeStepReducedByIterations = unitSystem.to_si(UnitSystem::measure::time, nonDimensionalMinTimeStepIterations);
-        timeStepControl_ = std::make_unique<PIDAndIterationCountTimeStepControl>(iterations, decayDampingFactor,
-                                                                                 growthDampingFactor, tol, minTimeStepReducedByIterations);
-        timeStepControlType_ = TimeStepControlType::PIDAndIterationCount;
-        useNewtonIteration_ = true;
-    }
-    else if (control == "iterationcount") {
-        const int iterations =  Parameters::Get<Parameters::TimeStepControlTargetIterations>(); // 30
-        const double decayrate = Parameters::Get<Parameters::TimeStepControlDecayRate>(); // 0.75
-        const double growthrate = Parameters::Get<Parameters::TimeStepControlGrowthRate>(); // 1.25
-        timeStepControl_ = std::make_unique<SimpleIterationCountTimeStepControl>(iterations, decayrate, growthrate);
-        timeStepControlType_ = TimeStepControlType::SimpleIterationCount;
-    }
-    else if (control == "newtoniterationcount") {
-        const int iterations =  Parameters::Get<Parameters::TimeStepControlTargetNewtonIterations>(); // 8
-        const double decayrate = Parameters::Get<Parameters::TimeStepControlDecayRate>(); // 0.75
-        const double growthrate = Parameters::Get<Parameters::TimeStepControlGrowthRate>(); // 1.25
-        timeStepControl_ = std::make_unique<SimpleIterationCountTimeStepControl>(iterations, decayrate, growthrate);
-        useNewtonIteration_ = true;
-        timeStepControlType_ = TimeStepControlType::SimpleIterationCount;
-    }
-    else if (control == "hardcoded") {
[…]
