### 2025-01-03 0ee1d2f
changed: use standard header guard structure for _impl.hpp clangd helpers
