+++ opm/simulators/wells/StandardWell.hpp
@@ -471,8 +471,6 @@ namespace Opm
 
 }
 
-#ifndef OPM_STANDARDWELL_IMPL_HEADER_INCLUDED
 #include "StandardWell_impl.hpp"
-#endif
 
 #endif // OPM_STANDARDWELL_HEADER_INCLUDED
