+++ opm/simulators/wells/MultisegmentWell.hpp
@@ -324,8 +324,6 @@ namespace Opm {
 
 }
 
-#ifndef OPM_MULTISEGMENTWELL_IMPL_HEADER_INCLUDED
 #include "MultisegmentWell_impl.hpp"
-#endif
 
 #endif // OPM_MULTISEGMENTWELL_HEADER_INCLUDED
