### 2025-01-03 79a85a7
still issues, starting the process of making flow compile and run
