### 2025-01-03 2f961ec
still issues, starting the process of making flow compile and run
