### 2025-01-04 1138a18
Cleanup after rebase on master
After rebasing on master some changes to AdaptiveTimeStepping.hpp and
AdaptiveTimeStepping_impl.hpp were missed
