+++ opm/simulators/timestepping/AdaptiveTimeStepping_impl.hpp
@@ -19,9 +19,31 @@
 
 #ifndef OPM_ADAPTIVE_TIME_STEPPING_IMPL_HPP
 #define OPM_ADAPTIVE_TIME_STEPPING_IMPL_HPP
+
+// Improve IDE experience
+#ifndef OPM_ADAPTIVE_TIME_STEPPING_HPP
+#include <config.h>
 #include <opm/simulators/timestepping/AdaptiveTimeStepping.hpp>
 #endif
 
+#include <opm/common/Exceptions.hpp>
+#include <opm/common/ErrorMacros.hpp>
+
+#include <opm/grid/utility/StopWatch.hpp>
+
+#include <opm/input/eclipse/Units/Units.hpp>
+#include <opm/input/eclipse/Units/UnitSystem.hpp>
+
+#include <opm/models/utils/parametersystem.hpp>
+
+#include <algorithm>
+#include <cassert>
+#include <cmath>
+#include <sstream>
+#include <stdexcept>
+
+#include <boost/date_time/posix_time/posix_time.hpp>
+#include <fmt/format.h>
 
 namespace Opm {
 /*********************************************
@@ -388,7 +410,7 @@ init_(const UnitSystem& unitSystem)
              time_step_control_,
              use_newton_iteration_) = detail::createController(unitSystem);
     // make sure growth factor is something reasonable
-    if (growthFactor_ < 1.0) {
+    if (this->growth_factor_ < 1.0) {
         OPM_THROW(std::runtime_error,
                   "Growth factor cannot be less than 1.");
     }
@@ -1246,3 +1268,5 @@ double AdaptiveTimeStepping<TypeTag>::SolutionTimeErrorSolverWrapper<Solver>::re
 }
 
 } // namespace Opm
+
+#endif // OPM_ADAPTIVE_TIME_STEPPING_IMPL_HPP
