### 2025-01-06 878b399
make a blackoil temp variant
