+++ opm/simulators/wells/MultisegmentWellEval.hpp
@@ -62,7 +62,7 @@ protected:
     //                                                         EvalR (Eval), EvalW, EvalRW
     // TODO: for now, we only use one type to save some implementation efforts, while improve later.
     using EvalWell = typename PrimaryVariables::EvalWell;
-    using Eval = DenseAd::Evaluation<Scalar, /*size=*/Indices::numEq+1>;
+    using Eval = DenseAd::Evaluation<Scalar, /*size=*/Indices::numDerivatives>;
     //using Eval = GetPropType<TypeTag, Properties::Evaluation>;
 
 public:
