### 2025-01-06 d3dac2a
make a blackoil temp variant
