+++ opm/simulators/wells/StandardWellEval.hpp
@@ -60,7 +60,7 @@ protected:
 
 public:
     using EvalWell = typename PrimaryVariables::EvalWell;
-    using Eval = DenseAd::Evaluation<Scalar, Indices::numEq+1>;
+    using Eval = DenseAd::Evaluation<Scalar, Indices::numDerivatives>;
     //using Eval = GetPropType<TypeTag, Properties::Evaluation>;
     using BVectorWell = typename StandardWellEquations<Scalar,Indices::numEq>::BVectorWell;
 
