### 2025-01-06 08d9805
make a blackoil temp variant
