### 2025-01-06 b04fa06
make a blackoil temp variant
