### 2025-01-06 e4e2c45
gives same results as thermal for spe1case2_temp/termal
