+++ opm/simulators/flow/TemperatureModel.hpp
@@ -118,6 +118,8 @@ public:
         for (unsigned globI = 0; globI < numCells; ++globI) {
             this->temperature_[globI] = simulator_.problem().initialFluidState(globI).temperature(0);
         }
+     
+        intQuants_.resize(numCells);
     }
 
     void beginTimeStep()
@@ -125,6 +127,11 @@ public:
         if (!this->doTemp())
             return;
 
+        // We copy the intensive quantities here to make it possible to update them
+        const unsigned int numCells = simulator_.model().numTotalDof();
+        for (unsigned globI = 0; globI < numCells; ++globI) {
+            intQuants_[globI] = simulator_.model().intensiveQuantities(globI, /*timeIdx*/ 0);
+        }
         updateStorageCache();
     }
 
@@ -136,6 +143,11 @@ public:
         if (!this->doTemp())
             return;
 
+        // We copy the intensive quantities here to make it possible to update them
+        const unsigned int numCells = simulator_.model().numTotalDof();
+        for (unsigned globI = 0; globI < numCells; ++globI) {
+            intQuants_[globI] = simulator_.model().intensiveQuantities(globI, /*timeIdx*/ 0);
+        }
         advanceTemperatureFields();
     }
 
@@ -182,6 +194,7 @@ protected:
     void advanceTemperatureFields()
     {
 
+
         for (int iter = 0; iter < 20; ++iter) {
         this->energyVector_ = 0.0;
         (*this->energyMatrix_) = 0.0;
@@ -198,7 +211,7 @@ protected:
             computeStorageTerm(globI, storage); 
             this->energyVector_[globI] += storefac * ( getValue(storage) - storage0_[globI] );
             if (globI == 0) {
-                std::cout << "storage " << storefac * ( getValue(storage) - storage0_[globI] ) << " " <<  storefac * storage.derivative(Indices::temperatureIdx) << std::endl;
+                std::cout << "storage " << getValue(storage) << " " << storage0_[globI] << " " << storefac * ( getValue(storage) - storage0_[globI] ) << " " <<  storefac * storage.derivative(Indices::temperatureIdx) << " " << dt/(3600*24) << std::endl;
             }
             (*this->energyMatrix_)[globI][globI][0][0] += storefac * storage.derivative(Indices::temperatureIdx); 
         }
@@ -211,13 +224,13 @@ protected:
             RateVector tmp(0.0);
             RateVector darcyFlux(0.0);
             const auto& nbInfos = neighborInfo[globI];
-            const IntensiveQuantities& intQuantsIn = simulator_.model().intensiveQuantities(globI, /*timeIdx*/ 0);
+            const IntensiveQuantities& intQuantsIn = intQuants_[globI];
             for (const auto& nbInfo : nbInfos) {
                 unsigned globJ = nbInfo.neighbor;
                 assert(globJ != globI);
                 tmp = 0.0;
                 darcyFlux = 0.0;
-                const IntensiveQuantities& intQuantsEx = simulator_.model().intensiveQuantities(globJ, /*timeIdx*/ 0);
+                const IntensiveQuantities& intQuantsEx = intQuants_[globJ];
                 LocalResidual::computeFlux(tmp, darcyFlux, globI, globJ, intQuantsIn, intQuantsEx, nbInfo.res_nbinfo, simulator_.problem().moduleParams());
                 //adres *= nbInfo.res_nbinfo.faceArea;
 
@@ -264,7 +277,7 @@ protected:
             sumNorm += std::abs(this->energyVector_[globI]);
         }
 
-        if (maxNorm < 1e-6 || sumNorm/numCells < 1e-8) {
[…]
