### 2025-01-06 7c5f9c5
BlackoilWellModel: move initialization of terminal_output_ to the generic class
it is now possible since parameter system does not need typetag
