### 2025-01-06 9c4eeab
gives same results as thermal for spe1case2_temp/termal
