### 2025-01-06 ae45913
gives same results as thermal for spe1case2_temp/termal
