### 2025-01-07 7f86a67
remove unnecessary template argument
