+++ tests/gpuistl/test_gpuPvt.cu
@@ -92,17 +92,17 @@ __global__ void gpuEvaluateUniformTabulated2DFunction(GpuTab gpuTab, Evaluation*
 
 // Kernel using a CO2 object on the GPU
 __global__ void gpuCO2GasViscosity(GpuViewCO2Tables gpuViewCo2Tables, Evaluation* temp, Evaluation* pressure, double* result) {
-    *result = GpuCO2::gasViscosity<Evaluation>(gpuViewCo2Tables, *temp, *pressure, true).value();
+    *result = GpuCO2::gasViscosity(gpuViewCo2Tables, *temp, *pressure, true).value();
 }
 
 // Kernel using a SimpleHuDuanH20 object on a GPU
 __global__ void huDuanLiquidDensity(Evaluation* temp, Evaluation* pressure, double* result) {
-    *result = HuDuan::liquidDensity<Evaluation>(*temp, *pressure, true).value();
+    *result = HuDuan::liquidDensity(*temp, *pressure, true).value();
 }
 
 // Kernel using a BrineDynamic object on a GPU
 __global__ void brineDynamicLiquidEnthalpy(Evaluation* temp, Evaluation* pressure, Evaluation* salinity, double* result) {
-    *result = BrineDyn::liquidEnthalpy<Evaluation>(*temp, *pressure, *salinity).value();
+    *result = BrineDyn::liquidEnthalpy(*temp, *pressure, *salinity).value();
 }
 
 // Kernel using a Brine_CO2 object on a GPU
@@ -112,12 +112,12 @@ __global__ void brineCO2GasDiffCoeff(GpuViewCO2Tables co2tables, Evaluation* tem
 
 // Kernel using a Co2GasPvt object on a GPU
 __global__ void co2GasPvtInternalEnergy(GpuViewCo2Pvt gpuViewCo2Pvt, Evaluation* temp, Evaluation* pressure, double* result) {
-    *result = gpuViewCo2Pvt.internalEnergy<Evaluation>(1, *temp, *pressure, Evaluation(0.4), Evaluation(0.0)).value();
+    *result = gpuViewCo2Pvt.internalEnergy(1, *temp, *pressure, Evaluation(0.4), Evaluation(0.0)).value();
 }
 
 // Kernel using a BrineCo2Pvt object on a GPU
 __global__ void brineCo2PvtInternalEnergy(GpuViewBrineCo2Pvt gpuViewBrineCo2Pvt, Evaluation* temp, Evaluation* pressure, Evaluation* rs, Evaluation* saltConcentration, double* result) {
-    *result = gpuViewBrineCo2Pvt.internalEnergy<Evaluation>(1, *temp, *pressure, *rs, *saltConcentration).value();
+    *result = gpuViewBrineCo2Pvt.internalEnergy(1, *temp, *pressure, *rs, *saltConcentration).value();
 }
 
 // Helper function to launch a kernel and retrieve the result on the CPU to reduce code duplicatoin
@@ -202,7 +202,7 @@ BOOST_FIXTURE_TEST_CASE(TestEvaluateUniformTabulated2DFunctionOnGpu, Fixture) {
 BOOST_FIXTURE_TEST_CASE(TestUseCO2OnGpu, Fixture) {
 
     // use the CO2 tables to aquire the viscosity at 290[K] and 2e5[Pa]
-    double viscosityReference = Opm::CO2<double, Opm::CO2Tables<double, std::vector<double>>>::gasViscosity<Evaluation>(co2Tables, temp, pressure, true).value();
[…]
