### 2025-01-07 ba9f3bc
add conduction
