+++ opm/simulators/wells/StandardWell_impl.hpp
@@ -2647,7 +2647,7 @@ namespace Opm
                 result += Base::restrictEval(cq_r_thermal);
             }
         }
-        std::cout << result << std::endl;
+        //std::cout << result << std::endl;
         return result * this->well_efficiency_factor_;
     }
 } // namespace Opm
