+++ opm/simulators/flow/FlowProblem.hpp
@@ -217,7 +217,7 @@ public:
                               simulator.vanguard().cartesianIndexMapper(),
                               simulator.vanguard().grid(),
                               simulator.vanguard().cellCentroids(),
-                              enableEnergy,
+                              enableEnergy || enableTemperature,
                               enableDiffusion,
                               enableDispersion)
         , wellModel_(simulator)
