### 2025-01-07 ca25664
add conduction
