+++ opm/simulators/flow/TemperatureModel.hpp
@@ -189,7 +189,6 @@ protected:
             computeStorageTerm(globI, storage);
             storage0_[globI] = storage;
         }
-        std::cout << "updateStorageCache" << std::endl;
     }
 
     void advanceTemperatureFields()
@@ -212,7 +211,7 @@ protected:
             computeStorageTerm(globI, storage); 
             this->energyVector_[globI] += storefac * ( getValue(storage) - storage0_[globI] );
             if (globI == 0) {
-                std::cout << "storage " << getValue(storage) << " " << storage0_[globI] << " " << storefac * ( getValue(storage) - storage0_[globI] ) << " " <<  storefac * storage.derivative(Indices::temperatureIdx) << " " << dt/(3600*24) << std::endl;
+                //std::cout << "storage " << getValue(storage) << " " << storage0_[globI] << " " << storefac * ( getValue(storage) - storage0_[globI] ) << " " <<  storefac * storage.derivative(Indices::temperatureIdx) << " " << dt/(3600*24) << std::endl;
             }
             (*this->energyMatrix_)[globI][globI][0][0] += storefac * storage.derivative(Indices::temperatureIdx); 
         }
@@ -256,35 +255,36 @@ protected:
             
                 this->energyVector_[globI] += getValue(flux);
                 if (globI == 0) {
-                    std::cout << "flux: " << globJ << " " << flux << std::endl;
+                    //std::cout << "flux: " << globJ << " " << flux << std::endl;
                 }
 
                 (*this->energyMatrix_)[globI][globI][0][0] += flux.derivative(Indices::temperatureIdx);
                 (*this->energyMatrix_)[globJ][globI][0][0] -= flux.derivative(Indices::temperatureIdx);
 
+                const Scalar inAlpha = simulator_.problem().thermalHalfTransmissibility(globI, globJ);
+                const Scalar outAlpha = simulator_.problem().thermalHalfTransmissibility(globJ, globI);
+                Evaluation heatFlux;
+                {
+                    short interiorDofIdx = 0; // NB
+                    short exteriorDofIdx = 1; // NB
 
-            //const Scalar inAlpha = nbInfo.inAlpha;
-            //const Scalar outAlpha = nbInfo.outAlpha;
-            const Scalar inAlpha = simulator_.problem().thermalHalfTransmissibility(globI, globJ);
-            const Scalar outAlpha = simulator_.problem().thermalHalfTransmissibility(globJ, globI);
-            std::cout << inAlpha << " " << outAlpha << std::endl;
-            Evaluation heatFlux;
-            {
-                short interiorDofIdx = 0; // NB
-                short exteriorDofIdx = 1; // NB
-
-                EnergyModule::ExtensiveQuantities::template updateEnergy(heatFlux,
-                                                                         interiorDofIdx, // focusDofIndex,
-                                                                         interiorDofIdx,
-                                                                         exteriorDofIdx,
-                                                                         intQuantsIn,
-                                                                         intQuantsEx,
-                                                                         intQuantsIn.fluidState(),
-                                                                         intQuantsEx.fluidState(),
-                                                                         inAlpha,
-                                                                         outAlpha,
-                                                                         nbInfo.res_nbinfo.faceArea);
-            }
+                    EnergyModule::ExtensiveQuantities::template updateEnergy(heatFlux,
+                                                                            interiorDofIdx, // focusDofIndex,
+                                                                            interiorDofIdx,
[…]
