### 2025-01-07 733e24f
Fix clang complaints about incorrect constexpr usage.
Also make one more if constexpr.
