### 2025-01-07 a840ece
changed: move NLDD BlackoilWellModel code to separate class
