### 2025-01-07 0782495
changed: move NLDD BlackoilWellModel code to separate class
