+++ opm/simulators/wells/BlackoilWellModelNldd.cpp
@@ -1,0 +1,170 @@
+/*
+  Copyright 2016 - 2019 SINTEF Digital, Mathematics & Cybernetics.
+  Copyright 2016 - 2018 Equinor ASA.
+  Copyright 2017 Dr. Blatt - HPC-Simulation-Software & Services
+  Copyright 2016 - 2018 Norce AS
+
+  This file is part of the Open Porous Media project (OPM).
+
+  OPM is free software: you can redistribute it and/or modify
+  it under the terms of the GNU General Public License as published by
+  the Free Software Foundation, either version 3 of the License, or
+  (at your option) any later version.
+
+  OPM is distributed in the hope that it will be useful,
+  but WITHOUT ANY WARRANTY; without even the implied warranty of
+  MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
+  GNU General Public License for more details.
+
+  You should have received a copy of the GNU General Public License
+  along with OPM.  If not, see <http://www.gnu.org/licenses/>.
+*/
+
+#include <config.h>
+#include <opm/simulators/wells/BlackoilWellModelNldd.hpp>
+
+namespace Opm {
+
+template <typename Scalar>
+std::vector<Scalar>
+BlackoilWellModelNlddGeneric<Scalar>::
+getPrimaryVarsDomain(const int domainIdx) const
+{
+    std::vector<Scalar> ret;
+    for (const auto& well : genWellModel_.genericWells()) {
+        if (this->well_domain_.at(well->name()) == domainIdx) {
+            const auto& pv = well->getPrimaryVars();
+            ret.insert(ret.end(), pv.begin(), pv.end());
+        }
+    }
+    return ret;
+}
+
+template <typename Scalar>
+void
+BlackoilWellModelNlddGeneric<Scalar>::
+setPrimaryVarsDomain(const int domainIdx, const std::vector<Scalar>& vars)
+{
+    std::size_t offset = 0;
+    for (const auto& well : genWellModel_.genericWells()) {
+        if (this->well_domain_.at(well->name()) == domainIdx) {
+            int num_pri_vars = well->setPrimaryVars(vars.begin() + offset);
+            offset += num_pri_vars;
+        }
+    }
+    assert(offset == vars.size());
+}
+
+template <typename Scalar>
+void
+BlackoilWellModelNlddGeneric<Scalar>::
+findWellDomains(const std::vector<const SubDomainIndices*>& domains)
+{
+    // TODO: This loop nest may be slow for very large numbers of
+    // domains and wells, but that has not been observed on tests so
+    // far.  Using the partition vector instead would be faster if we
+    // need to change.
+    for (const auto& wellPtr : genWellModel_.genericWells()) {
+        const int first_well_cell = wellPtr->cells().front();
+        for (const auto& domain : domains) {
+            auto cell_present = [domain](const auto cell)
+            {
+                return std::binary_search(domain->cells.begin(),
+                                          domain->cells.end(), cell);
+            };
+
+            if (cell_present(first_well_cell)) {
+                // Assuming that if the first well cell is found in a domain,
+                // then all of that well's cells are in that same domain.
+                this->well_domain_[wellPtr->name()] = domain->index;
+
+                // Verify that all of that well's cells are in that same domain.
+                for (int well_cell : wellPtr->cells()) {
+                    if (! cell_present(well_cell)) {
+                        OPM_THROW(std::runtime_error,
+                                  fmt::format("Well '{}' found on multiple domains.",
+                                              wellPtr->name()));
+                    }
+                }
+            }
+        }
+    }
+}
+
+template <typename Scalar>
[…]
