+++ opm/simulators/wells/BlackoilWellModel.hpp
@@ -30,8 +30,6 @@
 
 #include <opm/common/OpmLog/OpmLog.hpp>
 
-#include <opm/grid/utility/SparseTable.hpp>
-
 #include <opm/input/eclipse/Schedule/Group/Group.hpp>
 #include <opm/input/eclipse/Schedule/Group/GuideRate.hpp>
 #include <opm/input/eclipse/Schedule/Schedule.hpp>
@@ -41,7 +39,6 @@
 
 #include <opm/simulators/flow/countGlobalCells.hpp>
 #include <opm/simulators/flow/FlowBaseVanguard.hpp>
-#include <opm/simulators/flow/SubDomain.hpp>
 
 #include <opm/simulators/linalg/matrixblock.hh>
 
@@ -80,6 +77,9 @@
 #include <vector>
 
 namespace Opm {
+
+template<class Scalar> class BlackoilWellModelNldd;
+template<class T> class SparseTable;
 
 #if COMPILE_GPU_BRIDGE
 template<class Scalar> class WellContributions;
@@ -131,8 +131,6 @@ template<class Scalar> class WellContributions;
             // For computing average pressured used by gpmaint
             using AverageRegionalPressureType = RegionAverageCalculator::
                 AverageRegionalPressure<FluidSystem, std::vector<int> >;
-
-            using Domain = SubDomain<Grid>;
 
             explicit BlackoilWellModel(Simulator& simulator);
 
@@ -248,11 +246,6 @@ template<class Scalar> class WellContributions;
             // Check if well equations is converged.
             ConvergenceReport getWellConvergence(const std::vector<Scalar>& B_avg, const bool checkWellGroupControls = false) const;
 
-            // Check if well equations are converged locally.
-            ConvergenceReport getDomainWellConvergence(const Domain& domain,
-                                                       const std::vector<Scalar>& B_avg,
-                                                       DeferredLogger& local_deferredLogger) const;
-
             const SimulatorReportSingle& lastReport() const;
 
             void addWellContributions(SparseMatrixAdapter& jacobian) const;
@@ -274,7 +267,6 @@ template<class Scalar> class WellContributions;
             // at the beginning of each time step (Not report step)
             void prepareTimeStep(DeferredLogger& deferred_logger);
             void initPrimaryVariablesEvaluation() const;
-            void initPrimaryVariablesEvaluationDomain(const Domain& domain) const;
 
             std::pair<bool, bool>
             updateWellControls(const bool mandatory_network_balance, DeferredLogger& deferred_logger, const bool relax_network_tolerance = false);
@@ -292,15 +284,23 @@ template<class Scalar> class WellContributions;
 
             using PressureMatrix = Dune::BCRSMatrix<Opm::MatrixBlock<Scalar, 1, 1>>;
 
-            void addWellPressureEquations(PressureMatrix& jacobian, const BVector& weights,const bool use_well_weights) const;
-
-            void addWellPressureEquationsDomain([[maybe_unused]] PressureMatrix& jacobian,
-                                                [[maybe_unused]] const BVector& weights,
-                                                [[maybe_unused]] const bool use_well_weights,
-                                                [[maybe_unused]] const int domainIndex) const;
-
-
+            void addWellPressureEquations(PressureMatrix& jacobian,
+                                          const BVector& weights,
+                                          const bool use_well_weights) const;
             void addWellPressureEquationsStruct(PressureMatrix& jacobian) const;
+            void addWellPressureEquationsDomain(PressureMatrix& jacobian,
+                                                const BVector& weights,
+                                                const bool use_well_weights,
+                                                const int domainIndex) const
+            {
+                if (!nldd_) {
+                    OPM_THROW(std::logic_error, "Attempt to access NLDD data without a NLDD solver");
[…]
