### 2025-01-07 897a7a6
adding two regression tests for autochoke functionality
