+++ regressionTests.cmake
@@ -131,6 +131,22 @@ add_test_compareECLFiles(CASENAME network_01_reroute_std
 		                 DIR network
 		                 TEST_ARGS --enable-tuning=true --local-well-solve-control-switching=true)
 
+add_test_compareECLFiles(CASENAME network_model5_stdw_autochk
+		                 FILENAME NETWORK_MODEL5_STDW_AUTOCHK
+		                 SIMULATOR flow
+		                 ABS_TOL ${abs_tol}
+		                 REL_TOL ${rel_tol}
+		                 DIR network
+		                 TEST_ARGS --enable-tuning=true)
+
+add_test_compareECLFiles(CASENAME network_model5_stdw_autochk_groupguiderates
+		                 FILENAME NETWORK_MODEL5_STDW_AUTOCHK_GROUPGUIDERATES
+		                 SIMULATOR flow
+		                 ABS_TOL ${abs_tol}
+		                 REL_TOL ${rel_tol}
+		                 DIR network
+		                 TEST_ARGS --enable-tuning=true)
+
 add_test_compareECLFiles(CASENAME gas_precsalt
                          FILENAME GASWATER_VAPWAT_PRECSALT
                          SIMULATOR flow
