+++ compareECLFiles.cmake
@@ -423,7 +423,7 @@ add_test_runSimulator(CASENAME tuning_tsinit_nextstep
                       POST_COMMAND $<TARGET_FILE:test_tuning_tsinit_nextstep>)
 
 if (opm-common_EMBEDDED_PYTHON)
-  include (${CMAKE_CURRENT_SOURCE_DIR}/pyactionComparisons.cmake)
+  include (${CMAKE_CURRENT_SOURCE_DIR}/pyactionActionXComparisons.cmake)
 endif ()
 include (${CMAKE_CURRENT_SOURCE_DIR}/regressionTests.cmake)
 include (${CMAKE_CURRENT_SOURCE_DIR}/restartTests.cmake)
