### 2025-01-08 486f7cc
remove buffer and view cpu support
