+++ opm/simulators/linalg/gpuistl/GpuView.hpp
@@ -151,13 +151,7 @@ public:
 #ifndef NDEBUG
         assertHasElements();
 #endif
-#if OPM_IS_INSIDE_DEVICE_FUNCTION
         return m_dataPtr[0];
-#else
-        T value;
-        cudaMemcpy(&value, &m_dataPtr[0], sizeof(T), cudaMemcpyDeviceToHost);
-        return value;
-#endif
     }
 
     /**
@@ -168,13 +162,7 @@ public:
 #ifndef NDEBUG
         assertHasElements();
 #endif
-#if OPM_IS_INSIDE_DEVICE_FUNCTION
         return m_dataPtr[m_numberOfElements-1];
-#else
-        T value;
-        cudaMemcpy(&value, &m_dataPtr[m_numberOfElements - 1], sizeof(T), cudaMemcpyDeviceToHost);
-        return value;
-#endif
     }
 
     /**
