### 2025-01-08 8a1ccac
changed: rename BlackoilWellModelGeneric::hasWell to hasLocalWell
