### 2025-01-08 9894900
changed: rename BlackoilWellModel::hasWell to hasOpenLocalWell
this to avoid confusion as there is already a hasWell
in BlackoilWellModelGeneric which checks all wells
