+++ opm/simulators/flow/TemperatureModel.hpp
@@ -34,7 +34,6 @@
 
 #include <opm/simulators/flow/GenericTemperatureModel.hpp>
 #include <opm/simulators/linalg/findOverlapRowsAndColumns.hpp>
-//#include <opm/simulators/utils/VectorVectorDataHandle.hpp>
 
 #include <array>
 #include <cstddef>
@@ -109,24 +108,25 @@ public:
     void init(bool rst)
     {
         const unsigned int numCells = simulator_.model().numTotalDof();
-
         this->doInit(rst, numCells);
 
         if (!this->doTemp())
             return;
 
-        storage0_.resize(numCells);
+        // we need the storage term at start of the iteration (timeIdx = 1)
+        storage1_.resize(numCells);
 
+        // set the initial temperature
         for (unsigned globI = 0; globI < numCells; ++globI) {
             this->temperature_[globI] = simulator_.problem().initialFluidState(globI).temperature(0);
         }
-     
+        // keep a copy of the intensive quantities to simplify the update during
+        // the newton iterations
         intQuants_.resize(numCells);
-        const auto& elemMapper = simulator_.model().elementMapper(); //does not work.
-        // Set it up manually
-        //ElementMapper elemMapper(simulator_.vanguard().gridView(), Dune::mcmgElementLayout());
-        detail::findOverlapAndInterior(simulator_.vanguard().grid(), elemMapper, overlapRows_, interiorRows_);
 
+        // find and store the overlap cells
+        const auto& elemMapper = simulator_.model().elementMapper();
+        detail::findOverlapAndInterior(simulator_.vanguard().grid(), elemMapper, overlapRows_, interiorRows_);
     }
 
     void beginTimeStep()
@@ -181,167 +181,42 @@ public:
     {
         serializer(static_cast<BaseType&>(*this));
         serializer(intQuants_);
-        serializer(storage0_);
+        serializer(storage1_);
     }
 
 protected:
 
     void updateStorageCache()
     {
+        // we need the storage term at start of the iteration (timeIdx = 1)
         const unsigned int numCells = simulator_.model().numTotalDof();
-        //#ifdef _OPENMP
-        //#pragma omp parallel for
-        //#endif
+        #ifdef _OPENMP
+        #pragma omp parallel for
+        #endif
         for (unsigned globI = 0; globI < numCells; ++globI) {
             Scalar storage = 0.0;
             computeStorageTerm(globI, storage);
-            storage0_[globI] = storage;
+            storage1_[globI] = storage;
         }
     }
 
     void advanceTemperatureFields()
     {
-
-
-        for (int iter = 0; iter < 20; ++iter) {
-        this->energyVector_ = 0.0;
-        (*this->energyMatrix_) = 0.0;
-
-         
-
-        Scalar dt = simulator_.timeStepSize();
         const unsigned int numCells = simulator_.model().numTotalDof();
-        //#ifdef _OPENMP
-        //#pragma omp parallel for
-        //#endif
-        for (unsigned globI = 0; globI < numCells; ++globI) {
-            Scalar volume = simulator_.model().dofTotalVolume(globI);
-            Scalar storefac = volume / dt;
-            Evaluation storage = 0.0;
-            computeStorageTerm(globI, storage); 
-            this->energyVector_[globI] += storefac * ( getValue(storage) - storage0_[globI] );
-            if (globI == 0) {
-                //std::cout << "storage " << getValue(storage) << " " << storage0_[globI] << " " << storefac * ( getValue(storage) - storage0_[globI] ) << " " <<  storefac * storage.derivative(Indices::temperatureIdx) << " " << dt/(3600*24) << std::endl;
+        const int maximum_number_of_newton_iterations = 20;
+        // solve using Newton
[…]
