### 2025-01-08 a7b83b8
some more cleanups
