+++ opm/simulators/flow/FlowProblem.hpp
@@ -868,11 +868,12 @@ public:
     Scalar temperature(const Context& context, unsigned spaceIdx, unsigned timeIdx) const
     {
         // use the initial temperature of the DOF if temperature is not a primary
-        // variable       
+        // variable
         unsigned globalDofIdx = context.globalSpaceIndex(spaceIdx, timeIdx);
+        if (enableTemperature)
+            return temperatureModel_.temperature(globalDofIdx);
 
-        return temperatureModel_.temperature(globalDofIdx);
-        //return asImp_().initialFluidState(globalDofIdx).temperature(/*phaseIdx=*/0);
+        return asImp_().initialFluidState(globalDofIdx).temperature(/*phaseIdx=*/0);
     }
 
 
@@ -880,8 +881,10 @@ public:
     {
         // use the initial temperature of the DOF if temperature is not a primary
         // variable
-        return temperatureModel_.temperature(globalDofIdx);
-        //return asImp_().initialFluidState(globalDofIdx).temperature(/*phaseIdx=*/0);
+        if (enableTemperature)
+            return temperatureModel_.temperature(globalDofIdx);
+
+        return asImp_().initialFluidState(globalDofIdx).temperature(/*phaseIdx=*/0);
     }
 
     const SolidEnergyLawParams&
@@ -1365,7 +1368,7 @@ protected:
 
     void readThermalParameters_()
     {
-        if constexpr (true)
+        if constexpr (enableTemperature || enableEnergy)
         {
             const auto& simulator = this->simulator();
             const auto& vanguard = simulator.vanguard();
