### 2025-01-08 73ede83
changed: move WBP BlackoilWellModel code to separate class
model using has-a instead of is-a in BlackoilWellModelGeneric
