+++ opm/simulators/wells/BlackoilWellModelGeneric.cpp
@@ -94,10 +94,10 @@ BlackoilWellModelGeneric(Schedule& schedule,
     , summaryState_(summaryState)
     , eclState_(eclState)
     , comm_(comm)
+    , wbp_(*this)
     , phase_usage_(phase_usage)
     , terminal_output_(comm_.rank() == 0 &&
                        Parameters::Get<Parameters::EnableTerminalOutput>())
-    , wbpCalculationService_ { eclState.gridDims(), comm_ }
     , guideRate_(schedule)
     , active_wgstate_(phase_usage)
     , last_valid_wgstate_(phase_usage)
@@ -2031,138 +2031,6 @@ setPrimaryVarsDomain(const int domainIdx, const std::vector<Scalar>& vars)
         }
     }
     assert(offset == vars.size());
-}
-
-template<class Scalar>
-void BlackoilWellModelGeneric<Scalar>::
-registerOpenWellsForWBPCalculation()
-{
-    assert (this->wbpCalcMap_.size() == this->wells_ecl_.size());
-
-    for (auto& wbpCalc : this->wbpCalcMap_) {
-        wbpCalc.openWellIdx_.reset();
-    }
-
-    auto openWellIdx = typename std::vector<WellInterfaceGeneric<Scalar>*>::size_type{0};
-    for (const auto* openWell : this->well_container_generic_) {
-        this->wbpCalcMap_[openWell->indexOfWell()].openWellIdx_ = openWellIdx++;
-    }
-}
-
-template<class Scalar>
-typename ParallelWBPCalculation<Scalar>::EvaluatorFactory
-BlackoilWellModelGeneric<Scalar>::
-makeWellSourceEvaluatorFactory(const std::vector<Well>::size_type wellIdx) const
-{
-    using Span = typename PAvgDynamicSourceData<Scalar>::template SourceDataSpan<Scalar>;
-    using Item = typename Span::Item;
-
-    return [wellIdx, this]() -> typename ParallelWBPCalculation<Scalar>::Evaluator
-    {
-        if (! this->wbpCalcMap_[wellIdx].openWellIdx_.has_value()) {
-            // Well is stopped/shut.  Return evaluator for stopped wells.
-            return []([[maybe_unused]] const int connIdx, Span sourceTerm)
-            {
-                // Well/connection is stopped/shut.  Set all items to
-                // zero.
-
-                sourceTerm
-                    .set(Item::Pressure      , 0.0)
-                    .set(Item::PoreVol       , 0.0)
-                    .set(Item::MixtureDensity, 0.0)
-                    .set(Item::Depth         , 0.0)
-                    ;
-            };
-        }
-
-        // Well is open.  Return an evaluator for open wells/open connections.
-        return [this, wellPtr = this->well_container_generic_[*this->wbpCalcMap_[wellIdx].openWellIdx_]]
-            (const int connIdx, Span sourceTerm)
-        {
-            // Note: The only item which actually matters for the WBP
-            // calculation at the well reservoir connection level is the
-            // mixture density.  Set other items to zero.
-
-            const auto& connIdxMap =
-                this->conn_idx_map_[wellPtr->indexOfWell()];
-
-            const auto rho = wellPtr->
-                connectionDensity(connIdxMap.global(connIdx),
-                                  connIdxMap.open(connIdx));
-
-            sourceTerm
-                .set(Item::Pressure      , 0.0)
-                .set(Item::PoreVol       , 0.0)
-                .set(Item::MixtureDensity, rho)
-                .set(Item::Depth         , 0.0)
-                ;
-        };
-    };
-}
-
-template<class Scalar>
-void BlackoilWellModelGeneric<Scalar>::
-initializeWBPCalculationService()
-{
-    this->wbpCalcMap_.clear();
[…]
