### 2025-01-08 c0a96b6
Added dummy water phase to compositional simulator.
Note: compositional is now always three phase!
