### 2025-01-08 a0f55d4
Added dummy water phase to compositional simulator.
Note: compositional is now always three phase!
