### 2025-01-08 df6bdbe
Added dummy water phase to compositional simulator.
Note: compositional is now always three phase!
