+++ opm/simulators/linalg/gpuistl/GpuBuffer.hpp
@@ -27,6 +27,7 @@
 #include <opm/simulators/linalg/gpuistl/GpuView.hpp>
 #include <vector>
 #include <string>
+#include <cuda_runtime.h>
 
 
 namespace Opm::gpuistl
@@ -219,6 +220,15 @@ public:
     void copyFromHost(const T* dataPointer, size_t numberOfElements);
 
     /**
+     * @brief copyFromHost copies numberOfElements from the CPU memory dataPointer
+     * @param dataPointer raw pointer to CPU memory
+     * @param numberOfElements number of elements to copy
+     * @note This does synchronous transfer.
+     * @note assumes that this buffer has numberOfElements elements
+     */
+    void copyFromHost(const T* dataPointer, size_t numberOfElements, cudaStream_t stream);
+
+    /**
      * @brief copyFromHost copies numberOfElements to the CPU memory dataPointer
      * @param dataPointer raw pointer to CPU memory
      * @param numberOfElements number of elements to copy
@@ -235,6 +245,15 @@ public:
      * @note This assumes that the size of this buffer is equal to the size of the input vector.
      */
     void copyFromHost(const std::vector<T>& data);
+
+    /**
+     * @brief copyToHost copies data from an std::vector
+     * @param data the vector to copy from
+     *
+     * @note This does synchronous transfer.
+     * @note This assumes that the size of this buffer is equal to the size of the input vector.
+     */
+    void copyFromHost(const std::vector<T>& data, cudaStream_t stream);
 
     /**
      * @brief copyToHost copies data to an std::vector
