### 2025-01-08 a1f5901
implement cudagraphs for gpudilu update
for some reason I am not able to measure an any
speedup in the update.... it should see the same
absolute speedup as the apply
