### 2025-01-08 b865e80
Set limit to the number of threads if it is unset when we ask for the number.
