+++ opm/simulators/linalg/gpuistl/GpuBuffer.hpp
@@ -120,50 +120,6 @@ public:
     const T* data() const;
 
     /**
-     * @return fetch the first element in a GpuBuffer
-     */
-    __host__ __device__ T& front()
-    {
-#ifndef NDEBUG
-        assertHasElements();
-#endif
-        return m_dataOnDevice[0];
-    }
-
-    /**
-     * @return fetch the last element in a GpuBuffer
-     */
-    __host__ __device__ T& back()
-    {
-#ifndef NDEBUG
-        assertHasElements();
-#endif
-        return m_dataOnDevice[m_numberOfElements-1];
-    }
-
-    /**
-     * @return fetch the first element in a GpuBuffer
-     */
-    __host__ __device__ T front() const
-    {
-#ifndef NDEBUG
-        assertHasElements();
-#endif
-        return m_dataOnDevice[0];
-    }
-
-    /**
-     * @return fetch the last element in a GpuBuffer
-     */
-    __host__ __device__ T back() const
-    {
-#ifndef NDEBUG
-        assertHasElements();
-#endif
-        return m_dataOnDevice[m_numberOfElements-1];
-    }
-
-    /**
      * @brief copyFromHost copies data from a Dune::BlockVector
      * @param bvector the vector to copy from
      *
