### 2025-01-08 6c29cf6
use less template arguments
