### 2025-01-08 90bde05
some cleanup from review
