### 2025-01-09 f267df9
Avoid repeated rewriting.
