### 2025-01-09 64f9ebb
move hasOpenLocalWell to BlackoilWellModelGeneric
