+++ opm/simulators/wells/BlackoilWellModelGeneric.hpp
@@ -103,7 +103,12 @@ public:
 
     /// return true if wells are available in the reservoir
     bool wellsActive() const;
+
+    //! \brief Returns true if well is defined and has connections on current rank.
     bool hasWell(const std::string& wname) const;
+
+    //! \brief Returns true if well is defined, open and has connections on current rank.
+    bool hasOpenLocalWell(const std::string& well_name) const;
 
     /// return true if network is active (at least one network well in prediction mode)
     bool networkActive() const;
