+++ opm/simulators/wells/BlackoilWellModel_impl.hpp
@@ -2378,21 +2378,6 @@ namespace Opm {
         return *well;
     }
 
-    template<typename TypeTag>
-    bool
-    BlackoilWellModel<TypeTag>::
-    hasOpenLocalWell(const std::string& well_name) const
-    {
-        return std::any_of(well_container_.begin(), well_container_.end(),
-            [&well_name](const WellInterfacePtr& elem) -> bool
-        {
-            return elem->name() == well_name;
-        });
-    }
-
-
-
-
     template <typename TypeTag>
     int
     BlackoilWellModel<TypeTag>::
