+++ opm/simulators/flow/FlowProblem.hpp
@@ -1202,7 +1202,6 @@ public:
         serializer(wellModel_);
         serializer(aquiferModel_);
         serializer(tracerModel_);
-        serializer(temperatureModel_);
         serializer(*materialLawManager_);
     }
 
