### 2025-01-10 70e993c
BlackoilWellModel: move group switching reporting to separate method
this can be put in the generic TU
