### 2025-01-10 f59dd40
BlackoilWellModelGeneric::reportGroupSwitching: use {fmt}
