### 2025-01-10 6d003b2
BlackoilWellModel::computeWellTemperature: use if constexpr
