### 2025-01-10 df5c12f
Update workflow: Remove checking out the repository, that is actually not needed
