+++ pyactionActionXComparisons.cmake
@@ -196,6 +196,17 @@ add_test_compareSeparateECLFiles(CASENAME pyaction_WTMULT_insert_kw
                                  IGNORE_EXTRA_KW BOTH)
 
 if(MPI_FOUND)
+  add_test_compareSeparateECLFiles(CASENAME pyaction_gconprod_insert_kw_4_procs
+                                   DIR1 pyaction
+                                   FILENAME1 PYACTION_GCONPROD_INSERT_KW
+                                   DIR2 actionx
+                                   FILENAME2 ACTIONX_GCONPROD
+                                   SIMULATOR flow
+                                   ABS_TOL ${abs_tol}
+                                   REL_TOL ${rel_tol}
+                                   IGNORE_EXTRA_KW BOTH
+                                   MPI_PROCS 4)
+
   add_test_compareSeparateECLFiles(CASENAME pyaction_gconsump_insert_kw_4_procs
                                    DIR1 pyaction
                                    FILENAME1 PYACTION_GCONSUMP_INSERT_KW
@@ -207,6 +218,16 @@ if(MPI_FOUND)
                                    IGNORE_EXTRA_KW BOTH
                                    MPI_PROCS 4)
 
+  add_test_compareSeparateECLFiles(CASENAME pyaction_gruptree_insert_kw_4_procs
+                                   DIR1 pyaction
+                                   FILENAME1 PYACTION_GRUPTREE_INSERT_KW
+                                   DIR2 actionx
+                                   FILENAME2 ACTIONX_GRUPTREE
+                                   SIMULATOR flow
+                                   ABS_TOL ${abs_tol}
+                                   REL_TOL ${rel_tol}
+                                   IGNORE_EXTRA_KW BOTH
+                                   MPI_PROCS 4)
 
   add_test_compareSeparateECLFiles(CASENAME actionx_gefac_4_procs
                                    DIR1 model4
@@ -268,6 +289,17 @@ if(MPI_FOUND)
                                    FILENAME1 PYACTION_NEXT_INSERT_KW
                                    DIR2 actionx
                                    FILENAME2 ACTIONX_NEXT
+                                   SIMULATOR flow
+                                   ABS_TOL ${abs_tol}
+                                   REL_TOL ${rel_tol}
+                                   IGNORE_EXTRA_KW BOTH
+                                   MPI_PROCS 4)
+
+  add_test_compareSeparateECLFiles(CASENAME pyaction_wconprod_insert_kw_4_procs
+                                   DIR1 pyaction
+                                   FILENAME1 PYACTION_WCONPROD_INSERT_KW
+                                   DIR2 actionx
+                                   FILENAME2 ACTIONX_WCONPROD
                                    SIMULATOR flow
                                    ABS_TOL ${abs_tol}
                                    REL_TOL ${rel_tol}
