### 2025-01-10 5549670
Do not try to restore segment info for inactive wells
