### 2025-01-13 da892eb
fix another issue with heatflux and remove some unused variables
