### 2025-01-13 b3520dd
fix another issue with heatflux and remove some unused variables
