### 2025-01-13 fa02d68
refactor move_to_gpu
