+++ opm/simulators/flow/BlackoilModel.hpp
@@ -24,1318 +24,334 @@
 #ifndef OPM_BLACKOILMODEL_HEADER_INCLUDED
 #define OPM_BLACKOILMODEL_HEADER_INCLUDED
 
-#include <dune/common/timer.hh>
-
-#include <opm/common/ErrorMacros.hpp>
-#include <opm/common/Exceptions.hpp>
-#include <opm/common/OpmLog/OpmLog.hpp>
-
-#include <opm/input/eclipse/EclipseState/EclipseState.hpp>
-#include <opm/input/eclipse/EclipseState/Tables/TableManager.hpp>
-
-#include <opm/simulators/aquifers/AquiferGridUtils.hpp>
 #include <opm/simulators/aquifers/BlackoilAquiferModel.hpp>
 
 #include <opm/simulators/flow/BlackoilModelConvergenceMonitor.hpp>
 #include <opm/simulators/flow/BlackoilModelNldd.hpp>
-#include <opm/simulators/flow/BlackoilModelParameters.hpp>
 #include <opm/simulators/flow/BlackoilModelProperties.hpp>
-#include <opm/simulators/flow/countGlobalCells.hpp>
-#include <opm/simulators/flow/FlowProblemBlackoil.hpp>
-#include <opm/simulators/flow/NonlinearSolver.hpp>
+#include <opm/simulators/flow/FlowProblemBlackoilProperties.hpp>
 #include <opm/simulators/flow/RSTConv.hpp>
 
-#include <opm/simulators/timestepping/AdaptiveTimeStepping.hpp>
+#include <opm/simulators/linalg/ISTLSolver.hpp>
+
 #include <opm/simulators/timestepping/ConvergenceReport.hpp>
 #include <opm/simulators/timestepping/SimulatorReport.hpp>
 #include <opm/simulators/timestepping/SimulatorTimer.hpp>
 
+#include <opm/simulators/utils/BlackoilPhases.hpp>
 #include <opm/simulators/utils/ComponentName.hpp>
-#include <opm/simulators/utils/DeferredLoggingErrorHelpers.hpp>
-#include <opm/simulators/utils/ParallelCommunication.hpp>
-#include <opm/simulators/utils/phaseUsageFromDeck.hpp>
 
 #include <opm/simulators/wells/BlackoilWellModel.hpp>
 
-#include <algorithm>
-#include <cassert>
-#include <cmath>
-#include <filesystem>
-#include <fstream>
-#include <functional>
-#include <iomanip>
-#include <ios>
-#include <limits>
 #include <memory>
-#include <numeric>
-#include <sstream>
 #include <tuple>
-#include <utility>
 #include <vector>
 
 #include <fmt/format.h>
 
 namespace Opm {
 
-    /// A model implementation for three-phase black oil.
-    ///
-    /// The simulator is capable of handling three-phase problems
-    /// where gas can be dissolved in oil and vice versa. It
-    /// uses an industry-standard TPFA discretization with per-phase
-    /// upwind weighting of mobilities.
-    template <class TypeTag>
-    class BlackoilModel
-    {
-    public:
-        // ---------  Types and enums  ---------
-        using Simulator = GetPropType<TypeTag, Properties::Simulator>;
-        using Grid = GetPropType<TypeTag, Properties::Grid>;
-        using ElementContext = GetPropType<TypeTag, Properties::ElementContext>;
-        using IntensiveQuantities = GetPropType<TypeTag, Properties::IntensiveQuantities>;
-        using SparseMatrixAdapter = GetPropType<TypeTag, Properties::SparseMatrixAdapter>;
-        using SolutionVector = GetPropType<TypeTag, Properties::SolutionVector>;
[…]
