### 2025-01-13 c909283
BlackoilModel: move properties to separate header
