### 2025-01-13 0b57d01
fixed: don't run parallel pyaction tests without mpi
