+++ pyactionComparisons.cmake
@@ -25,17 +25,6 @@ add_test_compareSeparateECLFiles(CASENAME pyaction_gconsump_insert_kw
                                  REL_TOL ${rel_tol}
                                  IGNORE_EXTRA_KW BOTH)
 
-add_test_compareSeparateECLFiles(CASENAME pyaction_gconsump_insert_kw_4_procs
-                                 DIR1 pyaction
-                                 FILENAME1 PYACTION_GCONSUMP_INSERT_KW
-                                 DIR2 actionx
-                                 FILENAME2 ACTIONX_GCONSUMP
-                                 SIMULATOR flow
-                                 ABS_TOL ${abs_tol}
-                                 REL_TOL ${rel_tol}
-                                 IGNORE_EXTRA_KW BOTH
-                                 MPI_PROCS 4)
-
 add_test_compareSeparateECLFiles(CASENAME pyaction_gruptree_insert_kw
                                  DIR1 pyaction
                                  FILENAME1 PYACTION_GRUPTREE_INSERT_KW
@@ -56,17 +45,6 @@ add_test_compareSeparateECLFiles(CASENAME pyaction_mult+_insert_kw
                                  REL_TOL ${rel_tol}
                                  IGNORE_EXTRA_KW BOTH)
 
-add_test_compareSeparateECLFiles(CASENAME pyaction_mult+_insert_kw_4_procs
-                                 DIR1 pyaction
-                                 FILENAME1 PYACTION_MULT+_INSERT_KW
-                                 DIR2 actionx
-                                 FILENAME2 ACTIONX_MULT+
-                                 SIMULATOR flow
-                                 ABS_TOL ${abs_tol}
-                                 REL_TOL ${rel_tol}
-                                 IGNORE_EXTRA_KW BOTH
-                                 MPI_PROCS 4)
-
 add_test_compareSeparateECLFiles(CASENAME pyaction_multx+_insert_kw
                                  DIR1 pyaction
                                  FILENAME1 PYACTION_MULTX+_INSERT_KW
@@ -76,17 +54,6 @@ add_test_compareSeparateECLFiles(CASENAME pyaction_multx+_insert_kw
                                  ABS_TOL ${abs_tol}
                                  REL_TOL ${rel_tol}
                                  IGNORE_EXTRA_KW BOTH)
-
-add_test_compareSeparateECLFiles(CASENAME pyaction_multx+_insert_kw_4_procs
-                                 DIR1 pyaction
-                                 FILENAME1 PYACTION_MULTX+_INSERT_KW
-                                 DIR2 actionx
-                                 FILENAME2 ACTIONX_MULTX+
-                                 SIMULATOR flow
-                                 ABS_TOL ${abs_tol}
-                                 REL_TOL ${rel_tol}
-                                 IGNORE_EXTRA_KW BOTH
-                                 MPI_PROCS 4)
 
 add_test_compareSeparateECLFiles(CASENAME pyaction_multx-_insert_kw
                                  DIR1 pyaction
@@ -98,17 +65,6 @@ add_test_compareSeparateECLFiles(CASENAME pyaction_multx-_insert_kw
                                  REL_TOL ${rel_tol}
                                  IGNORE_EXTRA_KW BOTH)
 
-add_test_compareSeparateECLFiles(CASENAME pyaction_multx-_insert_kw_4_procs
-                                 DIR1 pyaction
-                                 FILENAME1 PYACTION_MULTX-_INSERT_KW
-                                 DIR2 actionx
-                                 FILENAME2 ACTIONX_MULTX-
-                                 SIMULATOR flow
-                                 ABS_TOL ${abs_tol}
-                                 REL_TOL ${rel_tol}
-                                 IGNORE_EXTRA_KW BOTH
-                                 MPI_PROCS 4)
-
[…]
