### 2025-01-13 495831c
Separate keyword validation for fully supported keywords
