### 2025-01-13 7bf47e6
regression test for gas water and multisegment wells
