### 2025-01-13 7590f01
BlackoilModel: make wasSwitched_ private
