### 2025-01-13 d7d83c4
Avoid setWsolvent call for wells shut by WCYCLE (assertion failure, at best)
