### 2025-01-13 653cdfc
BlackoilModel: move to has-a instead of is-a modelling of penalty cards
