### 2025-01-14 6e50ec9
Add test case for maybeChopSubStep()
