### 2025-01-14 ee73a49
Add test case for maybeChopSubStep()
