+++ tests/rescoup/test_chopstep.cpp
@@ -1,0 +1,287 @@
+// -*- mode: C++; tab-width: 4; indent-tabs-mode: nil; c-basic-offset: 4 -*-
+// vi: set et ts=4 sw=4 sts=4:
+/*
+  This file is part of the Open Porous Media project (OPM).
+
+  OPM is free software: you can redistribute it and/or modify
+  it under the terms of the GNU General Public License as published by
+  the Free Software Foundation, either version 3 of the License, or
+  (at your option) any later version.
+
+  OPM is distributed in the hope that it will be useful,
+  but WITHOUT ANY WARRANTY; without even the implied warranty of
+  MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
+  GNU General Public License for more details.
+
+  You should have received a copy of the GNU General Public License
+  along with OPM.  If not, see <http://www.gnu.org/licenses/>.
+
+  Consult the COPYING file in the top-level source directory of this
+  module for the precise wording of the license and the list of
+  copyright holders.
+*/
+#include "config.h"
+
+#define BOOST_TEST_MODULE ResCoup_ChopStep
+#include <opm/models/utils/propertysystem.hh>
+#include <opm/models/utils/parametersystem.hpp>
+#include <opm/models/utils/start.hh>
+
+#include <opm/input/eclipse/EclipseState/SummaryConfig/SummaryConfig.hpp>
+#include <opm/input/eclipse/Schedule/Schedule.hpp>
+#include <opm/input/eclipse/Schedule/ResCoup/ReservoirCouplingInfo.hpp>
+#include <opm/input/eclipse/Schedule/ResCoup/MasterGroup.hpp>
+#include <opm/input/eclipse/Schedule/ResCoup/Slaves.hpp>
+#include <opm/simulators/flow/FlowGenericVanguard.hpp>
+#include <opm/simulators/flow/Main.hpp>
+#include <opm/simulators/flow/TTagFlowProblemTPFA.hpp>
+#include <opm/simulators/flow/ReservoirCoupling.hpp>
+#include <opm/simulators/flow/ReservoirCouplingMaster.hpp>
+#include <opm/simulators/flow/ReservoirCouplingSlave.hpp>
+#include <opm/simulators/utils/DeferredLogger.hpp>
+#include <opm/simulators/flow/BlackoilModel.hpp>
+#include <opm/simulators/flow/FlowProblemBlackoil.hpp>
+
+#if HAVE_DUNE_FEM
+#include <dune/fem/misc/mpimanager.hh>
+#else
+#include <dune/common/parallel/mpihelper.hh>
+#endif
+
+#include <boost/test/unit_test.hpp>
+#include <boost/version.hpp>
+#if BOOST_VERSION / 100000 == 1 && BOOST_VERSION / 100 % 1000 < 71
+#include <boost/test/floating_point_comparison.hpp>
+#else
+#include <boost/test/tools/floating_point_comparison.hpp>
+#endif
+
+namespace Opm {
+
+class MainTestWrapper : public Main
+{
+public:
+    using TypeTag = Properties::TTag::FlowProblemTPFA;
+    using FlowMainPtr = std::unique_ptr<FlowMain<TypeTag>>;
+    using Simulator = GetPropType<TypeTag, Properties::Simulator>;
+
+    MainTestWrapper(int argc, char** argv) :
+    // NOTE: passing ownMPI=false to prevent each test from initializing MPI (which is illegal)
+          Main{argc, argv, /*ownMPI=*/false}
+        , exit_code_{EXIT_SUCCESS}
+        , flow_main_{nullptr}
+        , simulator_{nullptr}
+    {
[…]
