### 2025-01-14 79f5a1a
NonlinearSolver::Parameters: move to separate struct
and move some more code to the TU. small tidy up enabled
by typetag-less parameters system
