### 2025-01-14 cee75a1
FlowBaseVanguard: move some more code to FlowGenericVanguard
typetag less parameter system now allows us to initialize these variables
in class that holds them.
