### 2025-01-14 5122631
OutputBlackoilModule: move some more code to GenericOutputBlackoilModule
typetag less parameter system now allows us to initialize these variables
in class that holds them.
