### 2025-01-14 ef176d4
Damaris: move some more code to TU
possible with typetag less parameter system
