### 2025-01-14 eb782d9
Set maximum threads immediately after parsing paramteres. Remove hotfix.
