### 2025-01-14 3144de5
extend is_a_blackoil_system check
