### 2025-01-14 0995af2
WIP - currently broken, trying to avoid void*
