+++ flowexperimental/BlackOilIntensiveQuantitiesGlobalIndex.hpp
@@ -154,8 +154,8 @@ public:
     /*!
      * \copydoc IntensiveQuantities::update
      */
-    template <class DynamicFluidSystem = void*>
-    void update(const ElementContext& elemCtx, unsigned dofIdx, unsigned timeIdx, const DynamicFluidSystem& fluidSystem = nullptr)
+    template <class DynamicFluidSystem = FluidSystem>
+    void update(const ElementContext& elemCtx, unsigned dofIdx, unsigned timeIdx, const DynamicFluidSystem& fluidSystem = DynamicFluidSystem{})
     {
         ParentType::update(elemCtx, dofIdx, timeIdx);
 
