### 2025-01-14 3d3333d
refactor blackoilintensivequantities update
