+++ tests/gpuistl/test_gpuBlackOilFluidSystem.cu
@@ -1,728 +1,0 @@
-// -*- mode: C++; tab-width: 4; indent-tabs-mode: nil; c-basic-offset: 4 -*-
-// vi: set et ts=4 sw=4 sts=4:
-/*
-  This file is part of the Open Porous Media project (OPM).
-
-  OPM is free software: you can redistribute it and/or modify
-  it under the terms of the GNU General Public License as published by
-  the Free Software Foundation, either version 2 of the License, or
-  (at your option) any later version.
-
-  OPM is distributed in the hope that it will be useful,
-  but WITHOUT ANY WARRANTY; without even the implied warranty of
-  MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
-  GNU General Public License for more details.
-
-  You should have received a copy of the GNU General Public License
-  along with OPM.  If not, see <http://www.gnu.org/licenses/>.
-
-  Consult the COPYING file in the top-level source directory of this
-  module for the precise wording of the license and the list of
-  copyright holders.
-*/
-/*!
- * \file
- *
- * \brief This is the unit test for the black oil fluid system
- *
- * This test requires the presence of opm-parser.
- */
-#include "config.h"
-
-#if !HAVE_ECL_INPUT
-#error "The test for the black oil fluid system classes requires ecl input support in opm-common"
-#endif
-
-#include <boost/mpl/list.hpp>
-
-#define BOOST_TEST_MODULE EclBlackOilFluidSystem
-#include <boost/test/unit_test.hpp>
-
-#include <opm/material/fluidsystems/BlackOilFluidSystem.hpp>
-#include <opm/material/fluidstates/BlackOilFluidState.hpp>
-#include <opm/material/densead/Evaluation.hpp>
-
-#include <opm/material/common/Valgrind.hpp>
-
-#include <opm/input/eclipse/Parser/Parser.hpp>
-#include <opm/input/eclipse/Deck/Deck.hpp>
-#include <opm/input/eclipse/EclipseState/EclipseState.hpp>
-#include <opm/input/eclipse/Python/Python.hpp>
-#include <opm/input/eclipse/Schedule/Schedule.hpp>
-
-#include <type_traits>
-#include <cmath>
-
-// values of strings based on the SPE1 and NORNE cases of opm-data.
-static constexpr const char* deckString1 =
-    "RUNSPEC\n"
-    "\n"
-    "DIMENS\n"
-    "   10 10 3 /\n"
-    "\n"
-    "METRIC\n"
-    "\n"
-    "TABDIMS\n"
-    " * 2 /\n"
-    "\n"
-    "OIL\n"
-    "GAS\n"
-    "WATER\n"
-    "\n"
-    "DISGAS\n"
-    "VAPOIL\n"
-    "\n"
-    "GRID\n"
-    "\n"
-    "DX\n"
-    "   	300*1000 /\n"
-    "DY\n"
-    "	300*1000 /\n"
-    "DZ\n"
-    "	100*20 100*30 100*50 /\n"
-    "\n"
-    "TOPS\n"
-    "	100*1234 /\n"
-    "\n"
-    "PORO\n"
-    "  300*0.15 /\n"
-    "PROPS\n"
-    "\n"
-    "-- Copyright (C) 2015 Statoil\n"
-    "\n"
-    "PVTG\n"
-    "\n"
-    "-- PRESSURE       RSG        B-GAS     VISCOSITY\n"
-    "--   BAR                                 (CP)\n"
-    "\n"
-    "     50.00    0.00000497   0.024958     0.01441\n"
-    "              0.00000248   0.024958     0.01440\n"
-    "              0.00000000   0.024958     0.01440 /\n"
-    "\n"
[…]
