### 2025-01-14 5ba8e87
Uncomment OpmLog warning
