### 2025-01-15 c3489c5
Make 2 threads default and fix -1 for automatic
