### 2025-01-15 76667d6
Introduce struct to reduce number of constructor params
