### 2025-01-15 cd7da15
Add parameter Conserve inner energy and use it
