### 2025-01-15 47c4486
Remove false-positive warning with OMP_NUM_THREADS.
