### 2025-01-15 66b353d
Remove one extraneous registration of ThreadsPerProcess parameter
