### 2025-01-16 166ded0
Add tests for PartitionCells interface with simple and Zoltan
