+++ tests/test_partitionCells.cpp
@@ -1,5 +1,6 @@
 /*
-  Copyright 2021 SINTEF Digital, Mathematics and Cybernetics.
+  Copyright 2021,2025 SINTEF Digital, Mathematics and Cybernetics.
+  Copyright 2025 Equinor ASA
 
   This file is part of the Open Porous Media project (OPM).
 
@@ -19,11 +20,15 @@
 
 #include <config.h>
 
-#define BOOST_TEST_MODULE OPM_test_aspinPartition
+#define BOOST_TEST_MODULE OPM_test_partitionCells
+#define BOOST_TEST_NO_MAIN
 #include <boost/test/unit_test.hpp>
 
 #include <opm/simulators/flow/partitionCells.hpp>
+#include <dune/grid/common/gridview.hh>
+#include <opm/grid/CpGrid.hpp>
 
+#include <opm/input/eclipse/Schedule/Well/Well.hpp>
 
 BOOST_AUTO_TEST_CASE(FileBased)
 {
@@ -54,6 +59,95 @@ BOOST_AUTO_TEST_CASE(Simple2)
     BOOST_CHECK_EQUAL(num_part, 7);
     std::vector<int> expected = { 0, 0, 1, 1, 2, 2, 3, 4, 5, 6 };
     BOOST_CHECK_EQUAL_COLLECTIONS(expected.begin(), expected.end(), part.begin(), part.end());
-
 }
 
+BOOST_AUTO_TEST_CASE(PartitionCellsTestSimple)
+{
+    // Create a 3x2x2 grid (12 cells total)
+    Dune::CpGrid grid;
+    std::array<int, 3> dims{{3, 4, 1}};
+    std::array<double, 3> size{{3.0, 4.0, 1.0}};
+    grid.createCartesian(dims, size);
+
+    // Setup Zoltan control parameters
+    using Entity = typename Dune::CpGrid::LeafGridView::template Codim<0>::Entity;
+    Opm::ZoltanPartitioningControl<Entity> zoltan_ctrl;
+    zoltan_ctrl.domain_imbalance = 1.1;  // Allow 10% imbalance
+
+    // Initialize index function to return entity index
+    std::function<int(const Entity&)> index_func = [](const Entity& e) { return e.index(); };
+    zoltan_ctrl.index = index_func;
+
+    // Initialize local_to_global function to use index as global ID for this test
+    std::function<int(int)> l2g_func = [](int idx) { return idx; };
+    zoltan_ctrl.local_to_global = l2g_func;
+
+    // Test simple partitioning first
+    {
+        auto [part, num_part] = Opm::partitionCells("simple", 3, grid.leafGridView(),
+                                                   std::vector<Opm::Well>{},
+                                                   std::unordered_map<std::string, std::set<int>>{},
+                                                   zoltan_ctrl);
+
+        BOOST_CHECK_EQUAL(num_part, 3);
+        BOOST_CHECK_EQUAL(part.size(), 12);
+
+        // Check that all partition numbers are valid
+        for (const auto& p : part) {
+            BOOST_CHECK(p >= 0 && p < 3);
+        }
+    }
+}
+
+BOOST_AUTO_TEST_CASE(PartitionCellsTestZoltan)
+{
+    // Create a 3x2x2 grid (12 cells total)
+    Dune::CpGrid grid;
+    std::array<int, 3> dims{{3, 4, 1}};
+    std::array<double, 3> size{{3.0, 4.0, 1.0}};
+    grid.createCartesian(dims, size);
+
+    // Setup Zoltan control parameters
+    using Entity = typename Dune::CpGrid::LeafGridView::template Codim<0>::Entity;
[…]
