### 2025-01-16 0f1c2f3
fixing compilation related to conversion from std::optional
