### 2025-01-16 9a97bf4
Transmissibility: add a cache for centroids and use this in update()
The loops repeatedly calls distanceVector which recalculates
the centroid for a cell. This is not a trivial calculation
