+++ opm/simulators/flow/Transmissibility_impl.hpp
@@ -177,10 +177,12 @@ update(bool global, const TransUpdateQuantities update_quantities,
 
     const bool disableNNC = eclState_.getSimulationConfig().useNONNC();
 
-    if (map)
+    if (map) {
         extractPermeability_(map);
-    else
+    }
+    else {
         extractPermeability_();
+    }
 
     // reserving some space in the hashmap upfront saves quite a bit of time because
     // resizes are costly for hashmaps and there would be quite a few of them if we
@@ -225,8 +227,7 @@ update(bool global, const TransUpdateQuantities update_quantities,
         auto pinchTransCalcMode = eclGrid.getPinchOption();
         useSmallestMultiplier = eclGrid.getMultzOption() == PinchMode::ALL;
         pinchOption4ALL = (pinchTransCalcMode == PinchMode::ALL);
-        if (pinchOption4ALL)
-        {
+        if (pinchOption4ALL) {
             useSmallestMultiplier = false;
         }
     }
@@ -291,14 +292,14 @@ update(bool global, const TransUpdateQuantities update_quantities,
                         transBoundaryEnergyIs;
                 }
 
-                ++ boundaryIsIdx;
+                ++boundaryIsIdx;
                 continue;
             }
 
             if (!intersection.neighbor()) {
                 // elements can be on process boundaries, i.e. they are not on the
                 // domain boundary yet they don't have neighbors.
-                ++ boundaryIsIdx;
+                ++boundaryIsIdx;
                 continue;
             }
 
@@ -316,8 +317,9 @@ update(bool global, const TransUpdateQuantities update_quantities,
             // ghost cells and we need to use the cartesian index as this will be used when applying Z multipliers
             // To cover the case where both cells are part of an LGR and as a consequence might have
             // the same cartesian index, we tie their Cartesian indices and the ones on the leaf grid view.
-            if (std::tie(insideCartElemIdx, elemIdx) > std::tie(outsideCartElemIdx, outsideElemIdx))
+            if (std::tie(insideCartElemIdx, elemIdx) > std::tie(outsideCartElemIdx, outsideElemIdx)) {
                 continue;
+            }
 
             // local indices of the faces of the inside and
             // outside elements which contain the intersection
@@ -329,7 +331,7 @@ update(bool global, const TransUpdateQuantities update_quantities,
                 // *added to* by applyNncToGridTrans_() later.
                 assert(outsideFaceIdx == -1);
                 trans_[details::isId(elemIdx, outsideElemIdx)] = 0.0;
-                if (enableEnergy_  && !onlyTrans){
+                if (enableEnergy_ && !onlyTrans) {
                     thermalHalfTrans_[details::directionalIsId(elemIdx, outsideElemIdx)] = 0.0;
                     thermalHalfTrans_[details::directionalIsId(outsideElemIdx, elemIdx)] = 0.0;
                 }
@@ -380,42 +382,42 @@ update(bool global, const TransUpdateQuantities update_quantities,
             // convert half transmissibilities to full face
             // transmissibilities using the harmonic mean
             Scalar trans;
-            if (std::abs(halfTrans1) < 1e-30 || std::abs(halfTrans2) < 1e-30)
+            if (std::abs(halfTrans1) < 1e-30 || std::abs(halfTrans2) < 1e-30) {
                 // avoid division by zero
                 trans = 0.0;
-            else
-                trans = 1.0 / (1.0/halfTrans1 + 1.0/halfTrans2);
+            }
+            else {
[…]
