### 2025-01-16 3695ea8
Transmissibility::update(): use Dune::intersections() range generator
