### 2025-01-16 49a15a8
Transmissibility::update(): use Dune::intersections() range generator
