### 2025-01-16 23e1654
Transmissibility::update: use ThreadSafeMapBuilder for thermalHalfTrans
this in preparation for multithreading the loop
