### 2025-01-16 3401108
Transmissibility::update: use ThreadedMap for thermalHalfTrans
this in preparation for multithreading the loop
