### 2025-01-16 4a847d2
Transmissibility::update: use ThreadedMap for trans_
this in preparation for multithreading the loop
