### 2025-01-16 4db435c
Transmissibility::update: use ThreadedMap for diffusivity_
this in preparation for multithreading the loop
