### 2025-01-16 5cbbde6
Transmissibility::update: use ThreadSafeMapBuilder for thermalHalfTransBoundary
this in preparation for multithreading the loop
