### 2025-01-16 a502a86
Transmissibility::update: use ThreadSafeMapBuilder for diffusivity_
this in preparation for multithreading the loop
