### 2025-01-16 da696ff
Transmissibility::update: use ThreadSafeMapBuilder for trans_
this in preparation for multithreading the loop
