### 2025-01-16 45850c7
Transmissibility::update: use ThreadedMap for thermalHalfTransBoundary
this in preparation for multithreading the loop
