### 2025-01-16 c06d89b
Use the createThreadIterators() utility.
