### 2025-01-16 eb3cfa6
Get globalCell from level zero grid
