### 2025-01-16 8772f41
Transmissibility::update(): enable multithreading
