### 2025-01-16 5e65c43
Transmissibility::update(): enable multithreading
