+++ opm/simulators/flow/Transmissibility_impl.hpp
@@ -281,295 +281,295 @@ update(bool global, const TransUpdateQuantities update_quantities,
 #pragma omp parallel for
 #endif
     for (int chunk = 0; chunk < num_chunks; ++chunk) {
-    for (auto it = grid_chunk_iterators[chunk]; it != grid_chunk_iterators[chunk+1]; ++it) {
-        const auto& elem = *it;
-        unsigned elemIdx = elemMapper.index(elem);
+        for (auto it = grid_chunk_iterators[chunk]; it != grid_chunk_iterators[chunk+1]; ++it) {
+            const auto& elem = *it;
+            unsigned elemIdx = elemMapper.index(elem);
 
-        unsigned boundaryIsIdx = 0;
-        for (const auto& intersection : Dune::intersections(gridView_, elem)) {
-            // deal with grid boundaries
-            if (intersection.boundary()) {
-                // compute the transmissibilty for the boundary intersection
-                const auto& geometry = intersection.geometry();
-                const auto& faceCenterInside = geometry.center();
+            unsigned boundaryIsIdx = 0;
+            for (const auto& intersection : Dune::intersections(gridView_, elem)) {
+                // deal with grid boundaries
+                if (intersection.boundary()) {
+                    // compute the transmissibilty for the boundary intersection
+                    const auto& geometry = intersection.geometry();
+                    const auto& faceCenterInside = geometry.center();
 
-                auto faceAreaNormal = intersection.centerUnitOuterNormal();
-                faceAreaNormal *= geometry.volume();
+                    auto faceAreaNormal = intersection.centerUnitOuterNormal();
+                    faceAreaNormal *= geometry.volume();
 
-                Scalar transBoundaryIs;
-                computeHalfTrans_(transBoundaryIs,
+                    Scalar transBoundaryIs;
+                    computeHalfTrans_(transBoundaryIs,
+                                      faceAreaNormal,
+                                      intersection.indexInInside(),
+                                      distanceVector_(faceCenterInside,
+                                                      elemIdx),
+                                      permeability_[elemIdx]);
+
+                    // normally there would be two half-transmissibilities that would be
+                    // averaged. on the grid boundary there only is the half
+                    // transmissibility of the interior element.
+                    unsigned insideCartElemIdx = cartMapper_.cartesianIndex(elemIdx);
+                    applyMultipliers_(transBoundaryIs, intersection.indexInInside(), insideCartElemIdx, transMult);
+                    transBoundary.emplace(std::make_pair(elemIdx, boundaryIsIdx), transBoundaryIs);
+
+                    // for boundary intersections we also need to compute the thermal
+                    // half transmissibilities
+                    if (enableEnergy_ && !onlyTrans) {
+                        Scalar transBoundaryEnergyIs;
+                        computeHalfDiffusivity_(transBoundaryEnergyIs,
+                                                faceAreaNormal,
+                                                distanceVector_(faceCenterInside,
+                                                                elemIdx),
+                                                1.0);
+                        thermalHalfTransBoundary.emplace(std::make_pair(elemIdx, boundaryIsIdx),
+                                                          transBoundaryEnergyIs);
+                    }
+
+                    ++boundaryIsIdx;
+                    continue;
+                }
+
+                if (!intersection.neighbor()) {
+                    // elements can be on process boundaries, i.e. they are not on the
+                    // domain boundary yet they don't have neighbors.
+                    ++boundaryIsIdx;
+                    continue;
+                }
+
+                const auto& outsideElem = intersection.outside();
+                unsigned outsideElemIdx = elemMapper.index(outsideElem);
+
+                // Get the Cartesian indices of the origen cells (parent or equivalent cell on level zero), for CpGrid with LGRs.
[…]
