### 2025-01-16 f40f670
Retrieve welpi map for Pyaction calls as well and hand it over to the runPyAction function of the Schedule
