### 2025-01-16 b9e1db4
Add test to insert WELPI keyword to pyactionActionXComparisons.cmake
