+++ pyactionActionXComparisons.cmake
@@ -135,6 +135,16 @@ add_test_compareSeparateECLFiles(CASENAME pyaction_welpi_insert_kw
                                  REL_TOL ${rel_tol}
                                  IGNORE_EXTRA_KW BOTH)
 
+add_test_compareSeparateECLFiles(CASENAME pyaction_wpimult_insert_kw
+                                 DIR1 pyaction
+                                 FILENAME1 PYACTION_WPIMULT_INSERT_KW
+                                 DIR2 actionx
+                                 FILENAME2 ACTIONX_WPIMULT
+                                 SIMULATOR flow
+                                 ABS_TOL ${abs_tol}
+                                 REL_TOL ${rel_tol}
+                                 IGNORE_EXTRA_KW BOTH)
+
 add_test_compareSeparateECLFiles(CASENAME pyaction_wsegvalv_insert_kw
                                  DIR1 pyaction
                                  FILENAME1 PYACTION_WSEGVALV_INSERT_KW
@@ -272,6 +282,17 @@ if(MPI_FOUND)
                                  ABS_TOL ${abs_tol}
                                  REL_TOL ${rel_tol}
                                  IGNORE_EXTRA_KW BOTH,
+                                 MPI_PROCS 4)
+
+  add_test_compareSeparateECLFiles(CASENAME pyaction_wpimult_insert_kw_4_procs
+                                 DIR1 pyaction
+                                 FILENAME1 PYACTION_WPIMULT_INSERT_KW
+                                 DIR2 actionx
+                                 FILENAME2 ACTIONX_WPIMULT
+                                 SIMULATOR flow
+                                 ABS_TOL ${abs_tol}
+                                 REL_TOL ${rel_tol}
+                                 IGNORE_EXTRA_KW BOTH
                                  MPI_PROCS 4)
 
   add_test_compareSeparateECLFiles(CASENAME pyaction_wsegvalv_insert_kw_4_procs
