### 2025-01-17 9ce9fe4
Transmissibility::computeHalfTrans_: use return value instead of out-param
and make it static
