### 2025-01-17 50c5369
Transmissibility::computeHalfDiffusivity_: use return value instead of out-param
and make it static
