### 2025-01-17 2197a1e
Transmissibility::update: extract calculation of harmonic mean to a lambda
