### 2025-01-17 e105315
Transmissibility::update: add faceIdxDir lambda function
short loop increases readability
