### 2025-01-17 f9d6644
add gsatprod to group rates
