### 2025-01-17 6aef3ee
Transmissibility::update(): use lambdas to reduce code duplication
