+++ opm/simulators/flow/Transmissibility_impl.hpp
@@ -244,13 +244,15 @@ update(bool global, const TransUpdateQuantities update_quantities,
         centroids_cache_[elemIdx] = centroids_(elemIdx);
     }
 
-    auto harmonicMean = [](const Scalar half1, const Scalar half2) {
+    auto harmonicMean = [](const Scalar half1, const Scalar half2)
+    {
         return (std::abs(half1) < 1e-30 || std::abs(half2) < 1e-30)
             ? 0.0
             : 1.0 / (1.0 / half1 + 1.0 / half2);
     };
 
-    auto faceIdToDir = [](int insideFaceIdx) {
+    auto faceIdToDir = [](int insideFaceIdx)
+    {
         switch (insideFaceIdx) {
         case 0:
         case 1:
@@ -267,11 +269,50 @@ update(bool global, const TransUpdateQuantities update_quantities,
         }
     };
 
+    auto halfDiff = [](const DimVector& faceAreaNormal,
+                       const unsigned,
+                       const DimVector& distVector,
+                       const Scalar prop)
+    {
+        return computeHalfDiffusivity_(faceAreaNormal,
+                                       distVector,
+                                       prop);
+    };
+
     // compute the transmissibilities for all intersections
     for (const auto& elem : elements(gridView_)) {
         FaceInfo inside;
         FaceInfo outside;
+        DimVector faceAreaNormal;
+
         inside.elemIdx = elemMapper.index(elem);
+
+        auto computeHalf = [this, &faceAreaNormal, &inside, &outside]
+                           (const auto& halfComputer,
+                            const auto& prop1, const auto& prop2) -> std::array<Scalar,2>
+        {
+            return {
+                halfComputer(faceAreaNormal,
+                             inside.faceIdx,
+                             distanceVector_(inside.faceCenter, inside.elemIdx),
+                             prop1),
+                halfComputer(faceAreaNormal,
+                             outside.faceIdx,
+                             distanceVector_(outside.faceCenter, outside.elemIdx),
+                             prop2)
+            };
+        };
+
+        auto computeHalfMean = [&inside, &outside, &computeHalf, &ntg, &harmonicMean]
+                               (const auto& halfComputer, const auto& prop)
+        {
+            auto half = computeHalf(halfComputer, prop[inside.elemIdx], prop[outside.elemIdx]);
+            applyNtg_(half[0], inside, ntg);
+            applyNtg_(half[1], outside, ntg);
+
+            //TODO Add support for multipliers
+            return harmonicMean(half[0], half[1]);
+        };
 
         unsigned boundaryIsIdx = 0;
         for (const auto& intersection : intersections(gridView_, elem)) {
@@ -281,7 +322,7 @@ update(bool global, const TransUpdateQuantities update_quantities,
                 const auto& geometry = intersection.geometry();
                 inside.faceCenter = geometry.center();
 
-                auto faceAreaNormal = intersection.centerUnitOuterNormal();
+                faceAreaNormal = intersection.centerUnitOuterNormal();
                 faceAreaNormal *= geometry.volume();
 
                 Scalar transBoundaryIs =
@@ -363,8 +404,6 @@ update(bool global, const TransUpdateQuantities update_quantities,
                 continue;
             }
 
-            DimVector faceAreaNormal;
-
             typename std::is_same<Grid, Dune::CpGrid>::type isCpGrid;
             computeFaceProperties(intersection,
                                   inside,
@@ -372,23 +411,7 @@ update(bool global, const TransUpdateQuantities update_quantities,
                                   faceAreaNormal,
                                   isCpGrid);
 
-            Scalar halfTrans1 =
-                computeHalfTrans_(faceAreaNormal,
-                                  inside.faceIdx,
[…]
