### 2025-01-17 81de485
Transmissibility: introduce a FaceInfo struct
use this to group variables in ::update()
