+++ opm/simulators/flow/Transmissibility_impl.hpp
@@ -269,7 +269,9 @@ update(bool global, const TransUpdateQuantities update_quantities,
 
     // compute the transmissibilities for all intersections
     for (const auto& elem : elements(gridView_)) {
-        unsigned elemIdx = elemMapper.index(elem);
+        FaceInfo inside;
+        FaceInfo outside;
+        inside.elemIdx = elemMapper.index(elem);
 
         unsigned boundaryIsIdx = 0;
         for (const auto& intersection : intersections(gridView_, elem)) {
@@ -277,7 +279,7 @@ update(bool global, const TransUpdateQuantities update_quantities,
             if (intersection.boundary()) {
                 // compute the transmissibilty for the boundary intersection
                 const auto& geometry = intersection.geometry();
-                const auto& faceCenterInside = geometry.center();
+                inside.faceCenter = geometry.center();
 
                 auto faceAreaNormal = intersection.centerUnitOuterNormal();
                 faceAreaNormal *= geometry.volume();
@@ -285,24 +287,24 @@ update(bool global, const TransUpdateQuantities update_quantities,
                 Scalar transBoundaryIs =
                     computeHalfTrans_(faceAreaNormal,
                                       intersection.indexInInside(),
-                                      distanceVector_(faceCenterInside, elemIdx),
-                                      permeability_[elemIdx]);
+                                      distanceVector_(inside.faceCenter, inside.elemIdx),
+                                      permeability_[inside.elemIdx]);
 
                 // normally there would be two half-transmissibilities that would be
                 // averaged. on the grid boundary there only is the half
                 // transmissibility of the interior element.
-                unsigned insideCartElemIdx = cartMapper_.cartesianIndex(elemIdx);
-                applyMultipliers_(transBoundaryIs, intersection.indexInInside(), insideCartElemIdx, transMult);
-                transBoundary_.insert_or_assign(std::make_pair(elemIdx, boundaryIsIdx), transBoundaryIs);
+                inside.cartElemIdx = cartMapper_.cartesianIndex(inside.elemIdx);
+                applyMultipliers_(transBoundaryIs, intersection.indexInInside(), inside.cartElemIdx, transMult);
+                transBoundary_.insert_or_assign(std::make_pair(inside.elemIdx, boundaryIsIdx), transBoundaryIs);
 
                 // for boundary intersections we also need to compute the thermal
                 // half transmissibilities
                 if (enableEnergy_ && !onlyTrans) {
                     Scalar transBoundaryEnergyIs =
                         computeHalfDiffusivity_(faceAreaNormal,
-                                                distanceVector_(faceCenterInside, elemIdx),
+                                                distanceVector_(inside.faceCenter, inside.elemIdx),
                                                 1.0);
-                    thermalHalfTransBoundary_.insert_or_assign(std::make_pair(elemIdx, boundaryIsIdx),
+                    thermalHalfTransBoundary_.insert_or_assign(std::make_pair(inside.elemIdx, boundaryIsIdx),
                                                                transBoundaryEnergyIs);
                 }
 
@@ -318,76 +320,76 @@ update(bool global, const TransUpdateQuantities update_quantities,
             }
 
             const auto& outsideElem = intersection.outside();
-            unsigned outsideElemIdx = elemMapper.index(outsideElem);  
+            outside.elemIdx = elemMapper.index(outsideElem);
 
-            // Get the Cartesian indices of the origen cells (parent or equivalent cell on level zero), for CpGrid with LGRs.
-            // For genral grids and no LGRs, get the usual Cartesian Index.
-            unsigned insideCartElemIdx = this-> lookUpCartesianData_.template getFieldPropCartesianIdx<Grid>(elemIdx);
[…]
