### 2025-01-17 68dfd12
Transmissibility::applyNtg_: simplify by passing FaceInfo struct
and make it static
