### 2025-01-17 2dda16b
Transmissibility::applyAllZMultipliers: simplify by passing FaceInfo structs
