### 2025-01-17 6c292a9
Transmissibility::computeFaceProperties: simplify by passing FaceInfo structs
