### 2025-01-17 8cf2876
Support instantaneous flow rates in extended network (WEFAC and GEFAC item 3)
