### 2025-01-17 #7524 @prashjha
Hello @bprather, @marouenbg, @RupeshJey, @pratikvn, and @EgorOrachyov, I am looking for two reviewers for this JOSS submission. Please let me know if you will be interested in reviewing this submission. If you have questions, feel free to ping me here.

You can access the pdf from [this link](https://github.com/openjournals/joss-papers/blob/joss.07524/joss.07524/10.21105.joss.07524.pdf) and the corresponding repository is [here](https://github.com/OPM/opm-simulators).
