### 2025-01-30 #7740 @editorialbot
[REVIEW]: gpu-ISTL - Extending OPM Flow with GPU Linear Solvers
**Submitting author:** @multitalentloes (<a href="http://orcid.org/0009-0000-8913-3034">Tobias Meyer Andersen</a>)
**Repository:** https://github.com/OPM/opm-simulators
**Branch with paper.md** (empty if default branch): joss_gpuistl_paper
**Version:** 2024.10
