### 2025-02-27 #7740 @berenger-eu
@multitalentloes I think there are few typos in the pdf: "than can " , "sucessively", "incuring".
