### 2025-02-27 #7740 @berenger-eu
@multitalentloes I have few difficulties to follow the structure of the repo.
So the GPU code has been merged in the master branch, or it is located in the joss_gpuistl_paper branch?
Can you confirm that all the authors contributed and no one has been forgotten?
