### 2025-02-27 #7740 @multitalentloes
The OPM project has grown quite comprehensive and there is a lot going on in multiple repositories, so I understand it is hard to gauge how it all relates.

The main concern of this article is the GPU work that is already contributed to the master branch of `opm-simulators`. As mentioned in the pre-review, virtually all the work relevant to this article is placed in the subdirectory `opm/simulators/linalg/gpuistl`. There is also other GPU work done in `opm-simulators` primarily something called `gpubridge`, but this is not the concern of this article in any way.

If you are on a machine with an nvidia GPU it should be detected automatically by the build system if you have the necessary cuda stuff in your PATH. You might have to excplicitly deactive `gpubridge` to cmake. If you are using an AMD GPU, then you should also add the cmake flag `CONVERT_CUDA_TO_HIP=ON`.

As you can read from the projects home page, there are multiple repositories, so ensure the other ones are from the same release.

Regarding the authors: I can confirm that these are all the authors that have played any significant part in the development of the gpu-ISTL part of OPM-simulators.

Thank you for pointing out the typos, I will fix that soon:)
