### 2025-03-06 #7740 @berenger-eu
@multitalentloes
I wanted to build it on a cluster, but it seems more tricky than expected. I found this on the doc:
```
$ cat .buildall.sh
#!/bin/bash
for repo in opm-common opm-grid opm-simulators opm-upscaling
do
    echo "=== Cloning and building module: $repo"
    git clone https://github.com/OPM/$repo.git
    mkdir $repo/build
    cd $repo/build
    cmake ..
    make
    cd ../..
done
```
I am using this module `module load tools/boost/1.80.0 compiler/gcc/11.2.0 compiler/cuda/12.3 build/cmake/3.21.3`

It seems FMT is needed, correct ? Any advice ?

[build-result.txt](https://github.com/user-attachments/files/19103644/build-result.txt)
