### 2025-03-06 #7740 @multitalentloes
@berenger-eu Have you installed everything described in the [prerequisites for installing from source](https://opm-project.org/?page_id=239)? To me is looks like the CMake error stems from the DUNE, but yeah, `libfmt-dev` is present on the prerequisites page.

If this is not a helpful response then I can ask a colleague who has more experience with building OPM Flow on clusters to chime in here.
