### 2025-03-08 #7740 @berenger-eu
@multitalentloes I was able to install Ubuntu binaries. I guess the gpuistl is included, since it is in the main branch. I may have miss it, but I can I run the example of the paper?
