### 2025-03-10 #7740 @multitalentloes
@berenger-eu Unfortunately I do not think the prebuilt binaries come with support for gpu-ISTL as it is still in active development and a bit experimental, and the software is used operationally in the industry. I think you have to build from source.

I made an attempt to describe how the benchmarks were run [here](https://github.com/OPM/opm-simulators/blob/joss_gpuistl_paper/paper/caseStudyConfiguration.md).

The scaling study was on cases generated by this [repo](https://github.com/OPM/pyopmspe11), here are a few cases I have generated that you may use: [zip](https://github.com/user-attachments/files/19157479/spe11_cases.zip)
