### 2025-03-12 #7740 @berenger-eu
@multitalentloes  I installed with:
```
#!/bin/bash
for repo in opm-common opm-grid opm-simulators opm-upscaling
do
    echo "=== Cloning and building module: $repo"
    git clone https://github.com/OPM/$repo.git
    mkdir $repo/build
    cd $repo/build
    cmake ..
    make
    cd ../..
done
```

I then switch to the branch:
```
cd opm-simulators/
git checkout joss_gpuistl_paper
cd build/
make
```
But I get this error:
```
[  5%] Building CXX object CMakeFiles/opmsimulators.dir/opm/simulators/flow/ActionHandler.cpp.o
/opm/opm-simulators/opm/simulators/flow/ActionHandler.cpp: In member function 'std::unordered_map<std::__cxx11::basic_string<char>, Scalar> Opm::ActionHandler<Scalar>::fetchWellPI(int, const Opm::Action::ActionX&, const std::vector<std::__cxx11::basic_string<char> >&) const':
/opm/opm-simulators/opm/simulators/flow/ActionHandler.cpp:225:43: error: cannot convert 'const std::vector<std::__cxx11::basic_string<char> >' to 'const Opm::Action::Result::MatchingEntities&'
  225 |                                           matching_wells);
      |                                           ^~~~~~~~~~~~~~
      |                                           |
      |                                           const std::vector<std::__cxx11::basic_string<char> >
In file included from /opm/opm-common/opm/input/eclipse/Schedule/Action/Actions.hpp:23,
                 from /opm/opm-simulators/opm/simulators/flow/ActionHandler.cpp:33:
/opm/opm-common/opm/input/eclipse/Schedule/Action/ActionX.hpp:214:50: note:   initializing argument 2 of 'std::vector<std::__cxx11::basic_string<char> > Opm::Action::ActionX::wellpi_wells(const Opm::WellMatcher&, const Opm::Action::Result::MatchingEntities&) const'
  214 |                  const Result::MatchingEntities& matches) const;
      |                  ~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~^~~~~~~
/opm/opm-simulators/opm/simulators/flow/ActionHandler.cpp: In instantiation of 'void Opm::ActionHandler<Scalar>::applyActions(int, double, const TransFunc&) [with Scalar = double; TransFunc = std::function<void(bool)>]':
/opm/opm-simulators/opm/simulators/flow/ActionHandler.cpp:278:16:   required from here
/opm/opm-simulators/opm/simulators/flow/ActionHandler.cpp:159:13: error: no match for 'operator!' (operand type is 'const Opm::Action::Result')
  159 |         if (! actionResult) {
      |             ^~~~~~~~~~~~~~
/opm/opm-simulators/opm/simulators/flow/ActionHandler.cpp:159:13: note: candidate: 'operator!(bool)' (built-in)
/opm/opm-simulators/opm/simulators/flow/ActionHandler.cpp:159:13: note:   no known conversion for argument 1 from 'const Opm::Action::Result' to 'bool'
/opm/opm-simulators/opm/simulators/flow/ActionHandler.cpp:165:56: error: no matching function for call to 'Opm::Action::Result::wells() const'
  165 |         const auto& matching_wells = actionResult.wells();
      |                                      ~~~~~~~~~~~~~~~~~~^~
In file included from /opm/opm-common/opm/input/eclipse/Schedule/Action/ActionAST.hpp:23,
                 from /opm/opm-common/opm/input/eclipse/Schedule/Action/ActionX.hpp:23:
/opm/opm-common/opm/input/eclipse/Schedule/Action/ActionResult.hpp:330:13: note: candidate: 'Opm::Action::Result& Opm::Action::Result::wells(const std::vector<std::__cxx11::basic_string<char> >&)'
  330 |     Result& wells(const std::vector<std::string>& w);
      |             ^~~~~
/opm/opm-common/opm/input/eclipse/Schedule/Action/ActionResult.hpp:330:13: note:   candidate expects 1 argument, 0 provided
In file included from /opm/opm-simulators/opm/simulators/wells/BlackoilWellModelGeneric.hpp:37,
                 from /opm/opm-simulators/opm/simulators/flow/ActionHandler.cpp:42:
/opm/opm-simulators/opm/simulators/wells/ParallelWBPCalculation.hpp: In instantiation of 'class Opm::ParallelWBPCalculation<double>':
/opm/opm-simulators/opm/simulators/wells/BlackoilWellModelGeneric.hpp:564:44:   required from 'class Opm::BlackoilWellModelGeneric<double>'
/opm/opm-simulators/opm/simulators/flow/ActionHandler.cpp:187:26:   required from 'void Opm::ActionHandler<Scalar>::applyActions(int, double, const TransFunc&) [with Scalar = double; TransFunc = std::function<void(bool)>]'
/opm/opm-simulators/opm/simulators/flow/ActionHandler.cpp:278:16:   required from here
/opm/opm-simulators/opm/simulators/wells/ParallelWBPCalculation.hpp:168:5: error: no type named 'Result' in 'class Opm::PAvgCalculator<double>'
  168 |     averagePressures(const std::size_t calcIndex) const;
      |     ^~~~~~~~~~~~~~~~
/opm/opm-simulators/opm/simulators/flow/ActionHandler.cpp: In instantiation of 'void Opm::ActionHandler<Scalar>::evalUDQAssignments(unsigned int, Opm::UDQState&) [with Scalar = double]':
/opm/opm-simulators/opm/simulators/flow/ActionHandler.cpp:278:16:   required from here
/opm/opm-simulators/opm/simulators/flow/ActionHandler.cpp:270:20: error: no matching function for call to 'Opm::UDQConfig::eval_assign(const unsigned int&, Opm::Schedule&, Opm::WellMatcher, std::function<std::unique_ptr<Opm::SegmentMatcher>()>, Opm::SummaryState&, Opm::UDQState&) const'
  270 |     udq.eval_assign(episodeIdx,
      |     ~~~~~~~~~~~~~~~^~~~~~~~~~~~
  271 |                     this->schedule_,
      |                     ~~~~~~~~~~~~~~~~
  272 |                     this->schedule_.wellMatcher(episodeIdx),
      |                     ~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~
  273 |                     this->schedule_.segmentMatcherFactory(episodeIdx),
      |                     ~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~
  274 |                     this->summaryState_,
      |                     ~~~~~~~~~~~~~~~~~~~~
  275 |                     udq_state);
      |                     ~~~~~~~~~~
In file included from /opm/opm-simulators/opm/simulators/flow/ActionHandler.cpp:38:
/opm/opm-common/opm/input/eclipse/Schedule/UDQ/UDQConfig.hpp:372:14: note: candidate: 'void Opm::UDQConfig::eval_assign(const Opm::WellMatcher&, const Opm::GroupOrder&, SegmentMatcherFactory, Opm::SummaryState&, Opm::UDQState&) const'
  372 |         void eval_assign(const WellMatcher&    wm,
      |              ^~~~~~~~~~~
/opm/opm-common/opm/input/eclipse/Schedule/UDQ/UDQConfig.hpp:372:14: note:   candidate expects 5 arguments, 6 provided
/opm/opm-common/opm/input/eclipse/Schedule/UDQ/UDQConfig.hpp:634:14: note: candidate: 'void Opm::UDQConfig::eval_assign(Opm::UDQContext&) const'
  634 |         void eval_assign(UDQContext& context) const;
      |              ^~~~~~~~~~~
/opm/opm-common/opm/input/eclipse/Schedule/UDQ/UDQConfig.hpp:634:14: note:   candidate expects 1 argument, 6 provided
make[2]: *** [CMakeFiles/opmsimulators.dir/build.make:496: CMakeFiles/opmsimulators.dir/opm/simulators/flow/ActionHandler.cpp.o] Error 1
make[1]: *** [CMakeFiles/Makefile2:476: CMakeFiles/opmsimulators.dir/all] Error 2
make: *** [Makefile:146: all] Error 2
```
