### 2025-03-12 84c80dc
update pyopmspe11 source with DOI
