+++ paper/paper.bib
@@ -83,16 +83,16 @@ abstract = "We present version 3 of the open-source simulator for flow and trans
 @book{mrst, place={Cambridge}, title={An Introduction to Reservoir Simulation Using MATLAB/GNU Octave: User Guide for the MATLAB Reservoir Simulation Toolbox (MRST)}, publisher={Cambridge University Press}, author={Lie, Knut-Andreas}, year={2019}}
 
 @article{pyopmspe11,
-title = "pyopmspe11: A {Python} framework using {OPM Flow} for the {SPE11} benchmark project [IN REVIEW]",
+title = "pyopmspe11: A {Python} framework using {OPM Flow} for the {SPE11} benchmark project",
 journal = "Journal of Open Source Software",
-year = "2024",
+publisher = "The Open Journal",
+year = "2025",
 issn = "---",
-doi = "---",
-author = "David Landa-Marbán",
+doi = "10.21105/joss.07357",
+author = {David Landa-Marbán and Tor H. Sandve},
 keywords = "IN REVIEW",
-abstract = "--"
+url = "https://doi.org/10.21105/joss.07357"
 }
-David Landa-Marbán
 
 @article{Nordbotten2024,
   title = {The 11th Society of Petroleum Engineers Comparative Solution Project: Problem Definition},
