### 2025-03-12 #7740 @pratikvn
Some comments from my side:

## Paper related:

1. I think the references for AMD need to be updated to just say AMD instead of Inc. , AMD.
2. Thank you for the reproducer for the performance case-study. I could run the CPU simulation, after installing the binaries, but the GPU version setup on containers was not successful as running containers on clusters with GPUs is a little more challenging.

## Software related:

1. I also tried compiling the software (using the instructions here: https://opm-project.org/?page_id=231), but was unsuccessful.
   a. IMO, if there are in-built required dependencies, they **need** to be installed by the software. The overhead of installing all the dependencies is in my opinion a lot of work.

## Suggestions/Possibilities:

To ease the process for us and the users, do you think it would be possible to create a container and also deploy it to dockerhub, with all your dependencies (opm, dune, and other basic dependencies). I think this would significanlty simplify the build and install processes.

For example, if you create a docker image, it can then be also used on a HPC cluster with apptainer/singularity.
