### 2025-03-12 #7740 @multitalentloes
Hello @berenger-eu, this is just because the branch containing the `paper.md` is not rebased to be in sync with the other repos, I have rebased it now.

Here is a small script I wrote that use some specific commits and should compile everything if you have the dependencies and CUDA is automatically detectable by CMAKE:

```
#!/bin/bash
# specific commit hashes for each module that should work
declare -A repo_commits=(
    ["opm-common"]="2613c9f212222cbe8b412eaab17fa2725ee98b88"
    ["opm-grid"]="300bbb91521f44a5a8f0987d452b9657557a07b3"
    ["opm-simulators"]="84c80dc1378d8c24e7f466133cfd93ec316f14b0"
)

repo_order=("opm-common" "opm-grid" "opm-simulators")

for repo in "${repo_order[@]}"
do
    echo "=== Cloning and building module: $repo"
    git clone https://github.com/OPM/$repo.git
    cd $repo
    git checkout ${repo_commits[$repo]}
    mkdir build
    cd build
    cmake .. -DUSE_GPU_BRIDGE=OFF # by default when cuda is found, GPU_BRIDGE will be built, which is not compatible with gpu-ISTL
    make -j 15 # compile with 15 cores, change this if you have fewer available
    cd ../..
done
```
