### 2025-03-12 #7740 @multitalentloes
@pratikvn, I think your suggestion of using docker images for this is interesting, I will look into this and come come back to you on this.
