### 2025-03-13 #7740 @berenger-eu
I was able to build thanks to your script.
But I still cannot run:
```
# cat gpu_ilu0.json
{
    "tol": "0.01",
    "maxiter": "200",
    "verbosity": "0",
    "solver": "gpubicgstab",
    "preconditioner": {
        "type": "OPMGPUILU0",
        "ilulevel": "0"
    }
}

# ./opm-simulators/build/bin/flow SPE11C32CUBE_LONG/deck/SPE11C32CUBE_LONG.DATA --output-dir=./output --linear-solver=./gpu_ilu0.json --threads-per-process=16 --newton-min-iterations=1 --matrix-add-well-contributions=true

**********************************************************************
*                                                                    *
*                      This is flow 2025.04-pre                      *
*                                                                    *
* Flow is a simulator for fully implicit three-phase black-oil flow, *
*             including solvent and polymer capabilities.            *
*          For more information, see https://opm-project.org         *
*                                                                    *
**********************************************************************

Using 1 MPI processes with 16 OMP threads on each

Reading deck file 'SPE11C32CUBE_LONG/deck/SPE11C32CUBE_LONG.DATA'
    0 Reading RUNSPEC  in SPE11C32CUBE_LONG/deck/SPE11C32CUBE_LONG.DATA line 4
    1 Reading DIMENS   in SPE11C32CUBE_LONG/deck/SPE11C32CUBE_LONG.DATA line 6
    2 Reading EQLDIMS  in SPE11C32CUBE_LONG/deck/SPE11C32CUBE_LONG.DATA line 9
    3 Reading TABDIMS  in SPE11C32CUBE_LONG/deck/SPE11C32CUBE_LONG.DATA line 12
    4 Reading WATER    in SPE11C32CUBE_LONG/deck/SPE11C32CUBE_LONG.DATA line 15
    5 Reading GAS      in SPE11C32CUBE_LONG/deck/SPE11C32CUBE_LONG.DATA line 16
    6 Reading CO2STORE in SPE11C32CUBE_LONG/deck/SPE11C32CUBE_LONG.DATA line 17
    7 Reading DISGASW  in SPE11C32CUBE_LONG/deck/SPE11C32CUBE_LONG.DATA line 18
    8 Reading VAPWAT   in SPE11C32CUBE_LONG/deck/SPE11C32CUBE_LONG.DATA line 19
    9 Reading DIFFUSE  in SPE11C32CUBE_LONG/deck/SPE11C32CUBE_LONG.DATA line 20
   10 Reading THERMAL  in SPE11C32CUBE_LONG/deck/SPE11C32CUBE_LONG.DATA line 21
   11 Reading METRIC   in SPE11C32CUBE_LONG/deck/SPE11C32CUBE_LONG.DATA line 23
   12 Reading START    in SPE11C32CUBE_LONG/deck/SPE11C32CUBE_LONG.DATA line 25
   13 Reading UNIFIN   in SPE11C32CUBE_LONG/deck/SPE11C32CUBE_LONG.DATA line 29
   14 Reading UNIFOUT  in SPE11C32CUBE_LONG/deck/SPE11C32CUBE_LONG.DATA line 30
   15 Reading GRID     in SPE11C32CUBE_LONG/deck/SPE11C32CUBE_LONG.DATA line 32
   16 Reading INIT     in SPE11C32CUBE_LONG/deck/SPE11C32CUBE_LONG.DATA line 34
   17 Reading COORD    in /opm/SPE11C32CUBE_LONG/deck/GRID.INC line 3
   18 Reading ZCORN    in /opm/SPE11C32CUBE_LONG/deck/GRID.INC line 1161
   19 Reading PERMX    in /opm/SPE11C32CUBE_LONG/deck/PERMX.INC line 3
   20 Reading COPY     in SPE11C32CUBE_LONG/deck/SPE11C32CUBE_LONG.DATA line 42
   21 Reading MULTIPLY in SPE11C32CUBE_LONG/deck/SPE11C32CUBE_LONG.DATA line 47
   22 Reading PORO     in /opm/SPE11C32CUBE_LONG/deck/PORO.INC line 3
   23 Reading THCONR   in /opm/SPE11C32CUBE_LONG/deck/THCONR.INC line 3
   24 Reading BCCON    in SPE11C32CUBE_LONG/deck/SPE11C32CUBE_LONG.DATA line 57
   25 Reading DISPERC  in /opm/SPE11C32CUBE_LONG/deck/DISPERC.INC line 3
   26 Reading EDIT     in SPE11C32CUBE_LONG/deck/SPE11C32CUBE_LONG.DATA line 65
   27 Reading EQUALS   in /opm/SPE11C32CUBE_LONG/deck/PVBOUNDARIES.INC line 3
   28 Reading PROPS    in SPE11C32CUBE_LONG/deck/SPE11C32CUBE_LONG.DATA line 70
   29 Reading SGWFN    in /opm/SPE11C32CUBE_LONG/deck/TABLES.INC line 3
   30 Reading SPECROCK in SPE11C32CUBE_LONG/deck/SPE11C32CUBE_LONG.DATA line 75
   31 Reading REGIONS  in SPE11C32CUBE_LONG/deck/SPE11C32CUBE_LONG.DATA line 91
   32 Reading SATNUM   in /opm/SPE11C32CUBE_LONG/deck/SATNUM.INC line 3
   33 Reading FIPNUM   in /opm/SPE11C32CUBE_LONG/deck/FIPNUM.INC line 3
   34 Reading SOLUTION in SPE11C32CUBE_LONG/deck/SPE11C32CUBE_LONG.DATA line 98
   35 Reading EQUIL    in SPE11C32CUBE_LONG/deck/SPE11C32CUBE_LONG.DATA line 100
   36 Reading RPTRST   in SPE11C32CUBE_LONG/deck/SPE11C32CUBE_LONG.DATA line 103
   37 Reading RTEMPVD  in SPE11C32CUBE_LONG/deck/SPE11C32CUBE_LONG.DATA line 107
   38 Reading SUMMARY  in SPE11C32CUBE_LONG/deck/SPE11C32CUBE_LONG.DATA line 111
   39 Reading PERFORMA in SPE11C32CUBE_LONG/deck/SPE11C32CUBE_LONG.DATA line 113
   40 Reading FGIP     in SPE11C32CUBE_LONG/deck/SPE11C32CUBE_LONG.DATA line 114
   41 Reading FGIR     in SPE11C32CUBE_LONG/deck/SPE11C32CUBE_LONG.DATA line 115
   42 Reading FGIT     in SPE11C32CUBE_LONG/deck/SPE11C32CUBE_LONG.DATA line 116
   43 Reading RGCDI    in SPE11C32CUBE_LONG/deck/SPE11C32CUBE_LONG.DATA line 117
   44 Reading RGCDM    in SPE11C32CUBE_LONG/deck/SPE11C32CUBE_LONG.DATA line 119
   45 Reading RGIP     in SPE11C32CUBE_LONG/deck/SPE11C32CUBE_LONG.DATA line 121
   46 Reading RWCD     in SPE11C32CUBE_LONG/deck/SPE11C32CUBE_LONG.DATA line 123
   47 Reading WBHP     in SPE11C32CUBE_LONG/deck/SPE11C32CUBE_LONG.DATA line 125
   48 Reading WGIR     in SPE11C32CUBE_LONG/deck/SPE11C32CUBE_LONG.DATA line 127
   49 Reading WGIT     in SPE11C32CUBE_LONG/deck/SPE11C32CUBE_LONG.DATA line 129
   50 Reading BWPR     in SPE11C32CUBE_LONG/deck/SPE11C32CUBE_LONG.DATA line 131
   51 Reading SCHEDULE in SPE11C32CUBE_LONG/deck/SPE11C32CUBE_LONG.DATA line 136
   52 Reading RPTRST   in SPE11C32CUBE_LONG/deck/SPE11C32CUBE_LONG.DATA line 138
   53 Reading BCPROP   in SPE11C32CUBE_LONG/deck/SPE11C32CUBE_LONG.DATA line 141
   54 Reading TUNING   in SPE11C32CUBE_LONG/deck/SPE11C32CUBE_LONG.DATA line 147
   55 Reading SOURCE   in SPE11C32CUBE_LONG/deck/SPE11C32CUBE_LONG.DATA line 151
   56 Reading TSTEP    in SPE11C32CUBE_LONG/deck/SPE11C32CUBE_LONG.DATA line 193
   57 Reading TUNING   in SPE11C32CUBE_LONG/deck/SPE11C32CUBE_LONG.DATA line 196
   58 Reading SOURCE   in SPE11C32CUBE_LONG/deck/SPE11C32CUBE_LONG.DATA line 200
   59 Reading TSTEP    in SPE11C32CUBE_LONG/deck/SPE11C32CUBE_LONG.DATA line 242
   60 Reading TUNING   in SPE11C32CUBE_LONG/deck/SPE11C32CUBE_LONG.DATA line 245
   61 Reading SOURCE   in SPE11C32CUBE_LONG/deck/SPE11C32CUBE_LONG.DATA line 249
   62 Reading TSTEP    in SPE11C32CUBE_LONG/deck/SPE11C32CUBE_LONG.DATA line 291
   63 Reading TUNING   in SPE11C32CUBE_LONG/deck/SPE11C32CUBE_LONG.DATA line 294
   64 Reading SOURCE   in SPE11C32CUBE_LONG/deck/SPE11C32CUBE_LONG.DATA line 298
   65 Reading TSTEP    in SPE11C32CUBE_LONG/deck/SPE11C32CUBE_LONG.DATA line 340
   66 Reading TUNING   in SPE11C32CUBE_LONG/deck/SPE11C32CUBE_LONG.DATA line 343
   67 Reading SOURCE   in SPE11C32CUBE_LONG/deck/SPE11C32CUBE_LONG.DATA line 347
   68 Reading TSTEP    in SPE11C32CUBE_LONG/deck/SPE11C32CUBE_LONG.DATA line 389
   69 Reading TUNING   in SPE11C32CUBE_LONG/deck/SPE11C32CUBE_LONG.DATA line 392
   70 Reading SOURCE   in SPE11C32CUBE_LONG/deck/SPE11C32CUBE_LONG.DATA line 396
   71 Reading TSTEP    in SPE11C32CUBE_LONG/deck/SPE11C32CUBE_LONG.DATA line 438

Creating corner-point grid from keywords COORD, ZCORN and others
3 fluid phases are active

Processing dynamic information from
SPE11C32CUBE_LONG/deck/SPE11C32CUBE_LONG.DATA line 136
Initializing report step 0/26 at 2025-01-01 0 DAYS line 136
Processing keyword RPTRST at line 138
Processing keyword BCPROP at line 141
Processing keyword TUNING at line 147
Processing keyword SOURCE at line 151
Complete report step 1 (365000 DAYS) at 1855-03-26 (0 DAYS)

Initializing report step 2/26 at 1855-03-26 (0 DAYS) line 193
Processing keyword TUNING at line 196
Processing keyword SOURCE at line 200
Complete report step 2 (1825 DAYS) at 1860-03-24 (365000 DAYS)

Initializing report step 3/26 at 1860-03-24 (365000 DAYS) line 242
Complete report step 3 (1825 DAYS) at 1865-03-23 (366825 DAYS)

Initializing report step 4/26 at 1865-03-23 (366825 DAYS) line 242
Complete report step 4 (1825 DAYS) at 1870-03-22 (368650 DAYS)

Initializing report step 5/26 at 1870-03-22 (368650 DAYS) line 242
Complete report step 5 (1825 DAYS) at 1875-03-21 (370475 DAYS)

Processing grid
Total number of active cells: 34816 / total pore volume: 421655324102 RM3

===============Saturation Functions Diagnostics===============

System: Water-Gas system.
Relative permeability input format: Saturation Family II.
Number of saturation regions: 7

Warning: Dispersion is activated in combination with VAPWAT/VAPOIL.
Water/oil is still allowed to vaporize, but dispersion in the gas phase is ignored.
CO2STORE/CO2SOL is enabled.
The surface density of CO2 is 1.868433 KG/M3.
The surface density of brine is 998.107723 KG/M3.
The surface densities are computed using the reference pressure (1.013 BARSA) and reference temperature (15.56 C).

Warning: No explicit RVWVD table is given for EQUIL region 1.
and datum depth is not at the gas-water-contact.
Rvw is set to 0.0 in all cells.

================ Starting main simulation loop ===============

Report step  0/26 at day 0/730000, date = 01-Jan-2025
Using Newton nonlinear solver.
Restart file written for report step  0/26, date = 01-Jan-2025 00:00:00

Starting time step 0, stepsize 1 days, at day 0/365000, date = 01-Jan-2025

Error: [/opm/opm-simulators/opm/simulators/linalg/PreconditionerFactory_impl.hpp:727] Preconditioner type OPMGPUILU0 is not registered in the factory. Available types are: DILU DuneILU GS ILU0 ILUn Jac ParOverILU0 SOR SSOR amg cpr cprt famg kamg

Simulation aborted as program threw an unexpected exception: Error rethrown as CriticalError at [/opm/opm-simulators/opm/simulators/linalg/ISTLSolver.hpp:385].
Original error: [/opm/opm-simulators/opm/simulators/linalg/PreconditionerFactory_impl.hpp:727] Preconditioner type OPMGPUILU0 is not registered in the factory. Available types are: DILU DuneILU GS ILU0 ILUn Jac ParOverILU0 SOR SSOR amg cpr cprt famg kamg

Error hint: This is likely due to a faulty linear solver JSON specification. Check for errors related to missing nodes.
```
But maybe the test case I use is wrong?
