### 2025-03-13 #7740 @multitalentloes
Good that the script worked! I copied your json and managed to run the exact case, so that should not be an issue.

I suspect CUDA has not been automatically detected and that is why the correct preconditioners were not compiled... I am sorry for this inconvenience, we are working on creating a docker image that can hopefully make this part of setting everything up to test GPUs in OPM much more available...

To check whether it compiled anything for GPUs, can you check what is inside the /bin folder of the project, it should contain all the tests. Do you see executables with `gpu` in their name,  such as `test_GpuBuffer`, and can you run that one for instance?
