### 2025-03-18 #7740 @kjetilly
@pratikvn We are in the process of making Docker images with GPU support. The CUDA image is fully tested and ready to be used, while the AMD image is still undergoing some development.

I have created a [PR at opm-utilities](https://github.com/OPM/opm-utilities/pull/93) adding the Dockerfiles, and we plan to have the uploading apart of the release cycle of OPM Flow, meaning new Docker images will be uploaded twice a year with the latest release to the `openporousmedia` repository of Dockerhub. As for now, the Docker image is available as `kjetilly/opmreleases:2024.10_nvidia`. The container should also work in Podman and Singularity/Apptainer, but we have not tested this yet.

The Docker image also has all the necessary build components so it should in principle be possible to use as a development container.

I have written a README.md describing the use (which will be apart of the `opm-utilities` repo and eventually on the opm-project.org as a description), and am copying the CUDA section here:

## Prerequisities
In order to run this, you will need a working NVIDIA setup on your computer, including the NVIDIA drivers and the [NVIDIA Container Toolkit](https://docs.nvidia.com/datacenter/cloud-native/container-toolkit/latest/install-guide.html).

Remember to have the NVIDIA Container Toolkit configured after installation, that is, you should have run

```bash
sudo nvidia-ctk runtime configure --runtime=docker
sudo systemctl restart docker
```

## Running
First you should clone the [opm-data](https://github.com/opm/opm-data) repository:

```bash
git clone https://github.com/opm/opm-data
cd opm-data
```

Then create a JSON file describing the linear solver on the GPU

```JSON
{
    "tol": "0.01",
    "maxiter": "200",
    "verbosity": "0",
    "solver": "gpubicgstab",
    "preconditioner": {
        "type": "GPUDILU",
        "verbosity" : 0,
        "split_matrix": "true",
        "tune_gpu_kernels": "true",
        "mixed_precision_scheme": 1
    }
}
```

name the JSON file `solver.json`.

Assuming you are now in the `opm-data` directory with the newly created `solver.json` file, you can run OPM Flow with GPU support through Docker as

```bash
docker run -it --init --rm -w $(pwd) -v $(pwd):$(pwd) -u $(id -u):$(id -g) \
    --gpus all \
    kjetilly/opmreleases:24.10_nvidia \
    --linear-solver=solver.json \
    norne/NORNE_ATW2013.DATA \
    --output-dir=output \
    --matrix-add-well-contributions=true
```

We note that `norne/NORNE_ATW2013.DATA` can be replaced by any other `.DATA` DECK file.

## Using as a development image
In principle, it should be possible to use the image as a development container. Below we describe the most basic workflow that can compile OPM Flow using said container, though we are sure there are more sophisticated workflows available using extensions to VS Code, say. It is also possible to modify the Dockerfile itself, though that is not really a smooth workflow if one wishes to recompile often.

Create the bash script `build_opm_devtainer.sh` containing

```bash
#!/bin/bash
set -e

git clone https://github.com/opm/opm-common
git clone https://github.com/opm/opm-grid
git clone https://github.com/opm/opm-simulators
git clone https://github.com/opm/opm-upscaling
git clone https://github.com/opm/opm-utilities
cp opm-utilities/opm-super/CMakeLists.txt ./
mkdir -p build && cd build
cmake \
    -DUSE_OPENCL=OFF \
    -DUSE_GPU_BRIDGE=OFF \
    ..
make -j10 flow #NOTE: Adjust number of procs as needed. OPM uses a lot of memory compiling per proc (O(2GB))
```

then run

```bash
docker run --rm -it --init --entrypoint bash \
    -v $(pwd):$(pwd) -u $(id -u):$(id -g) -w $(pwd) \
     kjetilly/opmreleases:2024.10_nvidia \
    build_opm_devtainer.sh
```

You will then get a compiled version of OPM Flow on your local filesystem (in `$(pwd)/build/opm-simulators/bin/flow`) that you can run through the Docker image by

```bash
docker run --rm -it --init --gpus all \
    --entrypoint $(pwd)/build/opm-simulators/bin/flow \
    -v $(pwd):$(pwd) -u $(id -u):$(id -g) -w $(pwd) \
    kjetilly/opmreleases:2024.10_nvidia \
    --linear-solver=solver.json \
    --output-dir=output \
    --matrix-add-well-contributions=true \
    norne/NORNE_ATW2013.DATA
```

Where one assumes that `norne/NORNE_ATW2013.DATA` is accessible from the current directory (from [opm-data](https://github.com/opm/opm-data)) and `solver.josn` is the file above, that is, it should be

```json
{
    "tol": "0.01",
    "maxiter": "200",
    "verbosity": "0",
    "solver": "gpubicgstab",
    "preconditioner": {
        "type": "GPUDILU",
        "verbosity" : 0,
        "split_matrix": "true",
        "tune_gpu_kernels": "true",
        "mixed_precision_scheme": 1
    }
}
```
