### 2025-03-19 #7740 @multitalentloes
I think these docker images @kjetilly have written are a great addition to OPM that makes the gpuISTL part much more accessible, does it work for you @berenger-eu and @pratikvn? If it does, then I am sure we can add a description of this and make it an official thing in the upcoming april-release of OPM Flow.
