### 2025-03-19 #7740 @pratikvn
Thanks! Running with apptainer with the CUDA container works quite smoothly. I still have to check the Rocm container. But I cannot seem to reproduce the results from the paper (Figure 3).

### CPU run
```shell
mpirun -n 16 flow --linear-solver=cpu_ilu0.json \
                            --threads-per-process=1 --newton-min-iterations=1 \
                            --output-dir=output --matrix-add-well-contributions=true \
                            SPE11C32CUBE_LONG/deck/SPE11C32CUBE_LONG.DATA
```
with `cpu_ilu0.json`:
```json
{
        "tol": "0.01",
        "maxiter": "200",
        "verbosity": "0",
        "solver": "bicgstab",
        "preconditioner": {
                "type": "ILU0",
                "ilulevel": "0"
        }
}
```
with timings:
```
.
.
.
Number of MPI processes:        16
Threads per MPI process:         1
Setup time:                      1.00 s
  Deck input:                    0.45 s
Number of timesteps:          2006
Simulation time:                58.91 s
  Assembly time:                20.74 s (Wasted: 0.0 s; 0.0%)
    Well assembly:               0.00 s (Wasted: 0.0 s; 0.0%)
  Linear solve time:             8.73 s (Wasted: 0.0 s; 0.0%)
    Linear setup:                1.71 s (Wasted: 0.0 s; 0.0%)
  Props/update time:            25.22 s (Wasted: 0.0 s; 0.0%)
  Pre/post step:                 3.28 s (Wasted: 0.0 s; 0.0%)
  Output write time:             0.62 s
Overall Linearizations:       4229      (Wasted:     0; 0.0%)
Overall Newton Iterations:    2223      (Wasted:     0; 0.0%)
Overall Linear Iterations:   21273      (Wasted:     0; 0.0%)
```

### GPU run
```shell
mpirun -n 1 flow --linear-solver=gpu_ilu0.json \
                          --threads-per-process=1 --newton-min-iterations=1 \
                          --output-dir=output --matrix-add-well-contributions=true \
                          SPE11C32CUBE_LONG/deck/SPE11C32CUBE_LONG.DATA
```
with `gpu_ilu0.json`:
```json
{
        "tol": "0.01",
        "maxiter": "200",
        "verbosity": "0",
        "solver": "gpubicgstab",
        "preconditioner": {
                "type": "OPMGPUILU0",
                "ilulevel": "0",
                "tune_gpu_kernels": "true",
                "mixed_precision_scheme": 1
        }
}
```
with timings:
```
.
.
.
Number of MPI processes:         1
Threads per MPI process:        16
Setup time:                      0.84 s
  Deck input:                    0.21 s
Number of timesteps:          2006
Simulation time:               448.64 s
  Assembly time:               178.27 s (Wasted: 0.0 s; 0.0%)
    Well assembly:               0.00 s (Wasted: 0.0 s; 0.0%)
  Linear solve time:            66.37 s (Wasted: 0.0 s; 0.0%)
    Linear setup:                6.17 s (Wasted: 0.0 s; 0.0%)
  Props/update time:           175.53 s (Wasted: 0.0 s; 0.0%)
  Pre/post step:                27.10 s (Wasted: 0.0 s; 0.0%)
  Output write time:             0.25 s
Overall Linearizations:       4216      (Wasted:     0; 0.0%)
Overall Newton Iterations:    2210      (Wasted:     0; 0.0%)
Overall Linear Iterations:   17228      (Wasted:     0; 0.0%)
```
Maybe it can be expected that the assembly time and the overall simulation time for the GPU run to be higherm but maybe the linear solve time should be better for the GPU run ? It is 66.3s v/s 8.7s for the CPU.

I am using a H100 GPU, and have a Intel Xeon Gold 32 core CPU. Maybe there is something I can do to check that it is actually being run on the GPU ?
