### 2025-03-19 #7740 @pratikvn
Okay, it seems that I was running the small case with ~ 35k cells, for which GPU is slower. I am currently running a larger case with ~900k cells (for which the GPU is expected to be faster) and will update the timings after that finishes.
