### 2025-03-20 #7740 @multitalentloes
I am glad to hear that the container worked for you!

Your last observation is precisely it, @pratikvn. For very small cases CPUs are faster, and your case with 900k cells will very likely be quite a bit faster on GPU (at least the linear solve time!). The H100 is quite overkill for that case and will not be fully saturated.

Tip: If your simulation is taking a lot of time, you can speed up the GPU simulation time by using multiple threads. This will accelerate the assembly time and the props/update time, but not affect the linear solve time.
