### 2025-03-21 #7740 @berenger-eu
Ok, the container with CUDA worked for me.
So I will update the checkboxes, everything is fine for me.
Only change is did you include the description of the docker images, etc. in the documentation ? And maybe updating the paper to make sure to refer to it ?
