### 2025-03-21 #7740 @multitalentloes
We use Jenkins for CI which is typically triggered by code reviewers to ensure all tests pass and that no new warnings are introduced, you should see this being used in all PRs being merged.
