### 2025-03-21 #7740 @multitalentloes
@berenger-eu out of curiosity, since we have not run gpuISTL on the particular hardware you used, can you share your runtimes?
