### 2025-03-22 #7740 @pratikvn
So, for the overall simulation, the GPU run does seem to take a lot more time (But I guess that is expected as the large distributed systems are now only solved on one GPU). Here are a summary of the my runs:

| type | num_cells | num_ranks | total_time (s) | assembly (s) | linear_solve (s)  | update (s) | post (s)  | lin_iters|
| --- | --- | --- | --- |--- | --- |--- | --- | --- |
| cpu | ~38k | 16| 59|21|9|25|3|21273|
| gpu | ~38k | 1| 448|178|66|175|27|17228|
| cpu | ~900k | 16|8146|1017|5724|1132|263|305883
| gpu | ~900k | 1| 25202|11868|1735|9550|2008| 67961|

The results in the paper are reproducible, so I think that is good. But I have some questions:
1. It seems that if you have more MPI ranks, with as many MPI ranks as GPUs, then you dont really see an improvement in the linear solve time. Are the linear systems not distributed across GPUs ?
2. I do observe that the overall number of linear iterations is very different for the CPU and GPU cases. I understand that they are run on different number of MPI ranks, but should the reported number be that different ? Maybe I am missing something here ?
3. For the larger cases, assembly takes a significantly more amount of time. Is there a plan to also move that to the GPU ?
