### 2025-03-23 #7740 @pratikvn
Okay, it seems that with `split_matrix` set to true, you can also do distributed solves, which is quite nice. But atleast within the container setup, that does not seem to scale that well, but that might also be because of the small problem size as well. I didnt want to run it for the large problem size, as that takes quite a while to run.

| type | num_cells | num_ranks(threads_per_rank) | total_time (s) | assembly (s) | linear_solve (s)  | update (s) | post (s)  | lin_iters|
| --- | --- | --- | --- |--- | --- |--- | --- | --- |
| cpu | ~38k | 16 (1)| 59|21|9|25|3|21273|
| nvidia-gpu (H100) x1 | ~38k | 1 (32)| 448|178|66|175|27|17228|
| nvidia-gpu (H100) x2 | ~38k | 2 (16)| 270|99|58|96|14|18681|
| nvidia-gpu (H100) x4 | ~38k | 4 (8)| 104|17|44|33|9|18706|
