### 2025-03-24 #7740 @kjetilly
@pratikvn glad to hear the Docker container is working well. With regards to your questions:

>I do observe that the overall number of linear iterations is very different for the CPU and GPU cases. I understand that they are run on different number of MPI ranks, but should the reported number be that different ? Maybe I am missing something here ?

By default, the CPU path uses the CPRW preconditioner, which will usually result in much fewer linear iterations than the DILU/ILU precondtioners on the GPU. The benchmark plot in the paper is against the ILU0 preconditioner on the CPU. In a just published paper we have done a more thorough analysis of the performance, also against CPRW on the CPU (https://onepetro.org/spersc/proceedings-abstract/25RSC/25RSC/D021S009R004/646798 this is a paper where we go much more indept in the performance analysis but do not really describe the overall gpu-ISTL implementation as such), where the conclusion is that at least on an AMD 7950X, we beat the CPU CPRW with around 200k cells with a 4070 Ti or MI250X GPU), where we use all 16 cores on the CPU for CPRW.

Additionally, the `.json` file I supplied might have slightly different settings (tolerances) than the default.

You can do an apples to apples comparison by making JSON file containing:

```JSON
{
    "tol": "0.01",
    "maxiter": "200",
    "verbosity": "0",
    "solver": "bicgstab",
    "preconditioner": {
        "type": "DILU",
        "verbosity" : 0
    }
}
```
and save as say `solver_cpu_dilu.json`. You can then set `--linear-solver=solver_cpu_dilu.json` and it should run with the DILU preconditioner and the same tolerances/maxiters on the CPU.

Likewise, you could set the CPRW preconditioner to the same tolerances by supplying the following JSON file

```JSON
{
    "tol": "0.01",
    "maxiter": "200",
    "verbosity": "0",
    "solver": "bicgstab",
    "preconditioner": {
        "type": "cpr",
        "verbosity" : 0
    }
}
```

> For the larger cases, assembly takes a significantly more amount of time. Is there a plan to also move that to the GPU ?

Yes. This is in fact on-going work as we speak. We hope to have a proof-of-concept for the full assembly this year, at least for some select cases.

> I do observe better timings for the AMD GPU than the NVIDIA GPU, with the GPUs and the setup I use. Can you maybe point me to the files (.cu, .hip.cpp etc) which implement the kernels ? Maybe there are some possible improvements possible there ?

This is weird, though granted we have not benchmarked this on NVIDIA server grade/compute grade hardware (or at least not recently). For DILU, the relevant kernels are at https://github.com/OPM/opm-simulators/blob/master/opm/simulators/linalg/gpuistl/detail/preconditionerKernels/DILUKernels.cu , and they are called from https://github.com/OPM/opm-simulators/blob/master/opm/simulators/linalg/gpuistl/GpuDILU.cpp (`apply` and `update`). The `.cu` files are converted to `.hip` files on the fly if one compiles for HIP (ie. there are no `.hip` files in gpu-ISTL). Note that the master version could deveate a bit from the 2024.10 release which is shipped with the Docker image.

My hunch is however that we might not build for the correct CUDA architecture in the Docker container. In the building of the AMD container, I have supplied a myriad of HIP architectures, but we are only compiling for rather conservative architectures for the CUDA part. I have added the relevant H100 architecture (9.0) now and have uploaded it to `kjetilly/opmreleases:2024.10_nvidia`. If you want you could try to pull it again and run.

Though another source could also simply be that the 900k cell model is too small for the given GPU and that some latencies and other overheads dominate (say transferring to/and from the GPU). The (D)ILU preconditioners does have less parallelism due to dependencies within the linear system, so with 900k cells you might be under saturating the GPU quite a bit.

>It seems that if you have more MPI ranks, with as many MPI ranks as GPUs, then you dont really see an improvement in the linear solve time. Are the linear systems not distributed across GPUs ?

This should in fact be automatic (the `split_matrix` is a performance option for the DILU/ILU-preconditioners on the GPU and relates to the memory layout for the matrix, and should be irrelevant for the MPI partitioning). The only thing I can think of off the top of my head, except for model size, would be that the selection of the GPU is not done properly (handled here: https://github.com/OPM/opm-simulators/blob/master/opm/simulators/linalg/gpuistl/set_device.cpp ), but if you are dealing with a single-node, multiple cards setup, this should be OK. It could be that you get suboptimal selection of GPU versus CPU-memory for memory transfers, though.

However, 900k cells is a rather small model to be running multiple H100 GPUs on. I would think to get anything meaningful, you should go to say 30 million cells at least (the SPE11 models can be generated that large). Out of curiosity, which model are you trying to run here?

We do support GPU aware MPI, though the Docker image does not support it (and it is honestly a bit finicky to get working, and highly depends on the cluster setup). This could also give some speedup.

>I didnt want to run it for the large problem size, as that takes quite a while to run.

If you find the `SCHEDULE` line in the deck file, it is possible to simply remove timesteps to make it run for a shorter time.

I think also @multitalentloes can chime in with information I might have missed as well.
