### 2025-03-24 #7740 @kjetilly
@pratikvn another thing, just remove the ` "mixed_precision_scheme": 1` line of the GPU JSON file, that will ensure that the CPU and GPU uses the same precision scheme. I think this was the culprit in the end with regards to different number of linear iterations, since you actually specified ILU0 in the JSON I see.
