### 2025-03-24 #7740 @pratikvn
I see.
1. I think it makes sense that adding Hopper to the list of architectures should give better performance.
2. GPU Aware MPI would be nice, but I think for a container, it is not very important.
3. It makes sense that if distribution across the GPUs is the default, then within the container, mapping of the ranks to the GPUs might be non-optimal or it may be that one GPU is being over-subscribed. In general I think it should be possible for the user to set that through the `srun` or also with `mpirun`, so that should be okay.

I didnt try running the 900k cell model on multiple GPUs, but the even smaller 38k cell model. So the degradation is scaling is expected there.

I was using the `SPE11C32CUBE_LONG` (38k cells) and `SPE11C96CUBE_LONG` (900k cells) models that were provided by @multitalentloes .

For the purposes of this review, I am happy with the results, and I dont plan to run any more performance tests.
