### 2025-03-25 #7740 @kjetilly
@pratikvn thank you for your feedback!

> 1. Regarding the tests. I see that the opm-simulators repo has some test suite. Can you maybe highlight this on your webpage, and also add instructions on how users can run these tests ?

We have added a [dedicated page on how to run the unit tests](https://opm-project.org/?page_id=1853), which we link to from [the page where we describe how to compile OPM](https://opm-project.org/?page_id=231).

> 2. I am also not sure if there is a CI setup. I dont think that is a requirement for this paper, but a publicly available CI (for example with github CI: https://github.blog/news-insights/product-news/github-actions-now-supports-ci-cd/ , It is free for open source projects) would greatly enhance the sustainability of this project. I understand the availability of GPUs might be a concern, but I think you should be able to atleast compile the code without running the tests.

Maybe I don't quite get your point. We do have continuous integration handled by a privately administered Jenkins server. At the moment, only approved developers can trigger builds from Jenkins for pull requests. You can check eg. https://github.com/OPM/opm-simulators/pull/6099 for an example where Jenkins has been run on a pull request and failed due to some issues. I think you need to be logged in to GitHub to see the Jenkins status.

We do plan to add GPU servers to also run the GPU tests when needed. At the moment, the GPU code is simply compiled by Jenkins, but the unit tests are not run.

We furthermore do CI every time new commits lands in the master branch, and we have some weekly analysis as well.

Our Jenkins setup is (partially) [described in the GitHub repository](https://github.com/OPM/opm-simulators/tree/master/jenkins), including the build scripts used.

> 3. If you wish, you can also go a step further and automate your container deployment to the github container repository: ghcr.io , for example on every release.

Thanks for the tip! This is something we can consider for the next release.
