### 2025-03-25 #7740 @pratikvn
@kjetilly Perfect! Thank you for adding the information regarding the tests.

Thanks for pointing to the Jenkins setup. I somehow missed that when looking through the repo. I think that is already sufficient. I might also be worth it to advertise that on your [main github README page with a badge](https://docs.github.com/en/actions/monitoring-and-troubleshooting-workflows/monitoring-workflows/adding-a-workflow-status-badge).

@prashjha I am finished with my review, and all my concerns have been addressed. I recommend the paper for acceptance.
