### 2025-03-31 #7740 @multitalentloes
In the `Statement of need` section I added a new reference as we just got a paper published which uses `gpu-ISTL`
