### 2025-04-22 #7740 @multitalentloes
I have now published a zenodo-dataset containing the source code, and the test data used for the plot on https://zenodo.org/records/15259506

Is there anything else we need to do?
