### 2025-05-04 #7740 @prashjha
@multitalentloes, I also noticed the change in order of authors in the JOSS article and Zenodo archive.
