### 2025-05-04 #7740 @prashjha
@multitalentloes: Some action items for you:

1. Please address my comments on JOSS draft; see [comments](https://github.com/openjournals/joss-reviews/issues/7740#issuecomment-2849350703)
2. Check the author order in Zenodo and the JOSS article. Please confirm whether you are okay with different orders and briefly explain why.
3. Confirm version of the library is 2024.10.
4. Address the DOI errors (see above).
