### 2025-05-06 #7740 @multitalentloes
Thank you for the quick response @prashjha, I have now re-added the section on what commands were used to run the simulator to produce the case study results. Hopefully this provides a concise guide on how to run the simulator with `gpu-ISTL`.
