### 2025-05-07 #7740 @multitalentloes
I have now cited the Zenodo backup which contains the case files and expanded the `path/to/DATA/file` to actually fit with what you can download from there. Thank you for the suggestion, I think the section is better now that we are more specific.

Anything else I should fix, @prashjha?
