### 2025-05-09 #7740 @multitalentloes
@prashjha I have now reverted the file paths in the command to emphasize that it is a `.DATA` file that is the primary argument to flow. I also rewrote the section prior to that to explain what the data files used in the benchmark look like in the Zenodo dataset that they can access.

The error messages were a bit cryptic as they did not give any clues as to which lines caused them, but I have updated two suspicious lines in the `.bib` file that I guess was the problem, we will see if this is resolved once the paper acceptance pipeline is run again...
