### 2025-05-09 #7740 @prashjha
@multitalentloes, there are still some errors. Check link 1. I think the problem is that you have used the label 'given-names' instead of 'name' in the authors block in the paper.md. Another suggestion from link 1 is removing the PDF file from the folder. Try correcting the label first, and let me know so I can run the command again.

[1] https://github.com/openjournals/joss-reviews/issues/7365
[2] https://forum.pkp.sfu.ca/t/crossref-xml-export-throws-up-errors-in-validation-xml/62114
