### 2025-05-13 #7740 @danielskatz
👋 @multitalentloes - as track editor, I've proofread this paper, and found two small issues in references, in https://github.com/OPM/opm-simulators/pull/6268.  Please merge this, then we can accept and publish the work.
