### 2025-05-13 #7740 @editorialbot
:wave: @openjournals/csism-eics, this paper is ready to be accepted and published.
