### 2025-05-13 #7740 @danielskatz
Congratulations to @multitalentloes ([Tobias Meyer Andersen](http://orcid.org/0009-0000-8913-3034)) and co-authors on your publication!!

JOSS depends on volunteers, and we couldn't do this without you.

Also, note that the DOI isn't resolving, as the Crossref service that we use is undergoing weekly maintenance.  This should be complete in 75 minutes, and the DOI should start working sometime after that.  I'll keep this issue open until then.
