### 2025-05-13 #7740 @multitalentloes
The DOI resolves now for me - so I think all is good.

Thank you @berenger-eu  and @pratikvn for valuable feedback in your reviews, the addition of docker images for both GPU vendors were added to OPM Flow as of the last release, a very direct improvement from your testing and feedback.

Thank you @prashjha for thorough proof-reading as editor, and thank you @danielskatz too for proof-reading, and especially for initiating the pre-review which did not automatically start.

All of your volunteer work is greatly appreciated!
